@@ -18,6 +18,7 @@ package com.wasteofplastic.askyblock;
 
 import java.io.File;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Locale;
 import java.util.Set;
@@ -25,6 +26,8 @@ import java.util.Set;
 import org.bukkit.ChatColor;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
+
+import com.wasteofplastic.askyblock.Island.SettingsFlag;
 
 
 /**
@@ -53,6 +56,7 @@ public class ASLocale {
             "yellow",
             "white"
             ));
+
 
     // Localization Strings
     private FileConfiguration locale = null;
@@ -106,6 +110,7 @@ public class ASLocale {
     // ASkyBlock
     public String topTenheader;
     public String topTenGuiTitle;
+    public String topTenGuiHeading;
     public String topTenerrorNotReady;
     public String levelislandLevel;
     public String levelerrornotYourIsland;
@@ -197,6 +202,7 @@ public class ASLocale {
     public String islanderrorLevelNotReady;
     public String islanderrorInvalidPlayer;
     public String islandislandLevelis;
+    public String islandrequiredPointsToNextLevel;
     // Spawn
     public String islandhelpSpawn;
     // Teleport go
@@ -304,6 +310,7 @@ public class ASLocale {
     public String purgecalculating;
     public String purgenoneFound;
     public String purgethisWillRemove;
+    public String purgeEstimatedRunTime;
     public String purgewarning;
     public String purgetypeConfirm;
     public String purgepurgeCancelled;
@@ -325,7 +332,6 @@ public class ASLocale {
 
     // info
     public String adminInfoislandLocation;
-    public String adminInfoerrorNotPartOfTeam;
     public String adminInfoerrorNullTeamLeader;
     public String adminInfoerrorTeamMembersExist;
     public String adminHelpinfoIsland;
@@ -370,7 +376,7 @@ public class ASLocale {
     public String minishopBuyProblem;
     public String minishopYouCannotAfford;
     public String errorMinishopDisabled;
-    
+
     // Sign
     public String signLine1;
     public String signLine2;
@@ -413,6 +419,7 @@ public class ASLocale {
     public String islandhelpCoop;
     public String coopInvited;
     public String coopUseExpel;
+    public String cannotCoop;
 
     public String errorWrongWorld;
     public String islandcannotTeleport;
@@ -426,6 +433,7 @@ public class ASLocale {
     public String lockUnlocking;
     public String islandHelpLock;
     public String helpColor;
+    public String helpHeader;
     public String lockPlayerLocked;
     public String lockPlayerUnlocked;
     public String lockEnteringSpawn;
@@ -448,50 +456,18 @@ public class ASLocale {
     public String challengescompletedtimes;
     public String targetInNoPVPArea;
     // Island Guard Settings
+    public HashMap<SettingsFlag, String> igs;
     public String igsTitle;
     public String igsAllowed;
     public String igsDisallowed;
-    public String igsArmorStand;
-    public String igsBeacon;
-    public String igsBed;
-    public String igsBreakBlocks;
-    public String igsBreeding;
-    public String igsBrewing;
-    public String igsBucket;
-    public String igsChest;
     public String igsChestDamage;
-    public String igsWorkbench;
-    public String igsCropTrampling;
-    public String igsDoor;
-    public String igsEnchanting;
-    public String igsEnderPearl;
     public String igsFire;
-    public String igsFurnace;
-    public String igsGate;
-    public String igsHurtAnimals;
-    public String igsHurtMobs;
     public String igsCreeperDamage;
     public String igsCreeperGriefing;
     public String igsWitherDamage;
-    public String igsLeash;
-    public String igsLever;
-    public String igsSpawnEgg;
-    public String igsJukebox;
-    public String igsPlaceBlocks;
-    public String igsPortalUse;
-    public String igsPVP;
-    public String igsRedstone;
-    public String igsShears;
     public String igsTeleport;
     public String igsTNT;
-    public String igsVisitorDrop;
-    public String igsVisitorPickUp;
     public String igsVisitorKeep;
-    public String igsNetherPVP;
-    public String igsAnvil;
-    public String igsHorseRiding;
-    public String igsHorseInvAccess;
-    public String igsPressurePlate;
     public String igsPistonPush;
     public String igsSettingsGeneralTitle;
     public String igsSettingsGeneralDesc;
@@ -499,11 +475,7 @@ public class ASLocale {
     public String igsSettingsIslandDesc;
     public String igsSettingsSpawnTitle;
     public String igsSettingsSpawnDesc;
-    public String igsVillagerTrading;
-    public String igsChorusFruit;
-    public String igsJoinLeaveIslandMessage;
-    public String igsMobSpawning;
-    
+
     public String setHomeerrorNumHomes;
     public String schematicsTitle;
     public String islandhelpBan;
@@ -643,15 +615,75 @@ public class ASLocale {
 
     public String adminHelpAddRange;
 
+    public String errorBlockedByAPI;
+
     public String adminHelpcobbleStats;
 
     public String topTenerrorExcluded;
 
+    public String leaveWarning;
+
+    public String leaveCanceled;
+
+    public String unlimited;
+
+    public String resetsLeft;
+
+    public String islandblockValue;
+
+    public String islandblockWorthless;
+
+    public String errorNotABlock;
+
+    public String islandhelpValue;
+
+    public String adminSetRangeUpdated;
+
+    public String adminResetChallengeForAllError;
+
+    public String adminHelpresetChallengeForAll;
+
+    public String adminHelplistChallengeResets;
+
+    public String adminHelpclearChallengeReset;
+
+    public String adminResetChallengeForAllRepeating;
+
+    public String adminResetChallengeForAllReset;
+
+    public String adminHelpclearResetAll;
+
+    public String teamChatSpyStatusOff;
+
+    public String teamChatSpyStatusOn;
+
+    public String adminHelpsetLanguage;
+
+    public String purgeLimit;
+    
+    public String needRelog;
+
+
+    public String visitorsProtected;
+
+
+    public String biomePleaseWait;
+
+
+    public String coopHasInvited;
+
+
+    public String adminHelpSwitch;
+    public String adminSwitchOp;
+    public String adminSwitchRemoving;
+    public String adminSwitchAdding;
+
+
     /**
      * Creates a locale object full of localized strings for a language
-     * @param plugin
+     * @param plugin - ASkyBlock plugin object - ASkyBlock plugin
      * @param localeName - name of the yml file that will be used
-     * @param index 
+     * @param index - index number of the locale
      */
     public ASLocale(ASkyBlock plugin, String localeName, int index) {
         this.plugin = plugin;
@@ -659,12 +691,13 @@ public class ASLocale {
         this.localeName = localeName;
         getLocale(localeName);
         loadLocale();
-        if (!localeName.equalsIgnoreCase("locale")) {
-            localeObject = new Locale(localeName.substring(0, 2), localeName.substring(3, 5));
-        }       
+        //if (!localeName.equalsIgnoreCase("locale")) {
+        localeObject = new Locale(localeName.substring(0, 2), localeName.substring(3, 5));
+        //}       
     }
 
     /**
+     * @param localeName requested locale name
      * @return locale FileConfiguration object
      */
     public FileConfiguration getLocale(String localeName) {
@@ -676,6 +709,7 @@ public class ASLocale {
 
     /**
      * Reloads the locale file
+     * @param localeName - locale to reload
      */
     public void reloadLocale(String localeName) {
         //plugin.getLogger().info("DEBUG: loading local file " + localeName + ".yml");
@@ -698,20 +732,7 @@ public class ASLocale {
                 locale = YamlConfiguration.loadConfiguration(localeFile);
                 //locale.setDefaults(defLocale);
             } else {
-                // Use the default file
-                localeFile = new File(plugin.getDataFolder() + File.separator + "locale", "locale.yml");
-                if (localeFile.exists()) {
-                    locale = YamlConfiguration.loadConfiguration(localeFile);
-                } else {
-                    // Look for defaults in the jar                    
-                    if (plugin.getResource("locale/locale.yml") != null) {
-                        plugin.saveResource("locale/locale.yml", true);
-                        localeFile = new File(plugin.getDataFolder() + File.separator + "locale", "locale.yml");
-                        locale = YamlConfiguration.loadConfiguration(localeFile);
-                    } else {
-                        plugin.getLogger().severe("Could not find any locale file!");
-                    }
-                }
+                plugin.getLogger().severe("Could not find locale file! " + localeName);
             }
         }
     }
@@ -750,12 +771,14 @@ public class ASLocale {
 
         drankAcidAndDied = ChatColor.translateAlternateColorCodes('&', locale.getString("drankAcidAndDied", "drank acid and died."));
         drankAcid = ChatColor.translateAlternateColorCodes('&', locale.getString("drankAcid", "drank acid."));
+        errorBlockedByAPI = ChatColor.translateAlternateColorCodes('&', locale.getString("error.blockedbyapi", "A plugin using the API blocked this action."));
         errorTooShort = ChatColor.translateAlternateColorCodes('&', locale.getString("error.tooShort", "Too short. Minimum size is [length]."));
         errorTooLong = ChatColor.translateAlternateColorCodes('&', locale.getString("error.tooLong", "Too long. Maximum size is [length]."));
         errorUnknownPlayer = ChatColor.translateAlternateColorCodes('&', locale.getString("error.unknownPlayer", "That player is unknown."));
         errorNoPermission = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("error.noPermission", "You don't have permission to use that command!"));
         errorNoIsland = ChatColor.translateAlternateColorCodes('&', locale.getString("error.noIsland", "You do not have an island!"));
+        errorNotABlock = ChatColor.translateAlternateColorCodes('&', locale.getString("error.notABlock", "That is not a block"));
         errorNoIslandOther = ChatColor
                 .translateAlternateColorCodes('&', locale.getString("error.noIslandOther", "That player does not have an island!"));
         // "You must be on your island to use this command."
@@ -773,41 +796,66 @@ public class ASLocale {
         if (igsTitle.length() > 30) {
             igsTitle = igsTitle.substring(0,29);
         }
-        igsAnvil = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.anvil", "Anvil Use"));
         igsAllowed = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.allowed", "Allowed"));
         igsDisallowed = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.disallowed", "Disallowed"));
-        igsArmorStand = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.armorstand", "Armor Stand use"));
-        igsBeacon = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.beacon", "Beacon use"));
-        igsBed = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.bed", "Bed use"));
-        igsBreakBlocks = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.breakblocks", "Break blocks"));
-        igsBreeding = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.breeding", "Breeding"));
-        igsBrewing = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.brewingstand", "Potion Brewing"));
-        igsBucket = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.bucket", "Bucket use"));
-        igsChest = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.chest", "Chest use"));
+        // Developer - add new text for GUI icons here and in the local files
+        igs = new HashMap<SettingsFlag, String>();
+        igs.put(SettingsFlag.ACID_DAMAGE, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.aciddamage", "Acid Damage")));
+        igs.put(SettingsFlag.ANVIL, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.anvil", "Visitor anvil use")));
+        igs.put(SettingsFlag.ARMOR_STAND, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.armorstand", "Visitor armor stand use")));
+        igs.put(SettingsFlag.BEACON, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.beacon", "Visitor beacon use")));
+        igs.put(SettingsFlag.BED, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.bed", "Visitor bed use")));
+        igs.put(SettingsFlag.BREAK_BLOCKS, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.breakblocks", "Visitor block breaking")));
+        igs.put(SettingsFlag.BREEDING, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.breeding", "Visitor animal breeding")));
+        igs.put(SettingsFlag.BREWING, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.brewingstand", "Visitor potion brewing")));
+        igs.put(SettingsFlag.BUCKET, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.bucket", "Visitor bucket use")));
+        igs.put(SettingsFlag.COLLECT_LAVA, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.collectlava", "Visitor lava collection")));
+        igs.put(SettingsFlag.COLLECT_WATER, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.collectwater", "Visitor water collection")));
+        igs.put(SettingsFlag.CHEST, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.chest", "Visitor chest use")));
+        igs.put(SettingsFlag.CHORUS_FRUIT, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.chorusfruit", "Visitor chorus fruit use")));
+        igs.put(SettingsFlag.CRAFTING, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.workbench", "Visitor workbench use")));
+        igs.put(SettingsFlag.CREEPER_PAIN, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.creeperpain", "Creeper explosions hurt")));
+        igs.put(SettingsFlag.CROP_TRAMPLE, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.croptrample", "Visitor crop trampling")));
+        igs.put(SettingsFlag.DOOR, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.door", "Visitor door use")));
+        igs.put(SettingsFlag.EGGS, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.eggs", "Visitor egg throwing")));
+        igs.put(SettingsFlag.ENCHANTING, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.enchantingtable", "Visitor enchanting table use")));
+        igs.put(SettingsFlag.ENDER_PEARL, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.enderpearl", "Visitor enderpearl use")));
+        igs.put(SettingsFlag.ENTER_EXIT_MESSAGES, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.joinleaveislandmessages", "Enter/Exit Island Messages")));
+        igs.put(SettingsFlag.FIRE, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.fire", "Fire")));
+        igs.put(SettingsFlag.FIRE_EXTINGUISH, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.fireextinguish", "Visitor fire extinguish")));
+        igs.put(SettingsFlag.FIRE_SPREAD, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.firespread", "Fire spread")));
+        igs.put(SettingsFlag.FURNACE, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.furnace", "Visitor furnace use")));
+        igs.put(SettingsFlag.GATE, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.gate", "Visitor gate use")));
+        igs.put(SettingsFlag.HORSE_INVENTORY, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.horseinventoryaccess", "Visitor horse inventory access")));   
+        igs.put(SettingsFlag.HORSE_RIDING, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.horseriding", "Visitor horse riding")));   
+        igs.put(SettingsFlag.HURT_MOBS, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.hurtanimals", "Visitor animal hurting")));
+        igs.put(SettingsFlag.HURT_MONSTERS, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.hurtmonsters", "Visitor monster hurting")));
+        igs.put(SettingsFlag.LEASH, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.leash", "Visitor leash use")));
+        igs.put(SettingsFlag.LEVER_BUTTON, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.lever", "Visitor lever or button use")));
+        igs.put(SettingsFlag.MOB_SPAWN, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.animalspawning", "Allow/Disallow Animal Spawning")));
+        igs.put(SettingsFlag.MONSTER_SPAWN, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.monsterspawning", "Allow/Disallow Monster Spawning")));
+        igs.put(SettingsFlag.MUSIC, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.jukebox", "Visitor jukebox use")));
+        igs.put(SettingsFlag.PLACE_BLOCKS, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.placeblocks", "Visitor block placing")));
+        igs.put(SettingsFlag.PORTAL, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.portaluse", "Visitor portal use")));
+        igs.put(SettingsFlag.PRESSURE_PLATE, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.pressureplate", "Visitors activate pressure plates")));   
+        igs.put(SettingsFlag.PVP, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.pvp", "PvP")));
+        igs.put(SettingsFlag.MILKING, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.milking", "Visitor cow milking")));   
+        igs.put(SettingsFlag.NETHER_PVP, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.netherpvp", "Nether PvP")));
+        igs.put(SettingsFlag.REDSTONE, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.redstone", "Visitor redstone use")));
+        igs.put(SettingsFlag.SPAWN_EGGS, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.spawnegg", "Spawn egg use")));
+        igs.put(SettingsFlag.SHEARING, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.shears", "Visitor shears use")));
+        igs.put(SettingsFlag.VILLAGER_TRADING, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.villagertrading", "Visitor villager trading")));
+        igs.put(SettingsFlag.VISITOR_ITEM_DROP, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.visitordrop", "Visitor item drop")));
+        igs.put(SettingsFlag.VISITOR_ITEM_PICKUP, ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.visitorpickup", "Visitor item pick-up")));
+
         igsChestDamage = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.chestdamage", "Chest damage by TNT"));
-        igsWorkbench = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.workbench", "Workbench use"));
-        igsCropTrampling = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.croptrample", "Crop trampling"));
-        igsDoor = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.door", "Door use"));
-        igsEnchanting = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.enchantingtable", "Enchanting table use"));
-        igsEnderPearl = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.enderpearl", "Enderpearl use"));
-        igsFire = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.fire", "Fire"));
-        igsFurnace = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.furnace", "Furnace use"));
-        igsGate = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.gate", "Gate use"));
-        igsHurtAnimals = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.hurtanimals", "Hurting animals"));
-        igsHurtMobs = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.hurtmonsters", "Hurting monsters"));
         igsCreeperDamage = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.creeperdamage", "Creeper damage"));
         igsCreeperGriefing = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.creepergriefing", "Creeper griefing"));
         igsWitherDamage = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.withergriefing", "Wither griefing"));
-        igsLeash = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.leash", "Leash use"));
-        igsLever = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.lever", "Lever or Button Use"));
-        igsSpawnEgg = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.spawnegg", "Spawn egg use"));
-        igsJukebox = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.jukebox", "Jukebox use"));
-        igsPlaceBlocks = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.placeblocks", "Place blocks"));
-        igsPortalUse = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.portaluse", "Portal use"));
-        igsPVP = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.pvp", "PvP"));
-        igsNetherPVP = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.netherpvp", "Nether PvP"));
-        igsRedstone = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.redstone", "Redstone use"));
-        igsShears = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.shears", "Shears use"));
+        igsFire = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.fire", "Fire"));
+        igsTNT = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.TNTdamage", "TNT Damage"));
+        igsPistonPush = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.pistonpush", "Pistons can push blocks outside island area")); 
+        igsVisitorKeep = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.visitorkeepitems", "Visitor keep item on death"));   
         igsSettingsGeneralTitle = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.settingsgeneraltitle", "General Island World Settings"));
         if (igsSettingsGeneralTitle.length() > 30) {
             igsSettingsGeneralTitle = igsSettingsGeneralTitle.substring(0,29);
@@ -817,26 +865,14 @@ public class ASLocale {
         if (igsSettingsIslandTitle.length() > 30) {
             igsSettingsIslandTitle = igsSettingsIslandTitle.substring(0,29);
         }
-        igsSettingsIslandDesc = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.settingsislanddesc", "These rules apply to visitors on this island"));
+        igsSettingsIslandDesc = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.settingsislanddesc", "These rules apply to this island"));
         igsSettingsSpawnTitle = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.settingsspawntitle", "Spawn Settings"));
         if (igsSettingsSpawnTitle.length() > 30) {
             igsSettingsSpawnTitle = igsSettingsSpawnTitle.substring(0,29);
         }
         igsSettingsSpawnDesc = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.settingsspawndesc", "These rules apply at spawn."));
         igsTeleport = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.teleportwhenfalling", "Teleport when falling"));
-        igsTNT = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.TNTdamage", "TNT Damage"));
-        igsVisitorDrop = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.visitordrop", "Visitor item dropping"));
-        igsVisitorPickUp = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.visitorpickup", "Visitor item pick-up"));
-        igsVisitorKeep = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.visitorkeepitems", "Visitor keep item on death"));	
-        igsHorseRiding = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.horseriding", "Horse riding"));	
-        igsHorseInvAccess = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.horseinventoryaccess", "Horse inv access when riding"));	
-        igsPressurePlate = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.pressureplate", "Activate pressure plates"));	
-        igsPistonPush = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.pistonpush", "Pistons can push blocks outside island area"));	
-        igsVillagerTrading = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.villagertrading", "Villager trading"));
-        igsChorusFruit = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.chorusfruit", "Chorus Fruit Use"));
-        igsJoinLeaveIslandMessage = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.joinleaveislandmessages", "Enable Join & Leave Island Messages"));
-        igsMobSpawning = ChatColor.translateAlternateColorCodes('&', locale.getString("islandguardsettings.mobspawning", "Allow/Disallow Mob Spawning"));
-        
+
         lavaTip = ChatColor.translateAlternateColorCodes('&', locale.getString("lavaTip", "Changing obsidian back into lava. Be careful!"));
         warpswelcomeLine = ChatColor.translateAlternateColorCodes('&', locale.getString("warps.welcomeLine", "[WELCOME]"));
         warpswarpTip = ChatColor.translateAlternateColorCodes('&',
@@ -850,7 +886,7 @@ public class ASLocale {
         warpserrorNoPerm = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("warps.errorNoPerm", "You do not have permission to place Welcome Signs yet!"));
         warpserrorNotEnoughLevel = ChatColor.translateAlternateColorCodes('&', 
-        		locale.getString("warps.errorNotEnoughLevel", "You do not have enough island levels to create a Welcome Sign!"));
+                locale.getString("warps.errorNotEnoughLevel", "You do not have enough island levels to create a Welcome Sign!"));
         warpserrorNoPlace = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("warps.errorNoPlace", "You must be on your island to place a Welcome Sign!"));
         warpserrorDuplicate = ChatColor.translateAlternateColorCodes('&',
@@ -877,6 +913,7 @@ public class ASLocale {
         if (topTenGuiTitle.length() > 30) {
             topTenGuiTitle = topTenGuiTitle.substring(0, 29);
         }
+        topTenGuiHeading = ChatColor.translateAlternateColorCodes('&', locale.getString("topTen.guiHeading", "&E&L<!> Island: &6&N[name]&7 (#[rank])"));
         topTenerrorNotReady = ChatColor.translateAlternateColorCodes('&', locale.getString("topTen.errorNotReady", "Top ten list not generated yet!"));
         topTenerrorExcluded = ChatColor.translateAlternateColorCodes('&', locale.getString("topTen.errorExcluded", "FYI: You are excluded from the top ten as you do not have [perm]"));
         levelislandLevel = ChatColor.translateAlternateColorCodes('&', locale.getString("level.islandLevel", "Island level"));
@@ -961,9 +998,9 @@ public class ASLocale {
                 locale.getString("island.cannotTeleport", "You cannot teleport when falling!"));
         islandnew = ChatColor.translateAlternateColorCodes('&', locale.getString("island.new", "Creating a new island for you..."));
         islandSubTitle = locale.getString("island.subtitle", "by tastybento");
-        islandDonate = locale.getString("island.donate", "ASkyBlock by tastybento, click here to donate via PayPal!");
+        islandDonate = locale.getString("island.donate", "ASkyBlock by tastybento");
         islandTitle = locale.getString("island.title", "A SkyBlock");
-        islandURL = locale.getString("island.url", "https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=ZSBJG5J2E3B7U");
+        islandURL = locale.getString("island.url", "");
         islandSubTitleColor = colorCheck(locale.getString("island.subtitlecolor", "blue"));
         islandDonateColor = colorCheck(locale.getString("island.donatecolor", "aqua"));
         islandTitleColor = colorCheck(locale.getString("island.titlecolor", "gold"));
@@ -985,6 +1022,10 @@ public class ASLocale {
                 locale.getString("island.resetWait", "You have to wait [time] seconds before you can do that again."));
         islandresetConfirm = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("island.resetConfirm", "Type /[label] confirm within [seconds] seconds to delete your island and restart!").replace("[label]", Settings.ISLANDCOMMAND));
+        islandblockValue = ChatColor.translateAlternateColorCodes('&',
+                locale.getString("island.blockValue", "[name] placed here may be worth [value]"));
+        islandblockWorthless = ChatColor.translateAlternateColorCodes('&',
+                locale.getString("island.blockWorthless", "[name] is worthless"));
         islandhelpIsland = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("island.helpIsland", "start an island, or teleport to your island."));
         islandhelpTeleport = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpTeleport", "teleport to your island."));
@@ -999,6 +1040,7 @@ public class ASLocale {
         islandhelpLevelPlayer = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("island.helpLevelPlayer", "see another player's island level."));
         islandhelpTop = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpTop", "see the top ranked islands."));
+        islandhelpValue = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpValue", "see island level value of block in hand"));
         islandhelpWarps = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpWarps", "Lists all available welcome-sign warps."));
         islandhelpWarp = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpWarp", "Warp to <player>'s welcome sign."));
         islandhelpTeam = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpTeam", "view your team information."));
@@ -1015,8 +1057,8 @@ public class ASLocale {
         adminHelpHelp = ChatColor.translateAlternateColorCodes('&', locale.getString("adminHelp.help", "Admin Commands:"));
         adminHelpName = ChatColor.translateAlternateColorCodes('&', locale.getString("adminHelp.name", "Set name of player's island"));
         adminHelpResetName = ChatColor.translateAlternateColorCodes('&', locale.getString("adminHelp.resetname", "Reset name of player's island"));
-        islandHelpSelectLanguage = ChatColor.translateAlternateColorCodes('&', locale.getString("islandHelp.selectlanguage", "Select Language"));
-       
+        islandHelpSelectLanguage = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpSelectLanguage", "Select Language"));
+
         islandhelpAcceptReject = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("island.helpAcceptReject", "accept or reject an invitation."));
         islandhelpMakeLeader = ChatColor
@@ -1027,7 +1069,8 @@ public class ASLocale {
                 locale.getString("island.errorLevelNotReady", "Can't use that command right now! Try again in a few seconds."));
         islanderrorInvalidPlayer = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("island.errorInvalidPlayer", "That player is invalid or does not have an island!"));
-        islandislandLevelis = ChatColor.translateAlternateColorCodes('&', locale.getString("island.islandLevelis", "Island level is"));
+        islandislandLevelis = ChatColor.translateAlternateColorCodes('&', locale.getString("island.islandLevelis", "Island level is : [level]"));
+        islandrequiredPointsToNextLevel = ChatColor.translateAlternateColorCodes('&', locale.getString("island.requiredPointsToNextLevel", "You need [points] more points to reach level [next]!"));
 
         invitehelp = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("invite.help", "Use [/[label] invite <playername>] to invite a player to your island.").replace("[label]", Settings.ISLANDCOMMAND));
@@ -1079,6 +1122,10 @@ public class ASLocale {
                 locale.getString("leave.errorYouMustBeInWorld", "You must be in the island world to leave your team!"));
         leaveerrorLeadersCannotLeave = ChatColor.translateAlternateColorCodes('&', locale.getString("leave.errorLeadersCannotLeave",
                 "Leaders cannot leave an island. Make someone else the leader fist using /[label] makeleader <player>").replace("[label]", Settings.ISLANDCOMMAND));
+        leaveWarning = ChatColor.translateAlternateColorCodes('&', locale.getString("leave.warning",
+                "Are you sure you want to leave the team? Type '[label] leave' again to confirm.").replace("[label]", Settings.ISLANDCOMMAND));
+        leaveCanceled = ChatColor.translateAlternateColorCodes('&',
+                locale.getString("leave.canceled", "Leave canceled"));
         teamlistingMembers = ChatColor.translateAlternateColorCodes('&', locale.getString("team.listingMembers", "Listing your island members"));
         kickerrorPlayerNotInTeam = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("kick.errorPlayerNotInTeam", "That player is not in your team!"));
@@ -1115,10 +1162,15 @@ public class ASLocale {
         adminHelpunregister = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminHelp.unregister", "deletes a player without deleting the island blocks"));
         adminHelpdelete = ChatColor.translateAlternateColorCodes('&', locale.getString("adminHelp.delete", "delete an island (removes blocks)"));
+        adminHelpclearChallengeReset = ChatColor.translateAlternateColorCodes('&', locale.getString("adminHelp.clearChallengeReset", "clear a global challenge reset")); 
         adminHelpcompleteChallenge = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminHelp.completeChallenge", "marks a challenge as complete"));
         adminHelpresetChallenge = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminHelp.resetChallenge", "marks a challenge as incomplete"));
+        adminHelpresetChallengeForAll = ChatColor.translateAlternateColorCodes('&',
+                locale.getString("adminHelp.resetChallengeForAll", "globally resets a challenge for every player with an optional repetition"));
+        adminHelplistChallengeResets = ChatColor.translateAlternateColorCodes('&',
+                locale.getString("adminHelp.listChallengeResets", "lists challenge reset schedules, if any"));
         adminHelpresetAllChallenges = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminHelp.resetAllChallenges", "resets all of the player's challenges"));
         adminHelppurge = ChatColor.translateAlternateColorCodes('&',
@@ -1131,6 +1183,8 @@ public class ASLocale {
         adminHelpinfo = ChatColor.translateAlternateColorCodes('&', locale.getString("adminHelp.info", "check information on the given player"));
         adminHelpSetSpawn = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminHelp.setspawn", "sets the island world spawn to a location close to you"));
+        adminHelpsetLanguage = ChatColor.translateAlternateColorCodes('&',
+                locale.getString("adminHelp.setlanguage", "sets the default language and resets every player to that language"));
         adminHelpSetRange = ChatColor
                 .translateAlternateColorCodes('&', locale.getString("adminHelp.setrange", "changes the island's protection range"));
         adminHelpAddRange = ChatColor
@@ -1149,13 +1203,14 @@ public class ASLocale {
                 .translateAlternateColorCodes('&', locale.getString("adminHelp.add", "adds player to leader's team"));
         adminHelpcobbleStats = ChatColor
                 .translateAlternateColorCodes('&', locale.getString("adminHelp.cobblestats", "displays magic cobble generation stats"));
-
+        adminHelpSwitch = ChatColor
+                .translateAlternateColorCodes('&', locale.getString("adminHelp.switch", "toggles protection bypass"));
         settingsResetInProgress = ChatColor
                 .translateAlternateColorCodes('&', locale.getString("settingsReset.inprogress", "Protection settings being reset, please wait..."));
         settingsResetDone = ChatColor
                 .translateAlternateColorCodes('&', locale.getString("settingsReset.done", "Done."));
         adminHelpSettingsReset = ChatColor
-                .translateAlternateColorCodes('&', locale.getString("adminHelp.settingsreset", "Resets all islands to default protection settings"));
+                .translateAlternateColorCodes('&', locale.getString("adminHelp.settingsreset", "resets all islands to default protection settings"));
         adminHelplevel = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminHelp.level", "provide detailed island level report on player."));
         adminHelpinfoIsland = ChatColor.translateAlternateColorCodes('&',
@@ -1175,6 +1230,7 @@ public class ASLocale {
         purgenoneFound = ChatColor.translateAlternateColorCodes('&', locale.getString("purge.noneFound", "No inactive islands to remove."));
         purgethisWillRemove = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("purge.thisWillRemove", "[number] inactive islands found. Islands with level < [level] will be removed."));
+        purgeEstimatedRunTime = ChatColor.translateAlternateColorCodes('&', locale.getString("purge.estimatedRunTime", "Estimated run time is : [time]."));
         purgewarning = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("purge.warning", "DANGER! Do not run this with players on the server! MAKE BACKUP OF WORLD!"));
         purgetypeConfirm = ChatColor.translateAlternateColorCodes('&',
@@ -1188,6 +1244,7 @@ public class ASLocale {
         purgeStillChecking = ChatColor.translateAlternateColorCodes('&', locale.getString("purge.stillChecking", "Still checking player files..."));
         purgeSkyBlockFound = ChatColor.translateAlternateColorCodes('&', locale.getString("purge.skyblockFound", "There are [number] unowned islands. Do '/asadmin purge unowned confirm' to delete them within 20 seconds."));
         purgeAcidFound = ChatColor.translateAlternateColorCodes('&', locale.getString("purge.acidFound", "There are [number] unowned islands. Do '/acid purge unowned confirm' to delete them within 20 seconds."));
+        purgeLimit = ChatColor.translateAlternateColorCodes('&', locale.getString("purge.maxPurge", "Max purge is [number] islands. Run purge again afterwards to purge more."));
         adminAllowPurge = ChatColor.translateAlternateColorCodes('&', locale.getString("purge.preventName", "Purge protection removed"));
         adminPreventPurge = ChatColor.translateAlternateColorCodes('&', locale.getString("purge.allowName", "Island is protected from purging"));
         confirmerrorTimeLimitExpired = ChatColor.translateAlternateColorCodes('&',
@@ -1198,8 +1255,6 @@ public class ASLocale {
         registererrorBedrockNotFound = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("register.errorBedrockNotFound", "Error: unable to set the island!"));
         adminInfoislandLocation = ChatColor.translateAlternateColorCodes('&', locale.getString("adminInfo.islandLocation", "Island Location"));
-        adminInfoerrorNotPartOfTeam = ChatColor.translateAlternateColorCodes('&',
-                locale.getString("adminInfo.errorNotPartOfTeam", "That player is not a member of an island team."));
         adminInfoerrorNullTeamLeader = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminInfo.errorNullTeamLeader", "Team leader should be null!"));
         adminInfoerrorTeamMembersExist = ChatColor.translateAlternateColorCodes('&',
@@ -1221,10 +1276,6 @@ public class ASLocale {
         registererrorBedrockNotFound = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("register.errorBedrockNotFound", "Error: unable to set the island!"));
         adminInfoislandLocation = ChatColor.translateAlternateColorCodes('&', locale.getString("adminInfo.islandLocation", "Island Location"));
-        adminInfoerrorNotPartOfTeam = ChatColor.translateAlternateColorCodes('&',
-                locale.getString("adminInfo.errorNotPartOfTeam", "That player is not a member of an island team."));
-        adminInfoerrorNullTeamLeader = ChatColor.translateAlternateColorCodes('&',
-                locale.getString("adminInfo.errorNullTeamLeader", "Team leader should be null!"));
         adminInfoerrorTeamMembersExist = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminInfo.errorTeamMembersExist", "Player has team members, but shouldn't!"));
         resetChallengessuccess = ChatColor.translateAlternateColorCodes('&',
@@ -1262,7 +1313,11 @@ public class ASLocale {
                 locale.getString("boats.warning", "It's unsafe to exit the boat right now..."));
         adminHelpclearReset = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("adminHelp.clearreset", "resets the island reset limit for player."));
+        adminHelpclearResetAll = ChatColor.translateAlternateColorCodes('&',
+                locale.getString("adminHelp.clearresetall", "resets the island reset limit for all players."));
         resetYouHave = ChatColor.translateAlternateColorCodes('&', locale.getString("island.resetYouHave", "You have [number] resets left."));
+        resetsLeft = ChatColor.translateAlternateColorCodes('&', locale.getString("island.resetsLeft", "Resets left"));
+        unlimited = ChatColor.translateAlternateColorCodes('&', locale.getString("island.unlimitd", "Unlimited"));
         islandResetNoMore = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("island.resetNoMore", "No more resets are allowed for your island!"));
         clearedResetLimit = ChatColor.translateAlternateColorCodes('&', locale.getString("resetTo", "Cleared reset limit"));
@@ -1275,6 +1330,7 @@ public class ASLocale {
         if (biomePanelTitle.length() > 30) {
             biomePanelTitle = biomePanelTitle.substring(0, 29);
         }
+        biomePleaseWait = ChatColor.translateAlternateColorCodes('&', locale.getString("biome.pleasewait", "Please wait, changing biome..."));
         expelNotOnIsland = ChatColor.translateAlternateColorCodes('&',
                 locale.getString("expel.notonisland", "Player is not trespassing on your island!"));
         expelSuccess = ChatColor.translateAlternateColorCodes('&', locale.getString("expel.success", "You expelled [name]!"));
@@ -1298,16 +1354,18 @@ public class ASLocale {
         coopRemoved = ChatColor.translateAlternateColorCodes('&', locale.getString("coop.removed", "[name] remove your coop status!"));
         coopRemoveSuccess = ChatColor.translateAlternateColorCodes('&', locale.getString("coop.removesuccess", "[name] is no longer a coop player."));
         coopSuccess = ChatColor.translateAlternateColorCodes('&',
-                locale.getString("coop.success", "[name] is now a coop player until they log out or you expel them."));
+                locale.getString("coop.success", "[name] is now a coop player until you log out, uncoop or expel him."));
         coopMadeYouCoop = ChatColor.translateAlternateColorCodes('&',
-                locale.getString("coop.madeyoucoopy", "[name] made you a coop player until you log out or they expel you."));
+                locale.getString("coop.madeyoucoopy", "[name] made you a coop player until he logs out or expels you."));
         coopOnYourTeam = ChatColor.translateAlternateColorCodes('&', locale.getString("coop.onyourteam", "Player is already on your team!"));
         coopNotInCoop = ChatColor.translateAlternateColorCodes('&', locale.getString("coop.notincoop", "[name] is not in your coop!"));
         islandhelpCoop = ChatColor.translateAlternateColorCodes('&',locale.getString("coop.help", "temporarily give a player full access to your island"));
         islandhelpUnCoop = ChatColor.translateAlternateColorCodes('&',locale.getString("coop.uncoop", "remove full island access from player"));
         islandhelpListCoops = ChatColor.translateAlternateColorCodes('&',locale.getString("coop.listcoops", "list coop players"));
         coopInvited = ChatColor.translateAlternateColorCodes('&', locale.getString("coop.invited", "[name] made [player] a coop player!"));
+        coopHasInvited = ChatColor.translateAlternateColorCodes('&', locale.getString("coop.nameHasInvitedYou", "[name] has invited you to coop on their island!"));
         coopUseExpel = ChatColor.translateAlternateColorCodes('&', locale.getString("coop.useexpel", "Use expel to remove."));
+        cannotCoop = ChatColor.translateAlternateColorCodes('&', locale.getString("coop.cannotcoop", "Only island leader can coop a player."));
         lockIslandLocked = ChatColor.translateAlternateColorCodes('&', locale.getString("lock.islandlocked", "Island is locked to visitors"));
         lockNowEntering = ChatColor.translateAlternateColorCodes('&', locale.getString("lock.nowentering", "Now entering [name]'s island"));
         lockNowLeaving = ChatColor.translateAlternateColorCodes('&', locale.getString("lock.nowleaving", "Now leaving [name]'s island"));
@@ -1315,6 +1373,7 @@ public class ASLocale {
         lockUnlocking = ChatColor.translateAlternateColorCodes('&', locale.getString("lock.unlocking", "Unlocking island"));
         islandHelpLock = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpLock", "Locks island so visitors cannot enter it"));
         helpColor = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpColor", "&e"));
+        helpHeader = ChatColor.translateAlternateColorCodes('&', locale.getString("island.helpHeader", "[plugin] v[version] help:"));
         lockPlayerLocked = ChatColor.translateAlternateColorCodes('&', locale.getString("lock.playerlocked", "[name] locked the island"));
         lockPlayerUnlocked = ChatColor.translateAlternateColorCodes('&', locale.getString("lock.playerunlocked", "[name] unlocked the island"));
         lockEnteringSpawn = ChatColor.translateAlternateColorCodes('&', locale.getString("lock.enteringspawn", "Entering Spawn"));
@@ -1325,12 +1384,14 @@ public class ASLocale {
         }
         teamChatPrefix = ChatColor.translateAlternateColorCodes('&', locale.getString("teamchat.prefix", "[Team Chat]<{ISLAND_PLAYER}> "));
         teamChatHelp = ChatColor.translateAlternateColorCodes('&', locale.getString("teamchat.helpChat", "turn on/off team chat"));
+        teamChatSpyStatusOff = ChatColor.translateAlternateColorCodes('&', locale.getString("teamchat.spyoff", "Team chat spy is off"));
+        teamChatSpyStatusOn = ChatColor.translateAlternateColorCodes('&', locale.getString("teamchat.spyon", "Team chat spy is on"));
         teamChatStatusOff = ChatColor.translateAlternateColorCodes('&', locale.getString("teamchat.statusOff", "Team chat is off"));
         teamChatStatusOn = ChatColor.translateAlternateColorCodes('&', locale.getString("teamchat.statusOn", "Team chat is on"));
         teamChatNoTeamAround = ChatColor.translateAlternateColorCodes('&', locale.getString("teamchat.noTeamAround", "None of your team are online!"));
         teamChatNoTeam = ChatColor.translateAlternateColorCodes('&', locale.getString("teamchat.noTeam", "You are not in a team!"));
         adminLockerrorInGame = ChatColor.translateAlternateColorCodes('&', locale.getString("adminLock.errorInGame", "Must use command in-game while on an island!"));
-        errorNotOnIsland = ChatColor.translateAlternateColorCodes('&', locale.getString("error.notonisland", "You are not in an island space!"));
+        errorNotOnIsland = ChatColor.translateAlternateColorCodes('&', locale.getString("error.notOnIsland", "You are not in an island space!"));
         adminLockadminUnlockedIsland = ChatColor.translateAlternateColorCodes('&', locale.getString("adminLock.adminUnlockedIsland", "Admin unlocked your island"));
         adminLockadminLockedIsland = ChatColor.translateAlternateColorCodes('&', locale.getString("adminLock.adminLockedIsland", "Admin locked your island"));
         adminTopBreedersFinding = ChatColor.translateAlternateColorCodes('&', locale.getString("adminTopBreeders.finding", "Finding top breeders..."));
@@ -1356,6 +1417,9 @@ public class ASLocale {
         adminSetHomeNoneFound = ChatColor.translateAlternateColorCodes('&', locale.getString("adminSetHome.noneFound", "No safe location found!"));
         adminSetHomeHomeSet = ChatColor.translateAlternateColorCodes('&', locale.getString("adminSetHome.homeSet", "Home set to [location]"));
         adminSetHomeNotOnPlayersIsland = ChatColor.translateAlternateColorCodes('&', locale.getString("adminSetHome.notOnPlayersIsland", "You are not on the player's island"));
+        adminResetChallengeForAllError = ChatColor.translateAlternateColorCodes('&', locale.getString("adminResetChallengeForAll.error", "Format for repeat time must is [integer number][m/h/d] (minutes, hours, days), e.g. 5h"));
+        adminResetChallengeForAllRepeating = ChatColor.translateAlternateColorCodes('&', locale.getString("adminResetChallengeForAll.repeating", "Repeating [duration]"));
+        adminResetChallengeForAllReset = ChatColor.translateAlternateColorCodes('&', locale.getString("adminResetChallengeForAll.reset", "Reset [date]"));       
         adminHelpResetSign = ChatColor.translateAlternateColorCodes('&', locale.getString("adminHelp.resetSign", "Resets the sign you are looking at to the island owner"));
         adminResetSignNoSign = ChatColor.translateAlternateColorCodes('&', locale.getString("adminResetSign.noSign", "You must be looking at a sign post to run this command."));
         adminResetSignFound = ChatColor.translateAlternateColorCodes('&', locale.getString("adminResetSign.found", "Warp Sign found!"));
@@ -1398,6 +1462,16 @@ public class ASLocale {
         generalSuccess = ChatColor.translateAlternateColorCodes('&', locale.getString("general.success", "Success!"));
         adminHelpReserve = ChatColor.translateAlternateColorCodes('&', locale.getString("adminHelp.reserve", "reserves this spot for player's next island"));
         adminReserveIslandExists = ChatColor.translateAlternateColorCodes('&', locale.getString("adminReserve.islandExists", "There is an island here already! Register the player instead!"));
+        adminSetRangeUpdated = ChatColor.translateAlternateColorCodes('&', locale.getString("adminSetRange.updated", "Island range updated to [number]"));
+        
+        needRelog = ChatColor.translateAlternateColorCodes('&', locale.getString("needRelog", "You need to relog in order to see the changes."));
+        visitorsProtected = ChatColor.translateAlternateColorCodes('&', locale.getString("visitors.protected", "Visitors protected"));
+        
+        adminSwitchOp = ChatColor.translateAlternateColorCodes('&', locale.getString("adminSwitch.op", "Ops can always bypass protection. Deop to use command."));
+        adminSwitchAdding = ChatColor.translateAlternateColorCodes('&', locale.getString("adminSwitch.adding", "Adding protection bypass..."));
+        adminSwitchRemoving = ChatColor.translateAlternateColorCodes('&', locale.getString("adminSwitch.removing", "Removing protection bypass..."));
+        
+
     }
 
     /**
