@@ -1,5 +1,4 @@
 /*******************************************************************************
- * This file is part of ASkyBlock.
  *
  *     ASkyBlock is free software: you can redistribute it and/or modify
  *     it under the terms of the GNU General Public License as published by
@@ -18,15 +17,10 @@
 package com.wasteofplastic.askyblock;
 
 import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.TreeMap;
+import java.util.Map;
 import java.util.UUID;
 
-import org.apache.commons.lang.StringUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.GameMode;
@@ -34,31 +28,15 @@ import org.bukkit.Material;
 import org.bukkit.World;
 import org.bukkit.WorldCreator;
 import org.bukkit.WorldType;
-import org.bukkit.block.Biome;
-import org.bukkit.configuration.ConfigurationSection;
-import org.bukkit.configuration.file.YamlConfiguration;
-import org.bukkit.entity.Animals;
-import org.bukkit.entity.Entity;
-import org.bukkit.entity.EntityType;
-import org.bukkit.entity.Guardian;
-import org.bukkit.entity.Monster;
 import org.bukkit.entity.Player;
 import org.bukkit.event.HandlerList;
-import org.bukkit.event.entity.EntityDamageEvent.DamageCause;
 import org.bukkit.generator.ChunkGenerator;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.material.MaterialData;
-import org.bukkit.material.SpawnEgg;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.potion.PotionEffect;
-import org.bukkit.potion.PotionEffectType;
-import org.bukkit.potion.PotionType;
 import org.bukkit.scheduler.BukkitRunnable;
-import org.bukkit.scoreboard.Scoreboard;
-import org.bukkit.scoreboard.Team;
 
-import com.wasteofplastic.askyblock.Island.Flags;
 import com.wasteofplastic.askyblock.NotSetup.Reason;
 import com.wasteofplastic.askyblock.Settings.GameType;
 import com.wasteofplastic.askyblock.Updater.UpdateResult;
@@ -66,14 +44,15 @@ import com.wasteofplastic.askyblock.Updater.UpdateType;
 import com.wasteofplastic.askyblock.commands.AdminCmd;
 import com.wasteofplastic.askyblock.commands.Challenges;
 import com.wasteofplastic.askyblock.commands.IslandCmd;
+import com.wasteofplastic.askyblock.events.IslandPreDeleteEvent;
 import com.wasteofplastic.askyblock.events.IslandDeleteEvent;
 import com.wasteofplastic.askyblock.events.ReadyEvent;
 import com.wasteofplastic.askyblock.generators.ChunkGeneratorWorld;
 import com.wasteofplastic.askyblock.listeners.AcidEffect;
 import com.wasteofplastic.askyblock.listeners.ChatListener;
 import com.wasteofplastic.askyblock.listeners.CleanSuperFlat;
+import com.wasteofplastic.askyblock.listeners.EntityLimits;
 import com.wasteofplastic.askyblock.listeners.FlyingMobEvents;
-import com.wasteofplastic.askyblock.listeners.HeroChatListener;
 import com.wasteofplastic.askyblock.listeners.IslandGuard;
 import com.wasteofplastic.askyblock.listeners.IslandGuard1_8;
 import com.wasteofplastic.askyblock.listeners.IslandGuard1_9;
@@ -82,6 +61,8 @@ import com.wasteofplastic.askyblock.listeners.LavaCheck;
 import com.wasteofplastic.askyblock.listeners.NetherPortals;
 import com.wasteofplastic.askyblock.listeners.NetherSpawning;
 import com.wasteofplastic.askyblock.listeners.PlayerEvents;
+import com.wasteofplastic.askyblock.listeners.PlayerEvents2;
+import com.wasteofplastic.askyblock.listeners.PlayerEvents3;
 import com.wasteofplastic.askyblock.listeners.WorldEnter;
 import com.wasteofplastic.askyblock.listeners.WorldLoader;
 import com.wasteofplastic.askyblock.panels.BiomesPanel;
@@ -89,7 +70,7 @@ import com.wasteofplastic.askyblock.panels.ControlPanel;
 import com.wasteofplastic.askyblock.panels.SchematicsPanel;
 import com.wasteofplastic.askyblock.panels.SettingsPanel;
 import com.wasteofplastic.askyblock.panels.WarpPanel;
-import com.wasteofplastic.askyblock.util.SpawnEgg1_9;
+import com.wasteofplastic.askyblock.util.HeadGetter;
 import com.wasteofplastic.askyblock.util.Util;
 import com.wasteofplastic.askyblock.util.VaultHelper;
 
@@ -98,13 +79,12 @@ import com.wasteofplastic.askyblock.util.VaultHelper;
  *         Main ASkyBlock class - provides an island minigame in a sea of acid
  */
 public class ASkyBlock extends JavaPlugin {
+    private static final boolean DEBUG = false;
     // This plugin
     private static ASkyBlock plugin;
     // The ASkyBlock world
     private static World islandWorld = null;
     private static World netherWorld = null;
-    // No push scoreboard name
-    private final static String NO_PUSH_TEAM_NAME = "ASkyBlockNP";
     // Flag indicating if a new islands is in the process of being generated or
     // not
     private boolean newIsland = false;
@@ -112,8 +92,6 @@ public class ASkyBlock extends JavaPlugin {
     private File playersFolder;
     // Challenges object
     private Challenges challenges;
-    // Localization Strings
-    private HashMap<String,ASLocale> availableLocales = new HashMap<String,ASLocale>();
     // Players object
     private PlayerCache players;
     // Listeners
@@ -134,11 +112,6 @@ public class ASkyBlock extends JavaPlugin {
     // V1.8 or later
     private boolean onePointEight;
 
-    private boolean debug = false;
-
-    // Level calc
-    private boolean calculatingLevel = false;
-
     // Update object
     private Updater updateCheck = null;
 
@@ -153,6 +126,22 @@ public class ASkyBlock extends JavaPlugin {
 
     // Settings panel object
     private SettingsPanel settingsPanel;
+
+    // Acid Item Removal Task
+    private AcidTask acidTask;
+
+    // Player events listener
+    private PlayerEvents playerEvents;
+
+    // Metrics
+    private Metrics metrics;
+
+    // Localization Strings
+    private Map<String,ASLocale> availableLocales = new HashMap<>();
+
+    // Head getter
+    private HeadGetter headGetter;
+    private EntityLimits entityLimits;
 
     /**
      * Returns the World object for the island world named in config.yml.
@@ -176,30 +165,12 @@ public class ASkyBlock extends JavaPlugin {
                 getNetherWorld();
             }
             // Multiverse configuration
-
             if (!Settings.useOwnGenerator && Bukkit.getServer().getPluginManager().isPluginEnabled("Multiverse-Core")) {
-                Bukkit.getLogger().info("Trying to register generator with Multiverse ");
-                try {
-                    Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(),
-                            "mv import " + Settings.worldName + " normal -g " + plugin.getName());
-                    if (!Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(),
-                            "mv modify set generator " + plugin.getName() + " " + Settings.worldName)) {
-                        Bukkit.getLogger().severe("Multiverse is out of date! - Upgrade to latest version!");
-                    }
-                    if (Settings.createNether) {
-                        if (Settings.newNether) {
-                            Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(),
-                                    "mv import " + Settings.worldName + "_nether nether -g " + plugin.getName());
-                            Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(),
-                                    "mv modify set generator " + plugin.getName() + " " + Settings.worldName + "_nether");
-                        } else {
-                            Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(), "mv import " + Settings.worldName + "_nether nether");
-                        }
-                    }
-                } catch (Exception e) {
-                    Bukkit.getLogger().severe("Not successfull! Disabling " + plugin.getName() + "!");
-                    e.printStackTrace();
-                    Bukkit.getServer().getPluginManager().disablePlugin(plugin);
+                // Run sync
+                if (!Bukkit.isPrimaryThread()) {
+                    Bukkit.getScheduler().runTask(plugin, ASkyBlock::registerMultiverse);
+                } else {
+                    registerMultiverse();
                 }
             }
 
@@ -212,6 +183,35 @@ public class ASkyBlock extends JavaPlugin {
         }
 
         return islandWorld;
+    }
+
+    private static void registerMultiverse() {
+        Bukkit.getLogger().info("Trying to register generator with Multiverse ");
+        try {
+            Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(),
+                    "mv import " + Settings.worldName + " normal -g " + plugin.getName());
+            if (!Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(),
+                    "mv modify set generator " + plugin.getName() + " " + Settings.worldName)) {
+                Bukkit.getLogger().severe("Multiverse is out of date! - Upgrade to latest version!");
+            }
+
+            if (Settings.createNether) {
+                if (Settings.newNether) {
+
+                    Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(),
+
+                            "mv import " + Settings.worldName + "_nether nether -g " + plugin.getName());
+                    Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(),
+                            "mv modify set generator " + plugin.getName() + " " + Settings.worldName + "_nether");
+
+                } else {
+                    Bukkit.getServer().dispatchCommand(Bukkit.getServer().getConsoleSender(), "mv import " + Settings.worldName + "_nether nether");
+                }
+            }
+        } catch (Exception e) {
+            Bukkit.getLogger().severe("Not successfull! Disabling " + plugin.getName() + "!");
+            Bukkit.getServer().getPluginManager().disablePlugin(plugin);
+        }
     }
 
     /**
@@ -237,22 +237,34 @@ public class ASkyBlock extends JavaPlugin {
             }
             // Save the warps and do not reload the panel
             if (warpSignsListener != null) {
-                warpSignsListener.saveWarpList();
+                warpSignsListener.saveWarpList(false);
             }
             if (messages != null) {
-                messages.saveMessages();
+                messages.saveMessages(false);
             }
-            TopTen.topTenSave();
+            if (topTen != null) {
+                topTen.topTenSave();
+            }
             // Close the name database
             if (tinyDB != null) {
                 tinyDB.saveDB();
             }
             // Save the coops
             CoopPlay.getInstance().saveCoops();
+            // Remove temporary perms
+            if (playerEvents != null) {
+                playerEvents.removeAllTempPerms();
+            }
+            // Save entitiy limits
+            if (entityLimits != null) {
+                entityLimits.disable();
+            }
         } catch (final Exception e) {
             getLogger().severe("Something went wrong saving files!");
             e.printStackTrace();
         }
+
+        metrics = null;
     }
 
     /*
@@ -263,6 +275,8 @@ public class ASkyBlock extends JavaPlugin {
     public void onEnable() {
         // instance of this plugin
         plugin = this;
+        // Initialize the API
+        new ASkyBlockAPI(this);
         // Check server version - check for a class that only 1.8 has
         Class<?> clazz;
         try {
@@ -296,7 +310,7 @@ public class ASkyBlock extends JavaPlugin {
             return;
         }
         // Load all the configuration of the plugin and localization strings
-        if (!loadPluginConfig()) {
+        if (!PluginConfig.loadPluginConfig(this)) {
             // Currently, the only setup error is where the world_name does not match
             if (Settings.GAMETYPE.equals(Settings.GameType.ASKYBLOCK)) {
                 getCommand("island").setExecutor(new NotSetup(Reason.WORLD_NAME));
@@ -318,7 +332,6 @@ public class ASkyBlock extends JavaPlugin {
             getServer().getPluginManager().disablePlugin(this);
             return;
         }
-
         // This can no longer be run in onEnable because the plugin is loaded at
         // startup and so key variables are
         // not known to the server. Instead it is run one tick after startup.
@@ -338,6 +351,8 @@ public class ASkyBlock extends JavaPlugin {
         if (!playersFolder.exists()) {
             playersFolder.mkdir();
         }
+        if (DEBUG)
+            Bukkit.getLogger().info("DEBUG: Setting up player cache");
         players = new PlayerCache(this);
         // Set up commands for this plugin
         islandCmd = new IslandCmd(this);
@@ -370,15 +385,12 @@ public class ASkyBlock extends JavaPlugin {
         messages = new Messages(this);
         messages.loadMessages();
         // Register world load event
-        if (getServer().getVersion().contains("(MC: 1.8") || plugin.getServer().getVersion().contains("(MC: 1.7")) {
+        if (getServer().getVersion().contains("(MC: 1.8") || getServer().getVersion().contains("(MC: 1.7")) {
             getServer().getPluginManager().registerEvents(new WorldLoader(this), this);
         }
         // Metrics
-        try {
-            final Metrics metrics = new Metrics(this);
-            metrics.start();
-        } catch (final IOException localIOException) {
-        }
+        metrics = new Metrics(this);
+
         // Kick off a few tasks on the next tick
         // By calling getIslandWorld(), if there is no island
         // world, it will be created
@@ -390,14 +402,14 @@ public class ASkyBlock extends JavaPlugin {
                 getIslandWorld();
                 if (!Settings.useOwnGenerator && getServer().getWorld(Settings.worldName).getGenerator() == null) {
                     // Check if the world generator is registered correctly
-                    getLogger().severe("********* The Generator for " + plugin.getName() + " is not registered so the plugin cannot start ********");
+                    getLogger().severe("********* The Generator for " + ASkyBlock.this.getName() + " is not registered so the plugin cannot start ********");
                     getLogger().severe("If you are using your own generator or server.properties, set useowngenerator: true in config.yml");
                     getLogger().severe("Otherwise:");
                     getLogger().severe("Make sure you have the following in bukkit.yml (case sensitive):");
                     getLogger().severe("worlds:");
                     getLogger().severe("  # The next line must be the name of your world:");
                     getLogger().severe("  " + Settings.worldName + ":");
-                    getLogger().severe("    generator: " + plugin.getName());
+                    getLogger().severe("    generator: " + ASkyBlock.this.getName());
                     if (Settings.GAMETYPE.equals(Settings.GameType.ASKYBLOCK)) {
                         getCommand("island").setExecutor(new NotSetup(Reason.GENERATOR));
                         getCommand("asc").setExecutor(new NotSetup(Reason.GENERATOR));
@@ -407,74 +419,82 @@ public class ASkyBlock extends JavaPlugin {
                         getCommand("aic").setExecutor(new NotSetup(Reason.GENERATOR));
                         getCommand("acid").setExecutor(new NotSetup(Reason.GENERATOR));
                     }
-                    HandlerList.unregisterAll(plugin);
+                    HandlerList.unregisterAll(ASkyBlock.this);
                     return;
                 }
-                // Try to register Herochat
-                if (Bukkit.getServer().getPluginManager().isPluginEnabled("Herochat")) {
+
+                // Run game rule to keep things quiet
+                if (Settings.silenceCommandFeedback){
                     try {
-                        getServer().getPluginManager().registerEvents(new HeroChatListener(plugin), plugin);
-                    } catch (Exception e) {
-                        plugin.getLogger().severe("Could not register with Herochat");
-                    }
+                        getLogger().info("Silencing command feedback for Ops...");
+                        getServer().dispatchCommand(getServer().getConsoleSender(), "minecraft:gamerule sendCommandFeedback false");
+                        getLogger().info("If you do not want this, do /gamerule sendCommandFeedback true");
+                    } catch (Exception ignored) {} // do nothing
                 }
+
                 // Run these one tick later to ensure worlds are loaded.
-                getServer().getScheduler().runTask(plugin, new Runnable() {
-                    @Override
-                    public void run() {
-                        // load the list - order matters - grid first, then top
-                        // ten to optimize upgrades
-                        // Load grid
-                        if (grid == null) {
-                            grid = new GridManager(plugin);
-                        }
-                        // Register events
-                        registerEvents();
-
-                        // Load TinyDb
-                        if (tinyDB == null) {
-                            tinyDB = new TinyDB(plugin);
-                        }
-                        // Load warps
-                        getWarpSignsListener().loadWarpList();
-                        // Load the warp panel
-                        if (Settings.useWarpPanel) {
-                            warpPanel = new WarpPanel(plugin);
-                            getServer().getPluginManager().registerEvents(warpPanel, plugin);
-                        }						
-                        // Load the TopTen GUI
-                        if (!Settings.displayIslandTopTenInChat){
-                            topTen = new TopTen(plugin);
-                            getServer().getPluginManager().registerEvents(topTen, plugin);
-                        }
-                        // Minishop - must wait for economy to load before we can use
-                        // econ
-                        getServer().getPluginManager().registerEvents(new ControlPanel(plugin), plugin);
-                        // Settings
-                        settingsPanel = new SettingsPanel(plugin);
-                        getServer().getPluginManager().registerEvents(settingsPanel, plugin);
-                        // Biomes
-                        // Load Biomes
-                        biomes = new BiomesPanel(plugin);
-                        getServer().getPluginManager().registerEvents(biomes, plugin);
-
-                        TopTen.topTenLoad();
-
-                        // Add any online players to the DB
-                        for (Player onlinePlayer : plugin.getServer().getOnlinePlayers()) {
-                            tinyDB.savePlayerName(onlinePlayer.getName(), onlinePlayer.getUniqueId());
-                        }
-                        if (Settings.backupDuration > 0) {
-                            new AsyncBackup(plugin);
-                        }
-                        // Load the coops
-                        if (Settings.persistantCoops) {
-                            CoopPlay.getInstance().loadCoops();
-                        }
-                        getLogger().info("All files loaded. Ready to play...");
-                        // Fire event
-                        getServer().getPluginManager().callEvent(new ReadyEvent());
+                getServer().getScheduler().runTask(ASkyBlock.this, () -> {
+                    // load the list - order matters - grid first, then top
+                    // ten to optimize upgrades
+                    // Load grid
+                    if (grid == null) {
+                        grid = new GridManager(ASkyBlock.this);
                     }
+                    // Register events
+                    registerEvents();
+
+                    // Load TinyDb
+                    if (tinyDB == null) {
+                        tinyDB = new TinyDB(ASkyBlock.this);
+                    }
+                    // Run head getter
+                    headGetter = new HeadGetter(plugin);
+
+                    // Load warps
+                    getWarpSignsListener().loadWarpList();
+                    // Load the warp panel
+                    if (Settings.useWarpPanel) {
+                        plugin.getLogger().info("Loading warp panel...");
+                        warpPanel = new WarpPanel(ASkyBlock.this);
+                        getServer().getPluginManager().registerEvents(warpPanel, ASkyBlock.this);
+                    }
+                    topTen = new TopTen(ASkyBlock.this);
+                    // Load the TopTen GUI
+                    if (!Settings.displayIslandTopTenInChat){
+                        getServer().getPluginManager().registerEvents(topTen, ASkyBlock.this);
+                    }
+                    // Minishop - must wait for economy to load before we can use
+                    // econ
+                    getServer().getPluginManager().registerEvents(new ControlPanel(ASkyBlock.this), ASkyBlock.this);
+                    // Settings
+                    settingsPanel = new SettingsPanel(ASkyBlock.this);
+                    getServer().getPluginManager().registerEvents(settingsPanel, ASkyBlock.this);
+                    // Biomes
+                    // Load Biomes
+                    biomes = new BiomesPanel(ASkyBlock.this);
+                    getServer().getPluginManager().registerEvents(biomes, ASkyBlock.this);
+
+                    // Add any online players to the DB
+                    for (Player onlinePlayer : ASkyBlock.this.getServer().getOnlinePlayers()) {
+                        tinyDB.savePlayerName(onlinePlayer.getName(), onlinePlayer.getUniqueId());
+                    }
+                    if (Settings.backupDuration > 0) {
+                        new AsyncBackup(ASkyBlock.this);
+                    }
+                    // Load the coops
+                    if (Settings.persistantCoops) {
+                        CoopPlay.getInstance().loadCoops();
+                    }
+                    // Give temp permissions
+                    playerEvents.giveAllTempPerms();
+
+                    getLogger().info("All files loaded. Ready to play...");
+
+                    registerCustomCharts();
+                    getLogger().info("Metrics loaded.");
+
+                    // Fire event
+                    getServer().getPluginManager().callEvent(new ReadyEvent());
                 });
                 // Check for updates asynchronously
                 if (Settings.updateCheck) {
@@ -484,38 +504,38 @@ public class ASkyBlock extends JavaPlugin {
                         @Override
                         public void run() {
                             if (count++ > 20) {
-                                plugin.getLogger().info("No updates found. (No response from server after 20s)");
+                                ASkyBlock.this.getLogger().info("No updates found. (No response from server after 20s)");
                                 this.cancel();
                             } else {
                                 // Wait for the response
                                 if (updateCheck != null) {
                                     switch (updateCheck.getResult()) {
                                     case DISABLED:
-                                        plugin.getLogger().info("Updating has been disabled");
+                                        ASkyBlock.this.getLogger().info("Updating has been disabled");
                                         break;
                                     case FAIL_APIKEY:
-                                        plugin.getLogger().info("API key failed");
+                                        ASkyBlock.this.getLogger().info("API key failed");
                                         break;
                                     case FAIL_BADID:
-                                        plugin.getLogger().info("Bad ID");
+                                        ASkyBlock.this.getLogger().info("Bad ID");
                                         break;
                                     case FAIL_DBO:
-                                        plugin.getLogger().info("Could not connect to updating service");
+                                        ASkyBlock.this.getLogger().info("Could not connect to updating service");
                                         break;
                                     case FAIL_DOWNLOAD:
-                                        plugin.getLogger().info("Downloading failed");
+                                        ASkyBlock.this.getLogger().info("Downloading failed");
                                         break;
                                     case FAIL_NOVERSION:
-                                        plugin.getLogger().info("Could not recognize version");
+                                        ASkyBlock.this.getLogger().info("Could not recognize version");
                                         break;
                                     case NO_UPDATE:
-                                        plugin.getLogger().info("No update available.");
+                                        ASkyBlock.this.getLogger().info("No update available.");
                                         break;
                                     case SUCCESS:
-                                        plugin.getLogger().info("Success!");
+                                        ASkyBlock.this.getLogger().info("Success!");
                                         break;
                                     case UPDATE_AVAILABLE:
-                                        plugin.getLogger().info("Update available " + updateCheck.getLatestName());
+                                        ASkyBlock.this.getLogger().info("Update available " + updateCheck.getLatestName());
                                         break;
                                     default:
                                         break;
@@ -525,42 +545,12 @@ public class ASkyBlock extends JavaPlugin {
                                 }
                             }
                         }
-                    }.runTaskTimer(plugin, 0L, 20L); // Check status every second
+                    }.runTaskTimer(ASkyBlock.this, 0L, 20L); // Check status every second
                 }
-                // This part will kill monsters if they fall into the water
-                // because it
-                // is acid
-                if (Settings.mobAcidDamage > 0D || Settings.animalAcidDamage > 0D) {
-                    getServer().getScheduler().scheduleSyncRepeatingTask(plugin, new Runnable() {
-                        @Override
-                        public void run() {
-                            List<Entity> entList = islandWorld.getEntities();
-                            for (Entity current : entList) {
-                                if (plugin.isOnePointEight() && current instanceof Guardian) {
-                                    // Guardians are immune to acid too
-                                    continue;
-                                }
-                                if ((current instanceof Monster) && Settings.mobAcidDamage > 0D) {
-                                    if ((current.getLocation().getBlock().getType() == Material.WATER)
-                                            || (current.getLocation().getBlock().getType() == Material.STATIONARY_WATER)) {
-                                        ((Monster) current).damage(Settings.mobAcidDamage);
-                                        // getLogger().info("Killing monster");
-                                    }
-                                } else if ((current instanceof Animals) && Settings.animalAcidDamage > 0D) {
-                                    if ((current.getLocation().getBlock().getType() == Material.WATER)
-                                            || (current.getLocation().getBlock().getType() == Material.STATIONARY_WATER)) {
-                                        if (!current.getType().equals(EntityType.CHICKEN)) {
-                                            ((Animals) current).damage(Settings.animalAcidDamage);
-                                        } else if (Settings.damageChickens) {
-                                            ((Animals) current).damage(Settings.animalAcidDamage);
-                                        }
-                                        // getLogger().info("Killing animal");
-                                    }
-                                }
-                            }
-                        }
-                    }, 0L, 20L);
-                }
+
+                // Run acid tasks
+                acidTask = new AcidTask(ASkyBlock.this);
+
             }
         });
     }
@@ -576,9 +566,9 @@ public class ASkyBlock extends JavaPlugin {
             @Override
             public void run() {
                 if (Settings.GAMETYPE.equals(GameType.ASKYBLOCK)) {
-                    updateCheck = new Updater(plugin, 85189, plugin.getFile(), UpdateType.NO_DOWNLOAD, true); // ASkyBlock
+                    updateCheck = new Updater(ASkyBlock.this, 85189, ASkyBlock.this.getFile(), UpdateType.NO_DOWNLOAD, true); // ASkyBlock
                 } else {
-                    updateCheck = new Updater(plugin, 80095, plugin.getFile(), UpdateType.NO_DOWNLOAD, true); // AcidIsland
+                    updateCheck = new Updater(ASkyBlock.this, 80095, ASkyBlock.this.getFile(), UpdateType.NO_DOWNLOAD, true); // AcidIsland
                 }                
             }
         });
@@ -588,9 +578,9 @@ public class ASkyBlock extends JavaPlugin {
         if (updateCheck != null) {
             if (updateCheck.getResult().equals(UpdateResult.UPDATE_AVAILABLE)) {
                 // Player login
-                p.sendMessage(ChatColor.GOLD + updateCheck.getLatestName() + " is available! You are running V" + getDescription().getVersion());
-                p.sendMessage(ChatColor.RED + "Update at:");
-                p.sendMessage(ChatColor.RED + getUpdateCheck().getLatestFileLink());
+                Util.sendMessage(p, ChatColor.GOLD + updateCheck.getLatestName() + " is available! You are running V" + getDescription().getVersion());
+                Util.sendMessage(p, ChatColor.RED + "Update at:");
+                Util.sendMessage(p, ChatColor.RED + getUpdateCheck().getLatestFileLink());
             }
         }
 
@@ -612,6 +602,7 @@ public class ASkyBlock extends JavaPlugin {
         getWarpSignsListener().removeWarp(player);
         Island island = grid.getIsland(player);
         if (island != null) {
+            getServer().getPluginManager().callEvent(new IslandPreDeleteEvent(player, island));
             if (removeBlocks) {
                 grid.removePlayersFromIsland(island, player);
                 new DeleteIslandChunk(this, island);
@@ -695,850 +686,10 @@ public class ASkyBlock extends JavaPlugin {
     }
 
     /**
-     * @return the calculatingLevel
-     */
-    public boolean isCalculatingLevel() {
-        return calculatingLevel;
-    }
-
-    /**
      * @return the newIsland
      */
     public boolean isNewIsland() {
         return newIsland;
-    }
-
-    /**
-     * Loads the various settings from the config.yml file into the plugin
-     */
-    @SuppressWarnings("deprecation")
-    public boolean loadPluginConfig() {
-        // getLogger().info("*********************************************");
-        try {
-            getConfig();
-        } catch (final Exception e) {
-            e.printStackTrace();
-        }
-        //CompareConfigs.compareConfigs();
-        // Get the localization strings
-        // Look in the locale folder. If it is not there, then 
-        //getLocale();
-        // Add this to the config      
-        FileLister fl = new FileLister(this);
-        try {
-            int index = 1;
-            for (String code: fl.list()) {
-                availableLocales.put(code, new ASLocale(this,code, index++)); 
-            }
-        } catch (IOException e1) {
-            getLogger().severe("Could not add locales!");
-        }
-        // Default is locale.yml
-        availableLocales.put("locale", new ASLocale(this, "locale", 0));
-        // Assign settings
-        String configVersion = getConfig().getString("general.version", "");
-        //getLogger().info("DEBUG: config ver length " + configVersion.split("\\.").length);
-        // Ignore last digit if it is 4 digits long
-        if (configVersion.split("\\.").length == 4) {
-            configVersion = configVersion.substring(0, configVersion.lastIndexOf('.')); 
-        }
-        // Save for plugin version
-        String version = plugin.getDescription().getVersion();
-        //getLogger().info("DEBUG: version length " + version.split("\\.").length);
-        if (version.split("\\.").length == 4) {
-            version = version.substring(0, version.lastIndexOf('.')); 
-        }
-        if (configVersion.isEmpty() || !configVersion.equalsIgnoreCase(version)) {
-            // Check to see if this has already been shared
-            File newConfig = new File(plugin.getDataFolder(),"config.new.yml");
-            getLogger().warning("***********************************************************");
-            getLogger().warning("Config file is out of date. See config.new.yml for updates!");
-            getLogger().warning("config.yml version is '" + configVersion + "'");
-            getLogger().warning("Latest config version is '" + version + "'");
-            getLogger().warning("***********************************************************");
-            if (!newConfig.exists()) {
-                File oldConfig = new File(plugin.getDataFolder(),"config.yml");
-                File bakConfig = new File(plugin.getDataFolder(),"config.bak");
-                if (oldConfig.renameTo(bakConfig)) {
-                    plugin.saveResource("config.yml", false);
-                    oldConfig.renameTo(newConfig);
-                    bakConfig.renameTo(oldConfig);
-                } 
-            }
-        }
-        // Recover superflat
-        Settings.recoverSuperFlat = getConfig().getBoolean("general.recoversuperflat");
-        if (Settings.recoverSuperFlat) {
-            getLogger().warning("*********************************************************");
-            getLogger().warning("WARNING: Recover super flat mode is enabled");
-            getLogger().warning("This will regenerate any chunks with bedrock at y=0 when they are loaded");
-            getLogger().warning("Switch off when superflat chunks are cleared");
-            getLogger().warning("You should back up your world before running this");
-            getLogger().warning("*********************************************************");
-        }
-        // Hack skeleton spawners for 1.11
-        Settings.hackSkeletonSpawners = getConfig().getBoolean("schematicsection.hackskeletonspawners", true);
-        // Allow Obsidian Scooping
-        Settings.allowObsidianScooping = getConfig().getBoolean("general.allowobsidianscooping", true);
-        // Kicked players keep inventory
-        Settings.kickedKeepInv = getConfig().getBoolean("general.kickedkeepinv", false);
-        // Chat prefixes
-        Settings.chatLevelPrefix = getConfig().getString("general.chatlevelprefix","{ISLAND_LEVEL}");
-        Settings.chatChallengeLevelPrefix = getConfig().getString("general.chatchallanegelevelprefix","{ISLAND_CHALLENGE_LEVEL}");
-        Settings.chatIslandPlayer = getConfig().getString("general.chatislandplayer","{ISLAND_PLAYER}");
-        // Nether roof option
-        Settings.netherRoof = getConfig().getBoolean("general.netherroof", true);
-        // FTB Autoamtic Activators
-        Settings.allowAutoActivator = getConfig().getBoolean("general.autoactivator");
-        // Debug
-        Settings.debug = getConfig().getInt("debug", 0);
-        // Persistent coops
-        Settings.persistantCoops = getConfig().getBoolean("general.persistentcoops");
-        // Level logging
-        Settings.levelLogging = getConfig().getBoolean("general.levellogging");
-        // Allow pushing
-        Settings.allowPushing = getConfig().getBoolean("general.allowpushing", true);
-        // try to remove the team from the scoreboard
-        if (Settings.allowPushing) {
-            try {
-                Scoreboard scoreboard = getServer().getScoreboardManager().getMainScoreboard();
-                if (scoreboard != null) {
-                    Team pTeam = scoreboard.getTeam(NO_PUSH_TEAM_NAME);
-                    if (pTeam != null) {
-                        pTeam.unregister();
-                    }
-                }
-            } catch (Exception e) {
-                getLogger().warning("Problem removing no push from scoreboard.");
-            }
-        }
-        // Custom generator
-        Settings.useOwnGenerator = getConfig().getBoolean("general.useowngenerator", false);
-        // How often the grid will be saved to file. Default is 5 minutes
-        Settings.backupDuration = (getConfig().getLong("general.backupduration", 5) * 20 * 60);
-        // How long a player has to wait after deactivating PVP until they can activate PVP again
-        Settings.pvpRestartCooldown = getConfig().getLong("general.pvpcooldown",60);
-        // Max Islands
-        Settings.maxIslands = getConfig().getInt("general.maxIslands",0);
-        // Mute death messages
-        Settings.muteDeathMessages = getConfig().getBoolean("general.mutedeathmessages", false);
-        // Warp Restriction
-        Settings.warpLevelsRestriction = getConfig().getInt("general.warplevelrestriction", 10);
-        // Warp panel
-        Settings.useWarpPanel = getConfig().getBoolean("general.usewarppanel", true);
-        // Fast level calculation (this is really fast)
-        Settings.fastLevelCalc = getConfig().getBoolean("general.fastlevelcalc", true);
-        // Restrict wither
-        Settings.restrictWither = getConfig().getBoolean("general.restrictwither", true);
-        // Team chat
-        Settings.teamChat = getConfig().getBoolean("general.teamchat", true);
-        Settings.logTeamChat = getConfig().getBoolean("general.logteamchat", true);
-        // TEAMSUFFIX as island level
-        Settings.setTeamName = getConfig().getBoolean("general.setteamsuffix", false);
-        Settings.teamSuffix = getConfig().getString("general.teamsuffix","([level])");
-        // Immediate teleport
-        Settings.immediateTeleport = getConfig().getBoolean("general.immediateteleport", false);
-        // Make island automatically
-        Settings.makeIslandIfNone = getConfig().getBoolean("general.makeislandifnone", false);
-        // Use physics when pasting island block schematics
-        Settings.usePhysics = getConfig().getBoolean("general.usephysics", false);
-        // Use old display (chat instead of GUI) for Island top ten
-        Settings.displayIslandTopTenInChat = getConfig().getBoolean("general.islandtopteninchat", false);
-        // Run level calc at login
-        Settings.loginLevel = getConfig().getBoolean("general.loginlevel", false);
-        // Use economy or not
-        // In future expand to include internal economy
-        Settings.useEconomy = getConfig().getBoolean("general.useeconomy", true);
-        // Use the minishop or not
-        Settings.useMinishop = getConfig().getBoolean("general.useminishop", true);
-        // Check for updates
-        Settings.updateCheck = getConfig().getBoolean("general.checkupdates", true);
-        // Island reset commands
-        Settings.resetCommands = getConfig().getStringList("general.resetcommands");
-        Settings.leaveCommands = getConfig().getStringList("general.leavecommands");
-        Settings.startCommands = getConfig().getStringList("general.startcommands");
-        Settings.teamStartCommands = getConfig().getStringList("general.teamstartcommands");
-        Settings.useControlPanel = getConfig().getBoolean("general.usecontrolpanel", false);
-        // Check if /island command is allowed when falling
-        Settings.allowTeleportWhenFalling = getConfig().getBoolean("general.allowfallingteleport", true);
-        Settings.fallingCommandBlockList = getConfig().getStringList("general.blockingcommands");
-        // Visitor command banned list
-        Settings.visitorCommandBlockList = getConfig().getStringList("general.visitorbannedcommands");
-        // Max team size
-        Settings.maxTeamSize = getConfig().getInt("island.maxteamsize", 4);
-        // Deprecated settings - use permission askyblock.team.maxsize.<number> instead
-        Settings.maxTeamSizeVIP = getConfig().getInt("island.vipteamsize", 0);
-        Settings.maxTeamSizeVIP2 = getConfig().getInt("island.vip2teamsize", 0);
-        if (Settings.maxTeamSizeVIP > 0 || Settings.maxTeamSizeVIP2 > 0) {
-            getLogger().warning(Settings.PERMPREFIX + "team.vip and " + Settings.PERMPREFIX + "team.vip2 are deprecated!");
-            getLogger().warning("Use permission " + Settings.PERMPREFIX + "team.maxsize.<number> instead.");
-        }
-        // Max home number
-        Settings.maxHomes = getConfig().getInt("general.maxhomes",1);
-        if (Settings.maxHomes < 1) {
-            Settings.maxHomes = 1;
-        }
-        // Settings from config.yml
-        Settings.worldName = getConfig().getString("general.worldName");
-        // Check if the world name matches island.yml info
-        File islandFile = new File(plugin.getDataFolder(), "islands.yml");
-        if (islandFile.exists()) {
-            YamlConfiguration islandYaml = new YamlConfiguration();
-            try {
-                islandYaml.load(islandFile);
-                if (!islandYaml.contains(Settings.worldName)) {
-                    // Bad news, stop everything and tell the admin
-                    getLogger().severe("+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+");
-                    getLogger().severe("More set up is required. Go to config.yml and edit it.");
-                    getLogger().severe("");
-                    getLogger().severe("Check island world name is same as world in islands.yml.");
-                    getLogger().severe("If you are resetting and changing world, delete island.yml and restart.");
-                    getLogger().severe("");
-                    getLogger().severe("+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+");
-                    return false;
-                }
-            } catch (Exception e) {}        
-        }
-        Settings.createNether = getConfig().getBoolean("general.createnether", true);
-        if (!Settings.createNether) {
-            getLogger().info("The Nether is disabled");
-        }
-
-        String companion = getConfig().getString("island.companion", "COW").toUpperCase();
-        Settings.islandCompanion = null;
-        if (!companion.equalsIgnoreCase("NOTHING")) {
-            String commaList = "NOTHING, ";
-            for (EntityType type : EntityType.values()) {
-                if (companion.equalsIgnoreCase(type.toString())) {
-                    Settings.islandCompanion = type;
-                    break;
-                }
-                commaList += ", " + type.toString();
-            }
-            if (Settings.islandCompanion == null) {
-                getLogger().warning("Island companion is not recognized. Pick from " + commaList);
-            }
-        }
-        // Companion names
-        List<String> companionNames = getConfig().getStringList("island.companionnames");
-        Settings.companionNames = new ArrayList<String>();
-        for (String name : companionNames) {
-            Settings.companionNames.add(ChatColor.translateAlternateColorCodes('&', name));
-        }
-        Settings.islandDistance = getConfig().getInt("island.distance", 200);
-        if (Settings.islandDistance < 50) {
-            Settings.islandDistance = 50;
-            getLogger().info("Setting minimum island distance to 50");
-        }
-        Settings.islandXOffset = getConfig().getInt("island.xoffset", 0);
-        if (Settings.islandXOffset < 0) {
-            Settings.islandXOffset = 0;
-            getLogger().info("Setting minimum island X Offset to 0");
-        } else if (Settings.islandXOffset > Settings.islandDistance) {
-            Settings.islandXOffset = Settings.islandDistance;
-            getLogger().info("Setting maximum island X Offset to " + Settings.islandDistance);
-        }
-        Settings.islandZOffset = getConfig().getInt("island.zoffset", 0);
-        if (Settings.islandZOffset < 0) {
-            Settings.islandZOffset = 0;
-            getLogger().info("Setting minimum island Z Offset to 0");
-        } else if (Settings.islandZOffset > Settings.islandDistance) {
-            Settings.islandZOffset = Settings.islandDistance;
-            getLogger().info("Setting maximum island Z Offset to " + Settings.islandDistance);
-        }
-        long x = getConfig().getLong("island.startx", 0);
-        // Check this is a multiple of island distance
-        long z = getConfig().getLong("island.startz", 0);
-        Settings.islandStartX = Math.round((double) x / Settings.islandDistance) * Settings.islandDistance + Settings.islandXOffset;
-        Settings.islandStartZ = Math.round((double) z / Settings.islandDistance) * Settings.islandDistance + Settings.islandZOffset;
-
-        // ASkyBlock and AcidIsland difference
-        if (Settings.GAMETYPE.equals(Settings.GameType.ACIDISLAND)) {
-            Settings.acidDamage = getConfig().getDouble("general.aciddamage", 5D);
-            if (Settings.acidDamage > 100D) {
-                Settings.acidDamage = 100D;
-            } else if (Settings.acidDamage < 0D) {
-                Settings.acidDamage = 0D;
-            }
-            Settings.mobAcidDamage = getConfig().getDouble("general.mobaciddamage", 10D);
-            if (Settings.mobAcidDamage > 100D) {
-                Settings.mobAcidDamage = 100D;
-            } else if (Settings.mobAcidDamage < 0D) {
-                Settings.mobAcidDamage = 0D;
-            }
-            Settings.rainDamage = getConfig().getDouble("general.raindamage", 0.5D);
-            if (Settings.rainDamage > 100D) {
-                Settings.rainDamage = 100D;
-            } else if (Settings.rainDamage < 0D) {
-                Settings.rainDamage = 0D;
-            }
-            // The island's center is actually 5 below sea level
-            Settings.sea_level = getConfig().getInt("general.sealevel", 50);
-            if (Settings.sea_level < 0) {
-                Settings.sea_level = 0;
-            }
-            Settings.island_level = getConfig().getInt("general.islandlevel", 50) - 5;
-            if (Settings.island_level < 0) {
-                Settings.island_level = 0;
-            }
-        } else {
-            Settings.acidDamage = getConfig().getDouble("general.aciddamage", 0D);
-            if (Settings.acidDamage > 100D) {
-                Settings.acidDamage = 100D;
-            } else if (Settings.acidDamage < 0D) {
-                Settings.acidDamage = 0D;
-            }
-            Settings.mobAcidDamage = getConfig().getDouble("general.mobaciddamage", 0D);
-            if (Settings.mobAcidDamage > 100D) {
-                Settings.mobAcidDamage = 100D;
-            } else if (Settings.mobAcidDamage < 0D) {
-                Settings.mobAcidDamage = 0D;
-            }
-            Settings.rainDamage = getConfig().getDouble("general.raindamage", 0D);
-            if (Settings.rainDamage > 100D) {
-                Settings.rainDamage = 100D;
-            } else if (Settings.rainDamage < 0D) {
-                Settings.rainDamage = 0D;
-            }
-            // The island's center is actually 5 below sea level
-            Settings.sea_level = getConfig().getInt("general.sealevel", 0);
-            if (Settings.sea_level < 0) {
-                Settings.sea_level = 0;
-            }
-            Settings.island_level = getConfig().getInt("general.islandlevel", 120) - 5;
-            if (Settings.island_level < 0) {
-                Settings.island_level = 0;
-            }
-        }
-        Settings.animalAcidDamage = getConfig().getDouble("general.animaldamage", 0D);
-        if (Settings.animalAcidDamage > 100D) {
-            Settings.animalAcidDamage = 100D;
-        } else if (Settings.animalAcidDamage < 0D) {
-            Settings.animalAcidDamage = 0D;
-        }
-        Settings.damageChickens = getConfig().getBoolean("general.damagechickens", false);
-        // Damage Type
-        List<String> acidDamageType = getConfig().getStringList("general.damagetype");
-        Settings.acidDamageType.clear();
-        if (acidDamageType != null) {
-            for (String effect : acidDamageType) {
-                PotionEffectType newPotionType = PotionEffectType.getByName(effect);
-                if (newPotionType != null) {
-                    // Check if it is a valid addition
-                    if (newPotionType.equals(PotionEffectType.BLINDNESS) || newPotionType.equals(PotionEffectType.CONFUSION)
-                            || newPotionType.equals(PotionEffectType.HUNGER) || newPotionType.equals(PotionEffectType.POISON)
-                            || newPotionType.equals(PotionEffectType.SLOW) || newPotionType.equals(PotionEffectType.SLOW_DIGGING)
-                            || newPotionType.equals(PotionEffectType.WEAKNESS)) {
-                        Settings.acidDamageType.add(newPotionType);
-                    }
-                } else {
-                    getLogger().warning("Could not interpret acid damage modifier: " + effect + " - skipping");
-                    getLogger().warning("Types can be : SLOW, SLOW_DIGGING, CONFUSION,");
-                    getLogger().warning("BLINDNESS, HUNGER, WEAKNESS and POISON");
-                }
-            }
-        }
-
-        Settings.animalSpawnLimit = getConfig().getInt("general.animalspawnlimit", 15);
-        if (Settings.animalSpawnLimit < -1) {
-            Settings.animalSpawnLimit = -1;
-        }
-
-        Settings.monsterSpawnLimit = getConfig().getInt("general.monsterspawnlimit", 100);
-        if (Settings.monsterSpawnLimit < -1) {
-            Settings.monsterSpawnLimit = -1;
-        }
-
-        Settings.waterAnimalSpawnLimit = getConfig().getInt("general.wateranimalspawnlimit", 15);
-        if (Settings.waterAnimalSpawnLimit < -1) {
-            Settings.waterAnimalSpawnLimit = -1;
-        }
-
-        Settings.abandonedIslandLevel = getConfig().getInt("general.abandonedislandlevel", 10);
-        if (Settings.abandonedIslandLevel < 0) {
-            Settings.abandonedIslandLevel = 0;
-        }
-
-        Settings.island_protectionRange = getConfig().getInt("island.protectionRange", 100);
-        if (Settings.island_protectionRange % 2 != 0) {
-            Settings.island_protectionRange--;
-            getLogger().warning("Protection range must be even, using " + Settings.island_protectionRange);
-        }
-        if (Settings.island_protectionRange > Settings.islandDistance) {
-            if (!getConfig().getBoolean("island.overridelimit", false)) {
-                if (Settings.island_protectionRange > (Settings.islandDistance - 16)) {
-                    Settings.island_protectionRange = Settings.islandDistance - 16;
-                    getLogger().warning(
-                            "*** Island protection range must be " + (Settings.islandDistance - 16) + " or less, (island range -16). Setting to: "
-                                    + Settings.island_protectionRange);
-                }
-            } else {
-                Settings.island_protectionRange = Settings.islandDistance;
-            }
-        }
-        if (Settings.island_protectionRange < 0) {
-            Settings.island_protectionRange = 0;
-        }
-        Settings.resetChallenges = getConfig().getBoolean("general.resetchallenges", true);
-        Settings.resetMoney = getConfig().getBoolean("general.resetmoney", true);
-        Settings.clearInventory = getConfig().getBoolean("general.resetinventory", true);
-        Settings.resetEnderChest = getConfig().getBoolean("general.resetenderchest", false);
-
-        Settings.startingMoney = getConfig().getDouble("general.startingmoney", 0D);
-        Settings.respawnOnIsland = getConfig().getBoolean("general.respawnonisland", false);
-        Settings.newNether = getConfig().getBoolean("general.newnether", true);
-        Settings.netherTrees = getConfig().getBoolean("general.nethertrees", true);
-        // Nether spawn protection radius
-        Settings.netherSpawnRadius = getConfig().getInt("general.netherspawnradius", 25);
-        if (Settings.netherSpawnRadius < 0) {
-            Settings.netherSpawnRadius = 0;
-        } else if (Settings.netherSpawnRadius > 100) {
-            Settings.netherSpawnRadius = 100;
-        }
-
-        Settings.resetWait = getConfig().getInt("general.resetwait", 300);
-        if (Settings.resetWait < 0) {
-            Settings.resetWait = 0;
-        }
-        Settings.resetLimit = getConfig().getInt("general.resetlimit", 0);
-        if (Settings.resetWait < 0) {
-            Settings.resetWait = -1;
-        }
-        Settings.inviteWait = getConfig().getInt("general.invitewait", 60);
-        if (Settings.inviteWait < 0) {
-            Settings.inviteWait = 0;
-        }
-        Settings.levelWait = getConfig().getInt("general.levelwait", 60);
-        if (Settings.levelWait < 0) {
-            Settings.levelWait = 0;
-        }
-        // Seconds to wait for a confirmation of reset
-        Settings.resetConfirmWait = getConfig().getInt("general.resetconfirmwait", 10);
-        if (Settings.resetConfirmWait < 0) {
-            Settings.resetConfirmWait = 0;
-        }
-        Settings.damageOps = getConfig().getBoolean("general.damageops", false);
-        // Invincible visitors
-        Settings.invincibleVisitors = getConfig().getBoolean("general.invinciblevisitors", false);
-        if(Settings.invincibleVisitors){
-            Settings.visitorDamagePrevention = new HashSet<DamageCause>();
-            List<String> damageSettings = getConfig().getStringList("general.invinciblevisitorsoptions");
-            for (DamageCause cause: DamageCause.values()) {
-                if (damageSettings.contains(cause.toString())) {
-                    Settings.visitorDamagePrevention.add(cause);
-                }
-            }
-        }
-
-        // Settings.ultraSafeBoats =
-        // getConfig().getBoolean("general.ultrasafeboats", true);
-        Settings.logInRemoveMobs = getConfig().getBoolean("general.loginremovemobs", true);
-        Settings.islandRemoveMobs = getConfig().getBoolean("general.islandremovemobs", false);
-        List<String> mobWhiteList = getConfig().getStringList("general.mobwhitelist");
-        Settings.mobWhiteList.clear();
-        String valid = "BLAZE, CREEPER, SKELETON, SPIDER, GIANT, ZOMBIE, GHAST, PIG_ZOMBIE, "
-                + "ENDERMAN, CAVE_SPIDER, SILVERFISH,  WITHER, WITCH, ENDERMITE,"
-                + " GUARDIAN";
-        for (String mobName : mobWhiteList) {
-            if (valid.contains(mobName.toUpperCase())) {
-                try {
-                    Settings.mobWhiteList.add(EntityType.valueOf(mobName.toUpperCase()));
-                } catch (Exception e) {
-                    plugin.getLogger().severe("Error in config.yml, mobwhitelist value '" + mobName + "' is invalid.");
-                    plugin.getLogger().severe("Possible values are : Blaze, Cave_Spider, Creeper, Enderman, Endermite, Giant, Guardian, "
-                            + "Pig_Zombie, Silverfish, Skeleton, Spider, Witch, Wither, Zombie");
-                }
-            } else {
-                plugin.getLogger().severe("Error in config.yml, mobwhitelist value '" + mobName + "' is invalid.");
-                plugin.getLogger().severe("Possible values are : Blaze, Cave_Spider, Creeper, Enderman, Endermite, Giant, Guardian, "
-                        + "Pig_Zombie, Silverfish, Skeleton, Spider, Witch, Wither, Zombie");
-            }
-        }
-        // getLogger().info("DEBUG: island level is " + Settings.island_level);
-        // Get chest items
-        String chestItems = getConfig().getString("island.chestItems","");
-        if (!chestItems.isEmpty()) {
-            final String[] chestItemString = chestItems.split(" ");
-            // getLogger().info("DEBUG: chest items = " + chestItemString);
-            final ItemStack[] tempChest = new ItemStack[chestItemString.length];
-            for (int i = 0; i < tempChest.length; i++) {
-                String[] amountdata = chestItemString[i].split(":");
-                try {
-                    if (amountdata.length == 3 && amountdata[0].equalsIgnoreCase("MONSTER_EGG")) {
-                        try {
-                            EntityType type = EntityType.valueOf(amountdata[1].toUpperCase());
-                            if (Bukkit.getServer().getVersion().contains("(MC: 1.8") || Bukkit.getServer().getVersion().contains("(MC: 1.7")) {
-                                tempChest[i] = new SpawnEgg(type).toItemStack(Integer.parseInt(amountdata[2]));
-                            } else {
-                                try {
-                                    tempChest[i] = new SpawnEgg1_9(type).toItemStack(Integer.parseInt(amountdata[2]));
-                                } catch (Exception ex) {
-                                    tempChest[i] = new ItemStack(Material.MONSTER_EGG);
-                                    plugin.getLogger().severe("Monster eggs not supported with this server version.");
-                                }
-                            }
-                        } catch (Exception e) {
-                            Bukkit.getLogger().severe("Spawn eggs must be described by name. Try one of these (not all are possible):");
-                            for (EntityType type : EntityType.values()) {
-                                if (type.isSpawnable() && type.isAlive()) {
-                                    plugin.getLogger().severe(type.toString());
-                                }
-                            }
-                        }
-                    } else if (amountdata[0].equals("POTION")) {
-                        // getLogger().info("DEBUG: Potion length " +
-                        // amountdata.length);
-                        if (amountdata.length == 6) {
-                            tempChest[i] = Challenges.getPotion(amountdata, Integer.parseInt(amountdata[5]), "config.yml");
-                        } else {
-                            getLogger().severe("Problem loading chest item from config.yml so skipping it: " + chestItemString[i]);
-                            getLogger().severe("Potions for the chest must be fully defined as POTION:NAME:<LEVEL>:<EXTENDED>:<SPLASH/LINGER>:QTY");
-                        }
-                    } else {
-                        Material mat;
-                        if (StringUtils.isNumeric(amountdata[0])) {
-                            mat = Material.getMaterial(Integer.parseInt(amountdata[0]));
-                        } else {
-                            mat = Material.getMaterial(amountdata[0].toUpperCase());
-                        }
-                        if (amountdata.length == 2) {
-                            tempChest[i] = new ItemStack(mat, Integer.parseInt(amountdata[1]));
-                        } else if (amountdata.length == 3) {
-                            tempChest[i] = new ItemStack(mat, Integer.parseInt(amountdata[2]), Short.parseShort(amountdata[1]));
-                        }
-                    }
-                } catch (java.lang.IllegalArgumentException ex) {
-                    ex.printStackTrace();
-                    getLogger().severe("Problem loading chest item from config.yml so skipping it: " + chestItemString[i]);
-                    getLogger().severe("Error is : " + ex.getMessage());
-                    getLogger().info("Potential potion types are: ");
-                    for (PotionType c : PotionType.values())
-                        getLogger().info(c.name());
-                } catch (Exception e) {
-                    e.printStackTrace();
-                    getLogger().severe("Problem loading chest item from config.yml so skipping it: " + chestItemString[i]);
-                    getLogger().info("Potential material types are: ");
-                    for (Material c : Material.values())
-                        getLogger().info(c.name());
-                    // e.printStackTrace();
-                }
-            }
-            Settings.chestItems = tempChest;
-        } else {
-            // Nothing in the chest
-            Settings.chestItems = new ItemStack[0];
-        }
-        Settings.allowPvP = getConfig().getBoolean("island.allowPvP", false);
-        Settings.allowNetherPvP = getConfig().getBoolean("island.allowNetherPvP", false);
-        Settings.allowBreakBlocks = getConfig().getBoolean("island.allowbreakblocks", false);
-        Settings.allowPlaceBlocks = getConfig().getBoolean("island.allowplaceblocks", false);
-        Settings.allowBedUse = getConfig().getBoolean("island.allowbeduse", false);
-        Settings.allowBucketUse = getConfig().getBoolean("island.allowbucketuse", false);
-        Settings.allowShearing = getConfig().getBoolean("island.allowshearing", false);
-        Settings.allowEnderPearls = getConfig().getBoolean("island.allowenderpearls", false);
-        Settings.allowDoorUse = getConfig().getBoolean("island.allowdooruse", false);
-        Settings.allowLeverButtonUse = getConfig().getBoolean("island.allowleverbuttonuse", false);
-        Settings.allowCropTrample = getConfig().getBoolean("island.allowcroptrample", false);
-        Settings.allowChestAccess = getConfig().getBoolean("island.allowchestaccess", false);
-        Settings.allowFurnaceUse = getConfig().getBoolean("island.allowfurnaceuse", false);
-        Settings.allowRedStone = getConfig().getBoolean("island.allowredstone", false);
-        Settings.allowMusic = getConfig().getBoolean("island.allowmusic", false);
-        Settings.allowCrafting = getConfig().getBoolean("island.allowcrafting", false);
-        Settings.allowBrewing = getConfig().getBoolean("island.allowbrewing", false);
-        Settings.allowGateUse = getConfig().getBoolean("island.allowgateuse", false);
-        Settings.allowHurtMobs = getConfig().getBoolean("island.allowhurtmobs", true);
-        Settings.endermanDeathDrop = getConfig().getBoolean("island.endermandeathdrop", true);
-        Settings.allowEndermanGriefing = getConfig().getBoolean("island.allowendermangriefing", true);
-        Settings.allowCreeperDamage = getConfig().getBoolean("island.allowcreeperdamage", true);
-        Settings.allowCreeperGriefing = getConfig().getBoolean("island.allowcreepergriefing", false);
-        Settings.allowTNTDamage = getConfig().getBoolean("island.allowtntdamage", false);
-        Settings.allowMonsterEggs = getConfig().getBoolean("island.allowspawneggs", false);
-        Settings.allowBreeding = getConfig().getBoolean("island.allowbreeding", false);
-        Settings.allowFire = getConfig().getBoolean("island.allowfire", false);
-        Settings.allowFireSpread = getConfig().getBoolean("island.allowfirespread", false);
-        Settings.allowFireExtinguish = getConfig().getBoolean("island.allowfireextinguish", false);
-        Settings.allowChestDamage = getConfig().getBoolean("island.allowchestdamage", false);
-        Settings.allowLeashUse = getConfig().getBoolean("island.allowleashuse", false);
-        Settings.allowHurtMonsters = getConfig().getBoolean("island.allowhurtmonsters", true);
-        Settings.allowEnchanting = getConfig().getBoolean("island.allowenchanting", true);
-        Settings.allowAnvilUse = getConfig().getBoolean("island.allowanviluse", true);
-        Settings.allowVisitorKeepInvOnDeath = getConfig().getBoolean("island.allowvisitorkeepinvondeath", false);
-        Settings.allowVisitorItemDrop = getConfig().getBoolean("island.allowvisitoritemdrop", true);
-        Settings.allowVisitorItemPickup = getConfig().getBoolean("island.allowvisitoritempickup", true);
-        Settings.allowArmorStandUse = getConfig().getBoolean("island.allowarmorstanduse", false);
-        Settings.allowBeaconAccess = getConfig().getBoolean("island.allowbeaconaccess", false);
-        Settings.allowPortalUse = getConfig().getBoolean("island.allowportaluse", true);
-        Settings.allowPressurePlate = getConfig().getBoolean("island.allowpressureplates", true);
-        Settings.allowPistonPush = getConfig().getBoolean("island.allowpistonpush", true);
-        Settings.allowHorseRiding = getConfig().getBoolean("island.allowhorseriding", false);
-        Settings.allowHorseInvAccess = getConfig().getBoolean("island.allowhorseinventoryaccess", false);
-        Settings.allowVillagerTrading = getConfig().getBoolean("island.allowvillagertrading", true);
-        Settings.allowChorusFruit = getConfig().getBoolean("island.allowchorusfruit", false);
-        Settings.enableJoinAndLeaveIslandMessages = getConfig().getBoolean("island.enablejoinandleaveislandmessages", true);
-        Settings.allowMobDamageToItemFrames = getConfig().getBoolean("island.allowitemframedamage", false);
-        Settings.allowMobSpawning = getConfig().getBoolean("island.allowmobspawning", true);
-
-        // Default settings hashmap
-        Settings.defaultIslandSettings.clear();
-        Settings.defaultIslandSettings.put(Flags.allowAnvilUse, Settings.allowAnvilUse);
-        Settings.defaultIslandSettings.put(Flags.allowArmorStandUse, Settings.allowArmorStandUse);
-        Settings.defaultIslandSettings.put(Flags.allowBeaconAccess, Settings.allowBeaconAccess);
-        Settings.defaultIslandSettings.put(Flags.allowBedUse, Settings.allowBedUse);
-        Settings.defaultIslandSettings.put(Flags.allowBreakBlocks, Settings.allowBreakBlocks);
-        Settings.defaultIslandSettings.put(Flags.allowBreeding, Settings.allowBreeding);
-        Settings.defaultIslandSettings.put(Flags.allowBrewing, Settings.allowBrewing);
-        Settings.defaultIslandSettings.put(Flags.allowBucketUse, Settings.allowBucketUse);
-        Settings.defaultIslandSettings.put(Flags.allowChestAccess, Settings.allowChestAccess);
-        Settings.defaultIslandSettings.put(Flags.allowCrafting, Settings.allowCrafting);
-        Settings.defaultIslandSettings.put(Flags.allowCropTrample, Settings.allowCropTrample);
-        Settings.defaultIslandSettings.put(Flags.allowDoorUse, Settings.allowDoorUse);
-        Settings.defaultIslandSettings.put(Flags.allowEnchanting, Settings.allowEnchanting);
-        Settings.defaultIslandSettings.put(Flags.allowEnderPearls, Settings.allowEnderPearls);
-        Settings.defaultIslandSettings.put(Flags.allowFurnaceUse, Settings.allowFurnaceUse);
-        Settings.defaultIslandSettings.put(Flags.allowGateUse, Settings.allowGateUse);
-        Settings.defaultIslandSettings.put(Flags.allowHorseInvAccess, Settings.allowHorseInvAccess);
-        Settings.defaultIslandSettings.put(Flags.allowHorseRiding, Settings.allowHorseRiding);
-        Settings.defaultIslandSettings.put(Flags.allowHurtMobs, Settings.allowHurtMobs);
-        Settings.defaultIslandSettings.put(Flags.allowLeashUse, Settings.allowLeashUse);
-        Settings.defaultIslandSettings.put(Flags.allowLeverButtonUse, Settings.allowLeverButtonUse);
-        Settings.defaultIslandSettings.put(Flags.allowMusic, Settings.allowMusic);
-        Settings.defaultIslandSettings.put(Flags.allowPlaceBlocks, Settings.allowPlaceBlocks);
-        Settings.defaultIslandSettings.put(Flags.allowPortalUse, Settings.allowPortalUse);
-        Settings.defaultIslandSettings.put(Flags.allowPressurePlate, Settings.allowPressurePlate);
-        Settings.defaultIslandSettings.put(Flags.allowPvP, Settings.allowPvP);
-        Settings.defaultIslandSettings.put(Flags.allowNetherPvP, Settings.allowNetherPvP);
-        Settings.defaultIslandSettings.put(Flags.allowRedStone, Settings.allowRedStone);
-        Settings.defaultIslandSettings.put(Flags.allowShearing, Settings.allowShearing);
-        Settings.defaultIslandSettings.put(Flags.allowVillagerTrading, Settings.allowVillagerTrading);
-        Settings.defaultIslandSettings.put(Flags.allowChorusFruit, Settings.allowChorusFruit);
-        Settings.defaultIslandSettings.put(Flags.enableJoinAndLeaveIslandMessages, Settings.enableJoinAndLeaveIslandMessages);
-        Settings.defaultIslandSettings.put(Flags.allowMobSpawning, Settings.allowMobSpawning);
-
-        // Spawn Settings
-        Settings.allowSpawnCreeperPain = getConfig().getBoolean("spawn.allowcreeperpain", false);
-        Settings.allowSpawnHorseRiding = getConfig().getBoolean("spawn.allowhorseriding", false);
-        Settings.allowSpawnHorseInvAccess = getConfig().getBoolean("spawn.allowhorseinventoryaccess", false);
-        Settings.allowSpawnPressurePlate = getConfig().getBoolean("spawn.allowpressureplates", true);
-        Settings.allowSpawnDoorUse = getConfig().getBoolean("spawn.allowdooruse", true);
-        Settings.allowSpawnLeverButtonUse = getConfig().getBoolean("spawn.allowleverbuttonuse", true);
-        Settings.allowSpawnChestAccess = getConfig().getBoolean("spawn.allowchestaccess", true);
-        Settings.allowSpawnFurnaceUse = getConfig().getBoolean("spawn.allowfurnaceuse", true);
-        Settings.allowSpawnRedStone = getConfig().getBoolean("spawn.allowredstone", false);
-        Settings.allowSpawnMusic = getConfig().getBoolean("spawn.allowmusic", true);
-        Settings.allowSpawnCrafting = getConfig().getBoolean("spawn.allowcrafting", true);
-        Settings.allowSpawnBrewing = getConfig().getBoolean("spawn.allowbrewing", true);
-        Settings.allowSpawnGateUse = getConfig().getBoolean("spawn.allowgateuse", true);
-        Settings.allowSpawnMobSpawn = getConfig().getBoolean("spawn.allowmobspawn", false);
-        Settings.allowSpawnAnimalSpawn = getConfig().getBoolean("spawn.allowanimalspawn", true);
-        Settings.allowSpawnAnimalKilling = getConfig().getBoolean("spawn.allowanimalkilling", false);
-        Settings.allowSpawnMobKilling = getConfig().getBoolean("spawn.allowmobkilling", true);
-        Settings.allowSpawnMonsterEggs = getConfig().getBoolean("spawn.allowspawneggs", false);
-        Settings.allowSpawnEggs = getConfig().getBoolean("spawn.alloweggs", false);
-        Settings.allowSpawnBreakBlocks = getConfig().getBoolean("spawn.allowbreakblocks", false);
-        Settings.allowSpawnPlaceBlocks = getConfig().getBoolean("spawn.allowplaceblocks", false);
-        Settings.allowSpawnNoAcidWater = getConfig().getBoolean("spawn.allowspawnnoacidwater", false);
-        Settings.allowSpawnEnchanting = getConfig().getBoolean("spawn.allowenchanting", true);
-        Settings.allowSpawnAnvilUse = getConfig().getBoolean("spawn.allowanviluse", true);
-        Settings.allowSpawnBeaconAccess = getConfig().getBoolean("spawn.allowbeaconaccess", false);
-        Settings.allowSpawnPVP = getConfig().getBoolean("spawn.allowPVP", false);
-        Settings.allowSpawnMilking = getConfig().getBoolean("spawn.allowmilking", false);
-        Settings.allowSpawnLavaCollection = getConfig().getBoolean("spawn.allowlavacollection", false);
-        Settings.allowSpawnWaterCollection = getConfig().getBoolean("spawn.allowwatercollection", false);
-        Settings.allowSpawnVisitorItemDrop = getConfig().getBoolean("spawn.allowvisitoritemdrop", true);
-        Settings.allowSpawnVisitorItemPickup = getConfig().getBoolean("spawn.allowvisitoritempickup", true);
-        Settings.allowSpawnArmorStandUse = getConfig().getBoolean("spawn.allowarmorstanduse",false);
-        Settings.allowSpawnBedUse = getConfig().getBoolean("spawn.allowbeduse",false);
-        Settings.allowSpawnBreeding = getConfig().getBoolean("spawn.allowbreeding",false);
-        Settings.allowSpawnCropTrample = getConfig().getBoolean("spawn.allowcroptrample",false);
-        Settings.allowSpawnEnderPearls = getConfig().getBoolean("spawn.allowenderpearls",false);
-        Settings.allowSpawnLeashUse = getConfig().getBoolean("spawn.allowleashuse",false);
-        Settings.allowSpawnVillagerTrading = getConfig().getBoolean("spawn.allowvillagertrading", false);
-        Settings.allowSpawnChorusFruit = getConfig().getBoolean("spawn.allowchorusfruit", false);
-        Settings.allowSpawnFireExtinguish = getConfig().getBoolean("spawn.allowfireextinguish", false);
-
-        // Challenges
-        getChallenges();
-        // Challenge completion
-        Settings.broadcastMessages = getConfig().getBoolean("general.broadcastmessages", true);
-
-        // Levels
-        // Get the blockvalues.yml file
-        YamlConfiguration blockValuesConfig = Util.loadYamlFile("blockvalues.yml");
-        // Get the under water multiplier
-        Settings.deathpenalty = blockValuesConfig.getInt("deathpenalty", 0);
-        Settings.sumTeamDeaths = blockValuesConfig.getBoolean("sumteamdeaths");
-        Settings.maxDeaths = blockValuesConfig.getInt("maxdeaths", 10);
-        Settings.islandResetDeathReset = blockValuesConfig.getBoolean("islandresetdeathreset", true);
-        Settings.teamJoinDeathReset = blockValuesConfig.getBoolean("teamjoindeathreset", true);
-        Settings.underWaterMultiplier = blockValuesConfig.getDouble("underwater", 1D);
-        Settings.levelCost = blockValuesConfig.getInt("levelcost", 100);
-        if (Settings.levelCost < 1) {
-            Settings.levelCost = 1;
-            getLogger().warning("levelcost in blockvalues.yml cannot be less than 1. Setting to 1.");
-        }
-        Settings.blockLimits = new HashMap<MaterialData, Integer>();
-        if (blockValuesConfig.isSet("limits")) {
-            for (String material : blockValuesConfig.getConfigurationSection("limits").getKeys(false)) {
-                try {
-                    String[] split = material.split(":");
-                    byte data = 0;
-                    if (split.length>1) {
-                        data = Byte.valueOf(split[1]);
-                    }
-                    Material mat;
-                    if (StringUtils.isNumeric(split[0])) {
-                        mat = Material.getMaterial(Integer.parseInt(split[0]));
-                    } else {
-                        mat = Material.valueOf(split[0].toUpperCase());
-                    }
-                    MaterialData materialData = new MaterialData(mat);
-                    materialData.setData(data);
-                    Settings.blockLimits.put(materialData, blockValuesConfig.getInt("limits." + material, 0));
-                    if (debug) {
-                        getLogger().info("Maximum number of " + materialData + " will be " + Settings.blockLimits.get(materialData));
-                    }
-                } catch (Exception e) {
-                    getLogger().warning("Unknown material (" + material + ") in blockvalues.yml Limits section. Skipping...");
-                }
-            }
-        }
-        Settings.blockValues = new HashMap<MaterialData, Integer>();
-        if (blockValuesConfig.isSet("blocks")) {
-            for (String material : blockValuesConfig.getConfigurationSection("blocks").getKeys(false)) {
-                try {
-                    String[] split = material.split(":");
-                    byte data = 0;
-                    if (split.length>1) {
-                        data = Byte.valueOf(split[1]);
-                    }
-                    Material mat;
-                    if (StringUtils.isNumeric(split[0])) {
-                        mat = Material.getMaterial(Integer.parseInt(split[0]));
-                    } else {
-                        mat = Material.valueOf(split[0].toUpperCase());
-                    }
-                    MaterialData materialData = new MaterialData(mat);
-                    materialData.setData(data);
-                    Settings.blockValues.put(materialData, blockValuesConfig.getInt("blocks." + material, 0));
-                    if (debug) {
-                        getLogger().info(mat.toString() + " value is " + Settings.blockValues.get(mat));
-                    }
-                } catch (Exception e) {
-                    // e.printStackTrace();
-                    getLogger().warning("Unknown material (" + material + ") in blockvalues.yml blocks section. Skipping...");
-                }
-            }
-        } else {
-            getLogger().severe("No block values in blockvalues.yml! All island levels will be zero!");
-        }
-        // Biome Settings
-        Settings.biomeCost = getConfig().getDouble("biomesettings.defaultcost", 100D);
-        if (Settings.biomeCost < 0D) {
-            Settings.biomeCost = 0D;
-            getLogger().warning("Biome default cost is < $0, so set to zero.");
-        }
-        String defaultBiome = getConfig().getString("biomesettings.defaultbiome", "PLAINS");
-        try {
-            Settings.defaultBiome = Biome.valueOf(defaultBiome);
-        } catch (Exception e) {
-            getLogger().severe("Could not parse biome " + defaultBiome + " using PLAINS instead.");
-            Settings.defaultBiome = Biome.PLAINS;
-        }
-        Settings.breedingLimit = getConfig().getInt("general.breedinglimit", 0);
-        Settings.villagerLimit = getConfig().getInt("general.villagerlimit", 0);
-        Settings.limitedBlocks = new HashMap<String,Integer>();
-        ConfigurationSection entityLimits = getConfig().getConfigurationSection("general.entitylimits");
-        if (entityLimits != null) {
-            for (String entity: entityLimits.getKeys(false)) {
-                int limit = entityLimits.getInt(entity.toUpperCase(), -1);
-                if (limit > 0) {
-                    getLogger().info(entity.toUpperCase() + " will be limited to " + limit);
-                }
-                if (Material.getMaterial(entity.toUpperCase()) == null) {
-                    getLogger().warning("general.entitylimits section has unknown entity type: " + entity.toUpperCase() + " skipping...");
-                } else if (limit > -1) {
-                    Settings.limitedBlocks.put(entity.toUpperCase(), limit);
-                    if (entity.equalsIgnoreCase("REDSTONE_COMPARATOR")) {
-                        // Player can only ever place a redstone comparator in the OFF state
-                        Settings.limitedBlocks.put("REDSTONE_COMPARATOR_OFF", limit);
-                    } else if (entity.equalsIgnoreCase("BANNER")) {
-                        // To simplify banners, the banner is allowed and automatically made wall and standing banner
-                        Settings.limitedBlocks.put("WALL_BANNER", limit);
-                        Settings.limitedBlocks.put("STANDING_BANNER", limit);
-                    } else if (entity.equalsIgnoreCase("SIGN")) {
-                        // To simplify signs, the sign is allowed and automatically made wall and standing signs
-                        Settings.limitedBlocks.put("WALL_SIGN", limit);
-                        Settings.limitedBlocks.put("SIGN_POST", limit);
-                    }
-                }
-            }
-        }
-        // Legacy setting support for hopper limiting
-        if (Settings.limitedBlocks.isEmpty()) {
-            Settings.hopperLimit = getConfig().getInt("general.hopperlimit", -1);
-            if (Settings.hopperLimit > 0) {
-                Settings.limitedBlocks.put("HOPPER", Settings.hopperLimit);
-            }
-        }
-        Settings.mobLimit = getConfig().getInt("general.moblimit", 0);
-        Settings.removeCompleteOntimeChallenges = getConfig().getBoolean("general.removecompleteonetimechallenges", false);
-        Settings.addCompletedGlow = getConfig().getBoolean("general.addcompletedglow", true);
-        // Clean up blocks around edges when deleting islands
-        Settings.cleanUpBlocks = getConfig().getBoolean("island.cleanupblocks",false);
-        Settings.cleanRate = getConfig().getInt("island.cleanrate", 2);
-        if (Settings.cleanRate < 1) {
-            Settings.cleanRate = 1;
-        }
-        // No acid bottles or buckets
-        Settings.acidBottle = getConfig().getBoolean("general.acidbottles", true);
-        // Island name length
-        Settings.minNameLength = getConfig().getInt("island.minnamelength", 1);
-        Settings.maxNameLength = getConfig().getInt("island.maxnamelength", 20);
-        if (Settings.minNameLength < 0) {
-            Settings.minNameLength = 0;
-        }
-        if (Settings.maxNameLength < 1) {
-            Settings.maxNameLength = 1;
-        }
-        if (Settings.minNameLength > Settings.maxNameLength) {
-            Settings.minNameLength = Settings.maxNameLength;
-        }
-        // Magic Cobble Generator
-        Settings.useMagicCobbleGen = getConfig().getBoolean("general.usemagiccobblegen", false);
-        if(Settings.useMagicCobbleGen && getConfig().isSet("general.magiccobblegenchances")){
-            //getLogger().info("DEBUG: magic cobble gen enabled and chances section found");
-            Settings.magicCobbleGenChances = new TreeMap<Integer, TreeMap<Double,Material>>();
-            for(String level : getConfig().getConfigurationSection("general.magiccobblegenchances").getKeys(false)){
-                int levelInt = 0;
-                try{
-                    if(level.equals("default")) {
-                        levelInt = Integer.MIN_VALUE;
-                    } else {
-                        levelInt = Integer.parseInt(level);
-                    } 
-                    TreeMap<Double,Material> blockMapTree = new TreeMap<Double, Material>();
-                    double chanceTotal = 0;
-                    for(String block : getConfig().getConfigurationSection("general.magiccobblegenchances." + level).getKeys(false)){
-                        double chance = getConfig().getDouble("general.magiccobblegenchances." + level + "." + block, 0D);
-                        if(chance > 0 && Material.getMaterial(block) != null && Material.getMaterial(block).isBlock()) {
-                            // Store the cumulative chance in the treemap. It does not need to add up to 100%
-                            chanceTotal += chance;
-                            blockMapTree.put(chanceTotal, Material.getMaterial(block));
-                        }
-                    }
-                    if (!blockMapTree.isEmpty()) {
-                        Settings.magicCobbleGenChances.put(levelInt, blockMapTree);
-                    }
-                } catch(NumberFormatException e){
-                    // Putting the catch here means that an invalid level is skipped completely
-                    getLogger().severe("Unknown level '" + level + "' listed in magiccobblegenchances section! Must be an integer or 'default'. Skipping...");
-                }
-            }
-        }
-        // Disable offline redstone
-        Settings.disableOfflineRedstone = getConfig().getBoolean("general.disableofflineredstone", false);
-        // All done
-        return true;
     }
 
     /**
@@ -1552,8 +703,20 @@ public class ASkyBlock extends JavaPlugin {
         manager.registerEvents(new NetherSpawning(this), this);
         // Island Protection events
         manager.registerEvents(new IslandGuard(this), this);
+        // Island Entity Limits
+        entityLimits = new EntityLimits(this);
+        manager.registerEvents(entityLimits, this);
         // Player events
-        manager.registerEvents(new PlayerEvents(this), this);
+        playerEvents = new PlayerEvents(this);
+        manager.registerEvents(playerEvents, this);
+        try {
+            Class<?> clazz = Class.forName("org.bukkit.event.entity.EntityPickupItemEvent", false, getClassLoader());
+            if (clazz != null) {
+                manager.registerEvents(new PlayerEvents3(this), this);
+            }
+        } catch (ClassNotFoundException e) {
+            manager.registerEvents(new PlayerEvents2(this), this);
+        }
         // New V1.8 events
         if (onePointEight) {
             manager.registerEvents(new IslandGuard1_8(this), this);
@@ -1598,8 +761,6 @@ public class ASkyBlock extends JavaPlugin {
         if (Settings.recoverSuperFlat) {
             manager.registerEvents(new CleanSuperFlat(), this);
         }
-        // World loader
-        //manager.registerEvents(new WorldLoader(this), this);
     }
 
 
@@ -1608,8 +769,9 @@ public class ASkyBlock extends JavaPlugin {
      * Resets a player's inventory, armor slots, equipment, enderchest and
      * potion effects
      *
-     * @param player
+     * @param player - player
      */
+    @SuppressWarnings("deprecation")
     public void resetPlayer(Player player) {
         // getLogger().info("DEBUG: clear inventory = " +
         // Settings.clearInventory);
@@ -1639,7 +801,7 @@ public class ASkyBlock extends JavaPlugin {
         players.clearStartIslandRating(player.getUniqueId());
         // Save the player
         players.save(player.getUniqueId());
-        TopTen.topTenAddEntry(player.getUniqueId(), 0);
+        topTen.topTenAddEntry(player.getUniqueId(), 0);
         // Update the inventory
         player.updateInventory();
         if (Settings.resetEnderChest) {
@@ -1659,14 +821,6 @@ public class ASkyBlock extends JavaPlugin {
         // Enables warp signs in ASkyBlock
         warpSignsListener = new WarpSigns(this);
         manager.registerEvents(warpSignsListener, this);
-    }
-
-    /**
-     * @param calculatingLevel
-     *            the calculatingLevel to set
-     */
-    public void setCalculatingLevel(boolean calculatingLevel) {
-        this.calculatingLevel = calculatingLevel;
     }
 
     /**
@@ -1706,12 +860,13 @@ public class ASkyBlock extends JavaPlugin {
     }
 
     /**
+     * @param player - player
      * @return Locale for this player
      */
     public ASLocale myLocale(UUID player) {
         String locale = players.getLocale(player);
         if (locale.isEmpty() || !availableLocales.containsKey(locale)) {
-            return availableLocales.get("locale");
+            return availableLocales.get(Settings.defaultLanguage);
         }
         return availableLocales.get(locale);
     }
@@ -1720,7 +875,7 @@ public class ASkyBlock extends JavaPlugin {
      * @return System locale
      */
     public ASLocale myLocale() {
-        return availableLocales.get("locale");
+        return availableLocales.get(Settings.defaultLanguage);
     }
 
     /**
@@ -1791,7 +946,67 @@ public class ASkyBlock extends JavaPlugin {
     /**
      * @return the availableLocales
      */
-    public HashMap<String, ASLocale> getAvailableLocales() {
+    public Map<String, ASLocale> getAvailableLocales() {
         return availableLocales;
+    }
+
+    /**
+     * @param availableLocales the availableLocales to set
+     */
+    public void setAvailableLocales(HashMap<String, ASLocale> availableLocales) {
+        this.availableLocales = availableLocales;
+    }
+
+    /**
+     * @return the acidTask
+     */
+    public AcidTask getAcidTask() {
+        return acidTask;
+    }
+
+    /**
+     * @return the playerEvents
+     */
+    public PlayerEvents getPlayerEvents() {
+        return playerEvents;
+    }
+
+    /**
+     * Registers the custom charts for Metrics
+     */
+    public void registerCustomCharts(){
+        metrics.addCustomChart(new Metrics.SimplePie("challenges_count", () -> {
+
+            int count = challenges.getAllChallenges().size();
+            if(count <= 0) return "0";
+            else if(count <= 10) return "1-10";
+            else if(count <= 20) return "11-20";
+            else if(count <= 30) return "21-30";
+            else if(count <= 40) return "31-40";
+            else if(count <= 50) return "41-50";
+            else if(count <= 75) return "51-75";
+            else if(count <= 100) return "76-100";
+            else if(count <= 150) return "101-150";
+            else if(count <= 200) return "151-200";
+            else if(count <= 300) return "201-300";
+            else return "300+";
+        }));
+
+        metrics.addCustomChart(new Metrics.SingleLineChart("islands_count",
+            () -> getGrid().getIslandCount()));
+    }
+
+    /**
+     * @return the headGetter
+     */
+    public HeadGetter getHeadGetter() {
+        return headGetter;
+    }
+
+    /**
+     * @return the topTen
+     */
+    public TopTen getTopTen() {
+        return topTen;
     }
 }
