@@ -24,6 +24,7 @@ import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Set;
 import java.util.UUID;
 
@@ -36,28 +37,34 @@ import com.wasteofplastic.askyblock.commands.Challenges;
 import com.wasteofplastic.askyblock.panels.SetBiome;
 
 /**
- * Provides a programming interface
- * 
+ * Provides a programming interface. The aim of this class is to provide a fairly static set of calls for other plugins to use.
+ * If you try to hook into other classes in the code, they may change and break your code.
+ *
  * @author tastybento
  */
 public class ASkyBlockAPI {
-    private static ASkyBlockAPI instance = new ASkyBlockAPI(ASkyBlock.getPlugin());
+    private static ASkyBlockAPI instance;
 
     /**
      * @return the instance
      */
     public static ASkyBlockAPI getInstance() {
+        if (instance == null)
+            // Initialize the API
+            new ASkyBlockAPI(ASkyBlock.getPlugin());
         return instance;
     }
 
+
     private ASkyBlock plugin;
 
-    private ASkyBlockAPI(ASkyBlock plugin) {
+    protected ASkyBlockAPI(ASkyBlock plugin) {
         this.plugin = plugin;
+        instance = this;
     }
 
     /**
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return Map of all of the known challenges with a boolean marking
      *         them as complete (true) or incomplete (false). This is a view of the
      *         challenges map that only allows read operations.
@@ -65,9 +72,9 @@ public class ASkyBlockAPI {
     public Map<String, Boolean> getChallengeStatus(UUID playerUUID) {
         return Collections.unmodifiableMap(plugin.getPlayers().getChallengeStatus(playerUUID));
     }
-    
+
     /**
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return Map of all of the known challenges and how many times each
      *         one has been completed. This is a view of the challenges
      *         map that only allows read operations.
@@ -81,13 +88,29 @@ public class ASkyBlockAPI {
     }
 
     /**
+     * @deprecated
+     * Island level is now a long and the int value may not be accurate for very large values.
+     * Use getLongIslandLevel(UUID playerUUID) instead.
+     *
      * Returns the island level from the last time it was calculated. Note this
      * does not calculate the island level.
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      * @return the last level calculated for the island or zero if none.
      */
+    @Deprecated
     public int getIslandLevel(UUID playerUUID) {
+        return (int)plugin.getPlayers().getIslandLevel(playerUUID);
+    }
+
+    /**
+     * Returns the island level from the last time it was calculated. Note this
+     * does not calculate the island level.
+     *
+     * @param playerUUID - the player's UUID - player's UUID
+     * @return the last level calculated for the island or zero if none.
+     */
+    public long getLongIslandLevel(UUID playerUUID) {
         return plugin.getPlayers().getIslandLevel(playerUUID);
     }
 
@@ -96,25 +119,25 @@ public class ASkyBlockAPI {
      * You will need to check if the player is in a team and individually set the level of each team member.
      * This value will be overwritten if the players run the build-in level command or if the island level
      * is calculated some other way, e.g. at login or via an admin command.
-     * @param playerUUID
-     * @param level
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param level - level to set
      */
     public void setIslandLevel(UUID playerUUID, int level) {
         plugin.getPlayers().setIslandLevel(playerUUID, level);
     }
-    
+
     /**
      * Calculates the island level. Only the fast calc is supported.
      * The island calculation runs async and fires an IslandLevelEvent when completed
      * or use getIslandLevel(playerUUID). See https://gist.github.com/tastybento/e81d2403c03f2fe26642
      * for example code.
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      * @return true if player has an island, false if not
      */
     public boolean calculateIslandLevel(UUID playerUUID) {
-        if (plugin.getPlayers().inTeam(playerUUID) && !plugin.getPlayers().hasIsland(playerUUID)) {		
-            new LevelCalcByChunk(plugin, playerUUID, null, false);
+        if (plugin.getPlayers().hasIsland(playerUUID) || plugin.getPlayers().inTeam(playerUUID)) {
+            new LevelCalcByChunk(plugin, plugin.getGrid().getIsland(playerUUID), playerUUID, null, false);
             return true;
         }
         return false;
@@ -123,8 +146,8 @@ public class ASkyBlockAPI {
     /**
      * Provides the location of the player's island, either the team island or
      * their own
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - players UUID
      * @return Location of island
      */
     public Location getIslandLocation(UUID playerUUID) {
@@ -134,8 +157,8 @@ public class ASkyBlockAPI {
     /**
      * Returns the owner of an island from the location.
      * Uses the grid lookup and is quick
-     * 
-     * @param location
+     *
+     * @param location - location to check
      * @return UUID of owner
      */
     public UUID getOwner(Location location) {
@@ -144,8 +167,8 @@ public class ASkyBlockAPI {
 
     /**
      * Get Team Leader
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      * @return UUID of Team Leader or null if there is none. Use inTeam to
      *         check.
      */
@@ -156,8 +179,8 @@ public class ASkyBlockAPI {
     /**
      * Get a list of team members. This is a copy and changing the return value
      * will not affect the membership.
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      * @return List of team members, including the player. Empty if there are
      *         none.
      */
@@ -167,8 +190,8 @@ public class ASkyBlockAPI {
 
     /**
      * Provides location of the player's warp sign
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      * @return Location of sign or null if one does not exist
      */
     public Location getWarp(UUID playerUUID) {
@@ -177,8 +200,8 @@ public class ASkyBlockAPI {
 
     /**
      * Get the owner of the warp at location
-     * 
-     * @param location
+     *
+     * @param location - location of the warp
      * @return Returns name of player or empty string if there is no warp at
      *         that spot
      */
@@ -189,8 +212,8 @@ public class ASkyBlockAPI {
     /**
      * Status of island ownership. Team members do not have islands of their
      * own, only leaders do.
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      * @return true if player has an island, false if the player does not.
      */
     public boolean hasIsland(UUID playerUUID) {
@@ -198,7 +221,7 @@ public class ASkyBlockAPI {
     }
 
     /**
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return true if in a team
      */
     public boolean inTeam(UUID playerUUID) {
@@ -209,8 +232,8 @@ public class ASkyBlockAPI {
      * Determines if an island is at a location in this area location. Also
      * checks if the spawn island is in this area. Checks for bedrock within
      * limits and also looks in the file system. Quite processor intensive.
-     * 
-     * @param location
+     *
+     * @param location - location to check
      * @return true if there is an island in that location, false if not
      */
     public boolean islandAtLocation(Location location) {
@@ -220,10 +243,9 @@ public class ASkyBlockAPI {
     /**
      * Checks to see if a player is trespassing on another player's island. Both
      * players must be online.
-     * 
-     * @param owner
-     *            - owner or team member of an island
-     * @param target
+     *
+     * @param owner - player owner or team member of an island
+     * @param target - player who may be doing the trespassing
      * @return true if they are on the island otherwise false.
      */
     public boolean isOnIsland(Player owner, Player target) {
@@ -235,7 +257,7 @@ public class ASkyBlockAPI {
      * player's UUID is used. It can be displayed however you like to other
      * users. This is a copy of the set and changing it will not affect the
      * actual set of warps.
-     * 
+     *
      * @return String set of warps
      */
     public Set<UUID> listWarps() {
@@ -253,9 +275,9 @@ public class ASkyBlockAPI {
     /**
      * Checks if a specific location is within the protected range of an island
      * owned by the player
-     * 
-     * @param player
-     * @param location
+     *
+     * @param player - player object
+     * @param location - location to check
      * @return true if the location is on an island owner by player
      */
     public boolean locationIsOnIsland(final Player player, final Location location) {
@@ -266,10 +288,10 @@ public class ASkyBlockAPI {
      * Finds out if location is within a set of island locations and returns the
      * one that is there or null if not. The islandTestLocations should be the center
      * location of an island. The check is done to see if loc is inside the protected
-     * range of any of the islands given. 
-     * 
-     * @param islandTestLocations
-     * @param loc
+     * range of any of the islands given.
+     *
+     * @param islandTestLocations - Set of island locations to check
+     * @param loc - location to check
      * @return the island location that is in the set of locations or null if
      *         none
      */
@@ -280,9 +302,8 @@ public class ASkyBlockAPI {
     /**
      * Checks if an online player is on their island, on a team island or on a
      * coop island
-     * 
-     * @param player
-     *            - the player who is being checked
+     *
+     * @param player - the player who is being checked
      * @return - true if they are on their island, otherwise false
      */
     public boolean playerIsOnIsland(Player player) {
@@ -291,15 +312,15 @@ public class ASkyBlockAPI {
 
     /**
      * Sets all blocks in an island to a specified biome type
-     * 
-     * @param islandLoc
-     * @param biomeType
+     *
+     * @param islandLoc - island location
+     * @param biomeType - biome type
      * @return true if the setting was successful
      */
     public boolean setIslandBiome(Location islandLoc, Biome biomeType) {
         Island island = plugin.getGrid().getIslandAt(islandLoc);
         if (island != null) {
-            new SetBiome(plugin, island, biomeType);
+            new SetBiome(plugin, island, biomeType, null);
             return true;
         }
         return false;
@@ -307,9 +328,9 @@ public class ASkyBlockAPI {
 
     /**
      * Sets a message for the player to receive next time they login
-     * 
-     * @param playerUUID
-     * @param message
+     *
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param message - message to set
      * @return true if player is offline, false if online
      */
     public boolean setMessage(UUID playerUUID, String message) {
@@ -319,9 +340,9 @@ public class ASkyBlockAPI {
     /**
      * Sends a message to every player in the team that is offline If the player
      * is not in a team, nothing happens.
-     * 
-     * @param playerUUID
-     * @param message
+     *
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param message - message to send
      */
     public void tellOfflineTeam(UUID playerUUID, String message) {
         plugin.getMessages().tellOfflineTeam(playerUUID, message);
@@ -329,21 +350,18 @@ public class ASkyBlockAPI {
 
     /**
      * Player is in a coop or not
-     * 
-     * @param player
+     *
+     * @param player - player object to check
      * @return true if player is in a coop, otherwise false
      */
     public boolean isCoop(Player player) {
-        if (CoopPlay.getInstance().getCoopIslands(player).isEmpty()) {
-            return false;
-        }
-        return true;
+        return !CoopPlay.getInstance().getCoopIslands(player).isEmpty();
     }
 
     /**
      * Find out which coop islands player is a part of
-     * 
-     * @param player
+     *
+     * @param player - player object to check
      * @return set of locations of islands or empty if none
      */
     public Set<Location> getCoopIslands(Player player) {
@@ -368,7 +386,7 @@ public class ASkyBlockAPI {
 
     /**
      * Checks if a location is at spawn or not
-     * @param location
+     * @param location - location to check
      * @return true if at spawn
      */
     public boolean isAtSpawn(Location location) {
@@ -400,29 +418,53 @@ public class ASkyBlockAPI {
     }
 
     /**
+     * @deprecated
+     * Island levels are now stored as longs, so the int value may not be accurate
+     *
      * Get the top ten list
      * @return Top ten list
      */
+    @Deprecated
     public Map<UUID, Integer> getTopTen() {
-        return new HashMap<UUID, Integer>(TopTen.getTopTenList());
+        HashMap<UUID, Integer> result = new HashMap<UUID, Integer>();
+        for (Entry<UUID, Long> en : plugin.getTopTen().getTopTenList().entrySet()) {
+            result.put(en.getKey(), en.getValue().intValue());
+        }
+        return result;
+    }
+
+    /**
+     * Get the top ten list
+     * @return Top ten list
+     */
+    public Map<UUID, Long> getLongTopTen() {
+        return new HashMap<UUID, Long>(plugin.getTopTen().getTopTenList());
     }
 
     /**
      * Obtains a copy of the island object owned by playerUUID
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return copy of Island object
      */
     public Island getIslandOwnedBy(UUID playerUUID) {
-        return (Island)plugin.getGrid().getIsland(playerUUID).clone();
+        Island island = plugin.getGrid().getIsland(playerUUID);
+        if (island != null) {
+            return new Island(island);
+        }
+        return null;
     }
 
     /**
-     * Returns a copy of the Island object for an island at this location or null if one does not exist
-     * @param location
+     * Returns the Island object for an island at this location or null if one does not exist
+     * @param location - location requested
      * @return copy of Island object
      */
     public Island getIslandAt(Location location) {
-        return (Island)plugin.getGrid().getIslandAt(location);
+        Island island = plugin.getGrid().getIslandAt(location);
+        if (island != null) {
+            return new Island(island);
+        }
+        return null;
     }
 
     /**
@@ -452,7 +494,7 @@ public class ASkyBlockAPI {
 
     /**
      * Get name of the island owned by owner
-     * @param owner
+     * @param owner - player's UUID
      * @return Returns the name of owner's island, or the owner's name if there is none.
      */
     public String getIslandName(UUID owner) {
@@ -461,18 +503,47 @@ public class ASkyBlockAPI {
 
     /**
      * Set the island name
-     * @param owner
-     * @param name
+     * @param owner - player's UUID
+     * @param name - name to set
      */
     public void setIslandName(UUID owner, String name) {
         plugin.getGrid().setIslandName(owner, name);
     }
-    
+
     /**
      * Get all the challenges
      * @return challenges per level
      */
     public LinkedHashMap<String, List<String>> getAllChallenges(){
-    	return Challenges.getChallengeList();
+        return Challenges.getChallengeList();
+    }
+
+    /**
+     * Get the number of resets left for this player
+     * @param playerUUID - the player's UUID - player's UUID
+     * @return Number of resets left
+     */
+    public int getResetsLeft(UUID playerUUID) {
+        return plugin.getPlayers().getResetsLeft(playerUUID);
+    }
+
+    /**
+     * Set the number of resets left for this player
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param resets - number to set
+     */
+    public void setResetsLeft(UUID playerUUID, int resets) {
+        plugin.getPlayers().setResetsLeft(playerUUID, resets);
+    }
+
+    /**
+     * Find out if this player is a team leader or not. If the player is not in a team, the result will always be false.
+     * @param playerUUID - the player's UUID - player's UUID
+     * @return true if the player is in a team and is the leader
+     */
+    public boolean isLeader(UUID playerUUID) {
+        UUID leader = plugin.getPlayers().getTeamLeader(playerUUID);
+        return leader != null && leader.equals(playerUUID);
+
     }
 }
