@@ -23,6 +23,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.UUID;
 
+import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.World;
@@ -30,7 +31,6 @@ import org.bukkit.block.Biome;
 import org.bukkit.block.BlockState;
 import org.bukkit.block.Hopper;
 import org.bukkit.entity.Entity;
-import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Villager;
 
 import com.google.common.collect.HashMultiset;
@@ -40,11 +40,11 @@ import com.wasteofplastic.askyblock.util.Util;
 /**
  * Stores all the info about an island
  * Managed by GridManager
- * 
+ *
  * @author tastybento
- * 
+ *
  */
-public class Island implements Cloneable {
+public class Island {
     ASkyBlock plugin;
     // Coordinates of the island area
     private int minX;
@@ -73,66 +73,226 @@ public class Island implements Cloneable {
     private boolean locked = false;
     // Set if this island is a spawn island
     private boolean isSpawn = false;
-    // Stats variables
-    private HashMap<EntityType, Integer> entities = new HashMap<EntityType, Integer>();
     // Protection against deletion or not
     private boolean purgeProtected;
-    // The spawn point 
+    // The spawn point
     private Location spawnPoint;
     // Tile entities
     private Multiset<Material> tileEntityCount = HashMultiset.create();
     // Biome
-    Biome biome;
+    private Biome biome;
+
     // Island protection settings
-    private HashMap<Flags, Boolean> igs = new HashMap<Flags, Boolean>();
+    private static final List<String> islandSettingsKey = new ArrayList<>();
+    static {
+        islandSettingsKey.clear();
+        islandSettingsKey.add("");
+    }
+    private HashMap<SettingsFlag, Boolean> igs = new HashMap<>();
     private int levelHandicap;
     /**
      * Island Guard Setting flags
-     *
+     * Covers island, spawn and system settings
      */
-    public enum Flags {
-        allowAnvilUse,
-        allowArmorStandUse,
-        allowBeaconAccess,
-        allowBedUse,
-        allowBreakBlocks,
-        allowBreeding,
-        allowBrewing,
-        allowBucketUse,
-        allowChestAccess,
-        allowCrafting,
-        allowCropTrample,
-        allowDoorUse,
-        allowEnchanting,
-        allowEnderPearls,
-        allowFurnaceUse,
-        allowGateUse,
-        allowHorseInvAccess,
-        allowHorseRiding,
-        allowHurtMobs,
-        allowLeashUse,
-        allowLeverButtonUse,
-        allowMusic,
-        allowPlaceBlocks,
-        allowPortalUse,
-        allowPressurePlate,
-        allowPvP,
-        allowNetherPvP,
-        allowRedStone,
-        allowShearing,
-        allowVillagerTrading,
-        allowChorusFruit,
-        enableJoinAndLeaveIslandMessages,
-        allowMobSpawning
+    public enum SettingsFlag {
+        /**
+         * Water is acid above sea level
+         */
+        ACID_DAMAGE,
+        /**
+         * Anvil use
+         */
+        ANVIL,
+        /**
+         * Armor stand use
+         */
+        ARMOR_STAND,
+        /**
+         * Beacon use
+         */
+        BEACON,
+        /**
+         * Bed use
+         */
+        BED,
+        /**
+         * Can break blocks
+         */
+        BREAK_BLOCKS,
+        /**
+         * Can breed animals
+         */
+        BREEDING,
+        /**
+         * Can use brewing stand
+         */
+        BREWING,
+        /**
+         * Can empty or fill buckets
+         */
+        BUCKET,
+        /**
+         * Can collect lava
+         */
+        COLLECT_LAVA,
+        /**
+         * Can collect water
+         */
+        COLLECT_WATER,
+        /**
+         * Can open chests or hoppers or dispensers
+         */
+        CHEST,
+        /**
+         * Can eat and teleport with chorus fruit
+         */
+        CHORUS_FRUIT,
+        /**
+         * Can use the work bench
+         */
+        CRAFTING,
+        /**
+         * Allow creepers to hurt players (but not damage blocks)
+         */
+        CREEPER_PAIN,
+        /**
+         * Can trample crops
+         */
+        CROP_TRAMPLE,
+        /**
+         * Can open doors or trapdoors
+         */
+        DOOR,
+        /**
+         * Chicken eggs can be thrown
+         */
+        EGGS,
+        /**
+         * Can use the enchanting table
+         */
+        ENCHANTING,
+        /**
+         * Can throw ender pearls
+         */
+        ENDER_PEARL,
+        /**
+         * Can toggle enter/exit names to island
+         */
+        ENTER_EXIT_MESSAGES,
+        /**
+         * Fire use/placement in general
+         */
+        FIRE,
+        /**
+         * Can extinguish fires by punching them
+         */
+        FIRE_EXTINGUISH,
+        /**
+         * Allow fire spread
+         */
+        FIRE_SPREAD,
+        /**
+         * Can use furnaces
+         */
+        FURNACE,
+        /**
+         * Can use gates
+         */
+        GATE,
+        /**
+         * Can open horse or other animal inventories, e.g. llama
+         */
+        HORSE_INVENTORY,
+        /**
+         * Can ride an animal
+         */
+        HORSE_RIDING,
+        /**
+         * Can hurt friendly mobs, e.g. cows
+         */
+        HURT_MOBS,
+        /**
+         * Can hurt monsters
+         */
+        HURT_MONSTERS,
+        /**
+         * Can leash or unleash animals
+         */
+        LEASH,
+        /**
+         * Can use buttons or levers
+         */
+        LEVER_BUTTON,
+        /**
+         * Animals, etc. can spawn
+         */
+        MILKING,
+        /**
+         * Can do PVP in the nether
+         */
+        MOB_SPAWN,
+        /**
+         * Monsters can spawn
+         */
+        MONSTER_SPAWN,
+        /**
+         * Can operate jukeboxes, note boxes etc.
+         */
+        MUSIC,
+        /**
+         * Can place blocks
+         */
+        NETHER_PVP,
+        /**
+         * Can interact with redstone items, like diodes
+         */
+        PLACE_BLOCKS,
+        /**
+         * Can go through portals
+         */
+        PORTAL,
+        /**
+         * Will activate pressure plates
+         */
+        PRESSURE_PLATE,
+        /**
+         * Can do PVP in the overworld
+         */
+        PVP,
+        /**
+         * Cows can be milked
+         */
+        REDSTONE,
+        /**
+         * Spawn eggs can be used
+         */
+        SPAWN_EGGS,
+        /**
+         * Can shear sheep
+         */
+        SHEARING,
+        /**
+         * Can trade with villagers
+         */
+        VILLAGER_TRADING,
+        /**
+         * Visitors can drop items
+         */
+        VISITOR_ITEM_DROP,
+        /**
+         * Visitors can pick up items
+         */
+        VISITOR_ITEM_PICKUP
     }
 
 
     /**
      * New island by loading islands.yml
-     * @param plugin
+     * @param plugin - ASkyBlock plugin object
      * @param serial
+     * @param settingsKey
      */
-    public Island(ASkyBlock plugin, String serial) {
+    public Island(ASkyBlock plugin, String serial, List<String> settingsKey) {
         this.plugin = plugin;
         // Bukkit.getLogger().info("DEBUG: adding serialized island to grid ");
         // Deserialize
@@ -158,22 +318,14 @@ public class Island implements Cloneable {
             if (split.length > 6) {
                 // Bukkit.getLogger().info("DEBUG: " + split[6]);
                 // Get locked status
-                if (split[6].equalsIgnoreCase("true")) {
-                    this.locked = true;
-                } else {
-                    this.locked = false;
-                }
+                this.locked = split[6].equalsIgnoreCase("true");
                 // Bukkit.getLogger().info("DEBUG: " + locked);
             } else {
                 this.locked = false;
             }
             // Check if deletable
             if (split.length > 7) {
-                if (split[7].equalsIgnoreCase("true")) {
-                    this.purgeProtected = true;
-                } else {
-                    this.purgeProtected = false;
-                }
+                this.purgeProtected = split[7].equalsIgnoreCase("true");
             } else {
                 this.purgeProtected = false;
             }
@@ -190,57 +342,91 @@ public class Island implements Cloneable {
                 }
             }
             // Check if protection options there
-            if (!isSpawn) {
-                //plugin.getLogger().info("DEBUG: NOT SPAWN owner is " + owner + " location " + center);
-                // Set defaults
-                setDefaults();
-                // Load settings
-                if (split.length > 8) {
-                    // Parse the 8th string into island guard protection settings
-                    int index = 0;
-                    // Run through the enum and set
-                    for (Flags f : Flags.values()) {
-                        if (split[8].length() == index) {
-                            break;
-                        }
-                        this.igs.put(f, split[8].charAt(index++) == '1' ? true : false);
-                    }
-                }
-                // Get the biome
-                if (split.length > 9) {
-                    try {
-                        biome = Biome.valueOf(split[9]);
+            if (split.length > 8) {
+                setSettings(split[8], settingsKey);
+            } else {
+                setSettings(null, settingsKey);
+            }
 
-                    } catch (IllegalArgumentException ee) {
-                        // Unknown biome
-                    }
-                }
-                // Get island level handicap
-                if (split.length > 10) {
-                    try {
-                        this.levelHandicap = Integer.valueOf(split[10]);
-                    } catch (Exception e) {
-                        this.levelHandicap = 0;
-                    }
+            // Get the biome
+            if (split.length > 9) {
+                try {
+                    biome = Biome.valueOf(split[9]);
+
+                } catch (IllegalArgumentException ee) {
+                    // Unknown biome
                 }
             }
+            // Get island level handicap
+            if (split.length > 10) {
+                try {
+                    this.levelHandicap = Integer.valueOf(split[10]);
+                } catch (Exception e) {
+                    this.levelHandicap = 0;
+                }
+            }
+
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
 
     /**
-     * Resets the island protection settings to their default as set in config.yml
+     * Resets the protection settings to their default as set in config.yml for this island
      */
-    public void setDefaults() {
-        for (Flags flag: Flags.values()) {
-            this.igs.put(flag, Settings.defaultIslandSettings.get(flag));
+    public void setIgsDefaults() {
+        for (SettingsFlag flag: SettingsFlag.values()) {
+            if (!Settings.defaultIslandSettings.containsKey(flag)) {
+                // Default default
+                if (flag.equals(SettingsFlag.MOB_SPAWN) || flag.equals(SettingsFlag.MONSTER_SPAWN)) {
+                    this.igs.put(flag, true);
+                } else {
+                    this.igs.put(flag, false);
+                }
+            } else {
+                if (Settings.defaultIslandSettings.get(flag) == null) {
+                    //plugin.getLogger().info("DEBUG: null flag " + flag);
+                    if (flag.equals(SettingsFlag.MOB_SPAWN) || flag.equals(SettingsFlag.MONSTER_SPAWN)) {
+                        this.igs.put(flag, true);
+                    } else {
+                        this.igs.put(flag, false);
+                    }
+                } else {
+                    this.igs.put(flag, Settings.defaultIslandSettings.get(flag));
+                }
+            }
+        }
+    }
+
+    /**
+     * Reset spawn protection settings to their default as set in config.yml for this island
+     */
+    public void setSpawnDefaults() {
+        for (SettingsFlag flag: SettingsFlag.values()) {
+            if (!Settings.defaultSpawnSettings.containsKey(flag)) {
+                // Default default
+                if (flag.equals(SettingsFlag.MOB_SPAWN) || flag.equals(SettingsFlag.MONSTER_SPAWN)) {
+                    this.igs.put(flag, true);
+                } else {
+                    this.igs.put(flag, false);
+                }
+            } else {
+                if (Settings.defaultSpawnSettings.get(flag) == null) {
+                    if (flag.equals(SettingsFlag.MOB_SPAWN) || flag.equals(SettingsFlag.MONSTER_SPAWN)) {
+                        this.igs.put(flag, true);
+                    } else {
+                        this.igs.put(flag, false);
+                    }
+                } else {
+                    this.igs.put(flag, Settings.defaultSpawnSettings.get(flag));
+                }
+            }
         }
     }
 
     /**
      * Add a new island using the island center method
-     * @param plugin
+     * @param plugin - ASkyBlock plugin object
      * @param x
      * @param z
      */
@@ -253,11 +439,11 @@ public class Island implements Cloneable {
         // Calculate min minX and z
         this.minX = x - Settings.islandDistance / 2;
         this.minZ = z - Settings.islandDistance / 2;
-        this.minProtectedX = x - Settings.island_protectionRange / 2;
-        this.minProtectedZ = z - Settings.island_protectionRange / 2;
-        this.y = Settings.island_level;
+        this.minProtectedX = x - Settings.islandProtectionRange / 2;
+        this.minProtectedZ = z - Settings.islandProtectionRange / 2;
+        this.y = Settings.islandHeight;
         this.islandDistance = Settings.islandDistance;
-        this.protectionRange = Settings.island_protectionRange;
+        this.protectionRange = Settings.islandProtectionRange;
         this.world = ASkyBlock.getIslandWorld();
         this.center = new Location(world, x, y, z);
         this.createdDate = new Date().getTime();
@@ -266,47 +452,53 @@ public class Island implements Cloneable {
         this.votes = 0;
         this.owner = owner;
         // Island Guard Settings
-        setDefaults();
+        setIgsDefaults();
     }
 
-    /* (non-Javadoc)
-     * @see java.lang.Object#clone()
+    /**
+     * Copy constructor
+     * @param island - island to copy
      */
-    public Object clone() {
-        try {
-            return super.clone();
-        }
-        catch (CloneNotSupportedException e) {
-            // This should never happen
-            throw new InternalError(e.toString());
-        }
+    public Island(Island island) {
+        this.plugin = island.plugin;
+        this.biome = island.biome == null ? null : Biome.valueOf(island.biome.name());
+        this.center = island.center != null ? island.center.clone() : null;
+        this.createdDate = Long.valueOf(island.createdDate);
+        island.igs.forEach((k,v) -> this.igs.put(k, v));
+        this.islandDistance = Integer.valueOf(island.islandDistance);
+        this.isSpawn = Boolean.valueOf(island.isSpawn);
+        this.locked = Boolean.valueOf(island.locked);
+        this.levelHandicap = Integer.valueOf(island.levelHandicap);
+        this.minProtectedX = Integer.valueOf(island.minProtectedX);
+        this.minProtectedZ = Integer.valueOf(island.minProtectedZ);
+        this.minX = Integer.valueOf(island.minX);
+        this.minZ = Integer.valueOf(island.minZ);
+        this.owner = island.owner == null ? null : UUID.fromString(island.owner.toString());
+        this.password = island.password;
+        this.protectionRange = Integer.valueOf(island.protectionRange);
+        this.purgeProtected = Boolean.valueOf(island.purgeProtected);
+        this.spawnPoint = island.spawnPoint == null ? null : island.spawnPoint.clone();
+        this.tileEntityCount.addAll(island.tileEntityCount);
+        this.updatedDate = Long.valueOf(island.updatedDate);
+        this.votes = Integer.valueOf(island.votes);
+        this.world = island.world == null ? null : Bukkit.getWorld(island.world.getUID());
+        this.y = Integer.valueOf(island.y);
     }
 
     /**
      * Checks if a location is within this island's protected area
-     * 
-     * @param target
+     *
+     * @param target location to query
      * @return true if it is, false if not
      */
     public boolean onIsland(Location target) {
         if (world != null) {
             // If the new nether is being used, islands exist in the nether too
-            //plugin.getLogger().info("DEBUG: target x = " + target.getBlockX() + " target z = " + target.getBlockZ());
-            //plugin.getLogger().info("DEBUG: min prot x = " + minProtectedX + " min z = " + minProtectedZ);
-            //plugin.getLogger().info("DEBUG: max x = " + (minProtectedX + protectionRange) + " max z = " + (minProtectedZ + protectionRange));
-
             if (target.getWorld().equals(world) || (Settings.createNether && Settings.newNether && ASkyBlock.getNetherWorld() != null && target.getWorld().equals(ASkyBlock.getNetherWorld()))) {
-                if (target.getBlockX() >= minProtectedX && target.getBlockX() < (minProtectedX + protectionRange)
-                        && target.getBlockZ() >= minProtectedZ && target.getBlockZ() < (minProtectedZ + protectionRange)) {
-                    return true;
-                }
-                /*
-                if (target.getX() >= center.getBlockX() - protectionRange / 2 && target.getX() < center.getBlockX() + protectionRange / 2
-                        && target.getZ() >= center.getBlockZ() - protectionRange / 2 && target.getZ() < center.getBlockZ() + protectionRange / 2) {
-
-                    return true;
-                }
-                 */
+                return target.getBlockX() >= minProtectedX && target.getBlockX() < (minProtectedX
+                        + protectionRange)
+                        && target.getBlockZ() >= minProtectedZ && target.getBlockZ() < (minProtectedZ
+                                + protectionRange);
             }
         }
         return false;
@@ -314,26 +506,25 @@ public class Island implements Cloneable {
 
     /**
      * Checks if location is anywhere in the island space (island distance)
-     * 
-     * @param target
+     *
+     * @param target location to query
      * @return true if in the area
      */
     public boolean inIslandSpace(Location target) {
         if (target.getWorld().equals(ASkyBlock.getIslandWorld()) || target.getWorld().equals(ASkyBlock.getNetherWorld())) {
-            if (target.getX() >= center.getBlockX() - islandDistance / 2 && target.getX() < center.getBlockX() + islandDistance / 2
-                    && target.getZ() >= center.getBlockZ() - islandDistance / 2 && target.getZ() < center.getBlockZ() + islandDistance / 2) {
-                return true;
-            }
+            return target.getX() >= center.getBlockX() - islandDistance / 2
+                    && target.getX() < center.getBlockX() + islandDistance / 2
+                    && target.getZ() >= center.getBlockZ() - islandDistance / 2
+                    && target.getZ() < center.getBlockZ() + islandDistance / 2;
         }
         return false;
     }
 
     public boolean inIslandSpace(int x, int z) {
-        if (x >= center.getBlockX() - islandDistance / 2 && x < center.getBlockX() + islandDistance / 2 && z >= center.getBlockZ() - islandDistance / 2
-                && z < center.getBlockZ() + islandDistance / 2) {
-            return true;
-        }
-        return false;
+        return x >= center.getBlockX() - islandDistance / 2
+                && x < center.getBlockX() + islandDistance / 2
+                && z >= center.getBlockZ() - islandDistance / 2
+                && z < center.getBlockZ() + islandDistance / 2;
     }
 
     /**
@@ -528,28 +719,41 @@ public class Island implements Cloneable {
      */
     public String save() {
         // x:height:z:protection range:island distance:owner UUID
-        String result = "";
         String ownerString = "null";
         if (isSpawn) {
             // Bukkit.getLogger().info("DEBUG: island is spawn");
             ownerString = "spawn";
             if (spawnPoint != null) {
-                return center.getBlockX() + ":" + center.getBlockY() + ":" + center.getBlockZ() + ":" + protectionRange + ":" 
+                return center.getBlockX() + ":" + center.getBlockY() + ":" + center.getBlockZ() + ":" + protectionRange + ":"
                         + islandDistance + ":" + ownerString + ":" + locked + ":" + purgeProtected + ":SP:" + Util.getStringLocation(spawnPoint);
             }
-            return center.getBlockX() + ":" + center.getBlockY() + ":" + center.getBlockZ() + ":" + protectionRange + ":" 
+            return center.getBlockX() + ":" + center.getBlockY() + ":" + center.getBlockZ() + ":" + protectionRange + ":"
             + islandDistance + ":" + ownerString + ":" + locked + ":" + purgeProtected;
         }
         // Not spawn
         if (owner != null) {
             ownerString = owner.toString();
         }
+
+        return center.getBlockX() + ":" + center.getBlockY() + ":" + center.getBlockZ() + ":" + protectionRange + ":"
+        + islandDistance + ":" + ownerString + ":" + locked + ":" + purgeProtected + ":" + getSettings() + ":" + getBiome().toString() + ":" + levelHandicap;
+    }
+
+    /**
+     * @return Serialized set of settings
+     */
+    public String getSettings() {
+        String result = "";
         // Personal island protection settings - serialize enum into 1's and 0's representing the boolean values
+        //plugin.getLogger().info("DEBUG: igs = " + igs.toString());
         try {
-            for (Flags f: Flags.values()) {
+            for (SettingsFlag f: SettingsFlag.values()) {
+                //plugin.getLogger().info("DEBUG: flag f = " + f);
                 if (this.igs.containsKey(f)) {
+                    //plugin.getLogger().info("DEBUG: contains key");
                     result += this.igs.get(f) ? "1" : "0";
                 } else {
+                    //plugin.getLogger().info("DEBUG: does not contain key");
                     result += "0";
                 }
             }
@@ -557,16 +761,15 @@ public class Island implements Cloneable {
             e.printStackTrace();
             result = "";
         }
-        return center.getBlockX() + ":" + center.getBlockY() + ":" + center.getBlockZ() + ":" + protectionRange + ":" 
-        + islandDistance + ":" + ownerString + ":" + locked + ":" + purgeProtected + ":" + result + ":" + getBiome().toString() + ":" + levelHandicap;
+        return result;
     }
 
     /**
      * Get the Island Guard flag status
-     * @param flag
+     * @param flag - settings flag to check
      * @return true or false, or false if flag is not in the list
      */
-    public boolean getIgsFlag(Flags flag) {
+    public boolean getIgsFlag(SettingsFlag flag) {
         //plugin.getLogger().info("DEBUG: asking for " + flag + " = " + igs.get(flag));
         if (this.igs.containsKey(flag)) {
             return igs.get(flag);
@@ -576,17 +779,17 @@ public class Island implements Cloneable {
 
     /**
      * Set the Island Guard flag
-     * @param flag
-     * @param value
+     * @param flag - settings flag to check
+     * @param value - value to set true or false
      */
-    public void setIgsFlag(Flags flag, boolean value) {
+    public void setIgsFlag(SettingsFlag flag, boolean value) {
         this.igs.put(flag, value);
     }
 
     /**
      * Provides a list of all the players who are allowed on this island
      * including coop members
-     * 
+     *
      * @return a list of UUIDs that have legitimate access to the island
      */
     public List<UUID> getMembers() {
@@ -620,33 +823,6 @@ public class Island implements Cloneable {
         this.isSpawn = isSpawn;
     }
 
-    public void addEntity(EntityType type) {
-        if (this.entities.containsKey(type)) {
-            int sum = this.entities.get(type);
-            this.entities.put(type, (sum + 1));
-        } else {
-            this.entities.put(type, 1);
-        }
-    }
-
-    public int getEntity(EntityType type) {
-        if (this.entities.containsKey(type)) {
-            return this.entities.get(type);
-        }
-        return 0;
-    }
-
-    /**
-     * @return the entities
-     */
-    public HashMap<EntityType, Integer> getEntities() {
-        return entities;
-    }
-
-    public void clearStats() {
-        this.entities.clear();
-    }
-
     /**
      * @return the islandDeletable
      */
@@ -665,7 +841,7 @@ public class Island implements Cloneable {
      * @return Provides count of villagers within the protected island boundaries
      */
     public int getPopulation() {
-        int result = 0;	
+        int result = 0;
         for (int x = getMinProtectedX() /16; x <= (getMinProtectedX() + getProtectionSize() - 1)/16; x++) {
             for (int z = getMinProtectedZ() /16; z <= (getMinProtectedZ() + getProtectionSize() - 1)/16; z++) {
                 for (Entity entity : world.getChunkAt(x, z).getEntities()) {
@@ -673,7 +849,7 @@ public class Island implements Cloneable {
                         result++;
                     }
                 }
-            }  
+            }
         }
         return result;
     }
@@ -683,7 +859,7 @@ public class Island implements Cloneable {
      */
     public int getHopperCount() {
         tileEntityCount.clear();
-        int result = 0;	
+        int result = 0;
         for (int x = getMinProtectedX() /16; x <= (getMinProtectedX() + getProtectionSize() - 1)/16; x++) {
             for (int z = getMinProtectedZ() /16; z <= (getMinProtectedZ() + getProtectionSize() - 1)/16; z++) {
                 for (BlockState holder : world.getChunkAt(x, z).getTileEntities()) {
@@ -691,18 +867,19 @@ public class Island implements Cloneable {
                         result++;
                     }
                 }
-            }  
+            }
         }
         return result;
     }
 
     /**
-     * @param material
+     * @param material Bukkit material to check
+     * @param world - world to check
      * @return count of how many tile entities of type mat are on the island at last count. Counts are done when a player places
      * a tile entity.
      */
-    public int getTileEntityCount(Material material) {
-        int result = 0;	
+    public int getTileEntityCount(Material material, World world) {
+        int result = 0;
         for (int x = getMinProtectedX() /16; x <= (getMinProtectedX() + getProtectionSize() - 1)/16; x++) {
             for (int z = getMinProtectedZ() /16; z <= (getMinProtectedZ() + getProtectionSize() - 1)/16; z++) {
                 for (BlockState holder : world.getChunkAt(x, z).getTileEntities()) {
@@ -735,7 +912,7 @@ public class Island implements Cloneable {
                         result++;
                     }
                 }
-            }  
+            }
         }
         // Version 1.7.x counts differently to 1.8 (ugh)
         // In 1.7, the entity is present before it is cancelled and so gets counted.
@@ -757,11 +934,11 @@ public class Island implements Cloneable {
 
     /**
      * Toggles the Island Guard Flag
-     * @param flag
+     * @param flag  - settings flag to toggle
      */
-    public void toggleIgs(Flags flag) {
+    public void toggleIgs(SettingsFlag flag) {
         if (igs.containsKey(flag)) {
-            igs.put(flag, igs.get(flag) ? false : true);
+            igs.put(flag, !igs.get(flag));
         }
 
     }
@@ -796,4 +973,38 @@ public class Island implements Cloneable {
     public void setLevelHandicap(int levelHandicap) {
         this.levelHandicap = levelHandicap;
     }
+
+    /**
+     * Sets the settings for the island.
+     * @param settings - string of 0 and 1's that define the settings for the island
+     * @param settingsKey - a list showing the order and what flags each digit refers to
+     */
+    public void setSettings(String settings, List<String> settingsKey) {
+
+        // Start with defaults
+        if (isSpawn) {
+            setSpawnDefaults();
+        } else {
+            setIgsDefaults();
+        }
+        if(settings == null || settings.isEmpty())
+            return;
+        if (settingsKey.size() != settings.length()) {
+            plugin.getLogger().severe("Island settings does not match settings key in islands.yml. Using defaults.");
+            return;
+        }
+        for (int i = 0; i < settingsKey.size(); i++) {
+            try {
+                if (settings.charAt(i) == '0') {
+                    this.setIgsFlag(SettingsFlag.valueOf(settingsKey.get(i)), false);
+                } else {
+                    this.setIgsFlag(SettingsFlag.valueOf(settingsKey.get(i)), true);
+                }
+            } catch (Exception e) {
+                // do nothing - bad value, probably a downgrade
+            }
+        }
+
+    }
+
 }
