@@ -18,9 +18,10 @@
 package com.wasteofplastic.askyblock;
 
 import java.util.ArrayList;
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
 import org.bukkit.configuration.file.YamlConfiguration;
@@ -35,14 +36,14 @@ import com.wasteofplastic.askyblock.util.Util;
  * 
  */
 public class Messages {
-    private ASkyBlock plugin;
+    private final ASkyBlock plugin;
     // Offline Messages
-    private HashMap<UUID, List<String>> messages = new HashMap<UUID, List<String>>();
+    private final HashMap<UUID, List<String>> messages = new HashMap<>();
     private YamlConfiguration messageStore;
 
 
     /**
-     * @param plugin
+     * @param plugin - ASkyBlock plugin object
      */
     public Messages(ASkyBlock plugin) {
         this.plugin = plugin;
@@ -51,41 +52,40 @@ public class Messages {
     /**
      * Returns what messages are waiting for the player or null if none
      * 
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return List of messages
      */
     public List<String> getMessages(UUID playerUUID) {
-        List<String> playerMessages = messages.get(playerUUID);
-        return playerMessages;
+        return messages.get(playerUUID);
     }
 
     /**
      * Clears any messages for player
      * 
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      */
     public void clearMessages(UUID playerUUID) {
         messages.remove(playerUUID);
     }
 
-    public void saveMessages() {
+    public void saveMessages(boolean async) {
         if (messageStore == null) {
             return;
         }
         plugin.getLogger().info("Saving offline messages...");
         try {
             // Convert to a serialized string
-            final HashMap<String, Object> offlineMessages = new HashMap<String, Object>();
+            Map<String, Object> offlineMessages = new HashMap<>();
             for (UUID p : messages.keySet()) {
-                offlineMessages.put(p.toString(), messages.get(p));
+                if (p != null) {
+                    offlineMessages.put(p.toString(), messages.get(p));
+                }
             }
             // Convert to YAML
             messageStore.set("messages", offlineMessages);
-            Util.saveYamlFile(messageStore, "messages.yml");
-            return;
+            Util.saveYamlFile(messageStore, "messages.yml", async);
         } catch (Exception e) {
             e.printStackTrace();
-            return;
         }
     }
 
@@ -114,7 +114,7 @@ public class Messages {
     /**
      * Provides the messages for the player
      * 
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return List of messages
      */
     public List<String> get(UUID playerUUID) {
@@ -124,7 +124,7 @@ public class Messages {
     /**
      * Stores a message for player
      * 
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @param playerMessages
      */
     public void put(UUID playerUUID, List<String> playerMessages) {
@@ -135,8 +135,8 @@ public class Messages {
     /**
      * Sends a message to every player in the team that is offline
      * 
-     * @param playerUUID
-     * @param message
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param message - message to send
      */
     public void tellOfflineTeam(UUID playerUUID, String message) {
         // getLogger().info("DEBUG: tell offline team called");
@@ -146,20 +146,17 @@ public class Messages {
         }
         UUID teamLeader = plugin.getPlayers().getTeamLeader(playerUUID);
         List<UUID> teamMembers = plugin.getPlayers().getMembers(teamLeader);
-        for (UUID member : teamMembers) {
-            // getLogger().info("DEBUG: trying UUID " + member.toString());
-            if (plugin.getServer().getPlayer(member) == null) {
-                // Offline player
-                setMessage(member, message);
-            }
-        }
+        // getLogger().info("DEBUG: trying UUID " + member.toString());
+        // Offline player
+        teamMembers.stream().filter(member -> plugin.getServer().getPlayer(member) == null)
+            .forEach(member -> setMessage(member, message));
     }
 
     /**
      * Tells all online team members something happened
      * 
-     * @param playerUUID
-     * @param message
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param message - message to send
      */
     public void tellTeam(UUID playerUUID, String message) {
         // getLogger().info("DEBUG: tell offline team called");
@@ -173,7 +170,7 @@ public class Messages {
             // getLogger().info("DEBUG: trying UUID " + member.toString());
             if (!member.equals(playerUUID) && plugin.getServer().getPlayer(member) != null) {
                 // Online player
-                plugin.getServer().getPlayer(member).sendMessage(message);
+                Util.sendMessage(plugin.getServer().getPlayer(member), message);
             }
         }
     }
@@ -181,8 +178,8 @@ public class Messages {
     /**
      * Sets a message for the player to receive next time they login
      * 
-     * @param playerUUID
-     * @param message
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param message - message to set
      * @return true if player is offline, false if online
      */
     public boolean setMessage(UUID playerUUID, String message) {
@@ -191,20 +188,26 @@ public class Messages {
         // Check if player is online
         if (player != null) {
             if (player.isOnline()) {
-                // player.sendMessage(message);
+                // Util.sendMessage(player, message);
                 return false;
             }
         }
-        // Player is offline so store the message
-        // getLogger().info("DEBUG: player is offline - storing message");
+        storeMessage(playerUUID, message);
+        return true;
+    }
+
+    /**
+     * Stores a message without any online check
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param message - message to store
+     */
+    public void storeMessage(UUID playerUUID, String message) {
         List<String> playerMessages = get(playerUUID);
         if (playerMessages != null) {
             playerMessages.add(message);
         } else {
-            playerMessages = new ArrayList<String>(Arrays.asList(message));
+            playerMessages = new ArrayList<>(Collections.singletonList(message));
         }
         put(playerUUID, playerMessages);
-        return true;
     }
-
 }
