@@ -21,53 +21,67 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
-import org.bukkit.Bukkit;
 import org.bukkit.Location;
-import org.bukkit.World;
+import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 
-import com.google.common.collect.Lists;
+import com.google.common.collect.ImmutableMap;
 import com.wasteofplastic.askyblock.util.VaultHelper;
 
 /**
  * Provides a memory cache of online player information
  * This is the one-stop-shop of player info
  * If the player is not cached, then a request is made to Players to obtain it
- * 
+ *
  * @author tastybento
  */
 public class PlayerCache {
-    private HashMap<UUID, Players> playerCache = new HashMap<UUID, Players>();
+
+    private final Map<UUID, Players> playerCache = new HashMap<>();
     private final ASkyBlock plugin;
-    private Set<UUID> inTeleport = new HashSet<UUID>();
+    private final Set<UUID> inTeleport = new HashSet<>();
 
     public PlayerCache(ASkyBlock plugin) {
         this.plugin = plugin;
         // final Collection<? extends Player> serverPlayers =
         // Bukkit.getServer().getOnlinePlayers();
-        for (Player p : getOnlinePlayers()) {
-            if (p.isOnline()) {
-                final Players playerInf = new Players(plugin, p.getUniqueId());
-                // Make sure parties are working correctly
-                if (playerInf.inTeam() && playerInf.getTeamIslandLocation() == null) {
-                    final Players leaderInf = new Players(plugin, playerInf.getTeamLeader());
-                    playerInf.setTeamIslandLocation(leaderInf.getIslandLocation());
-                    playerInf.save();
+        for (Player p : plugin.getServer().getOnlinePlayers()) {
+            if (p.getUniqueId() != null) {
+                try {
+                    final Players playerInf = new Players(plugin, p.getUniqueId());
+                    // Make sure parties are working correctly
+                    if (playerInf.inTeam() && playerInf.getTeamIslandLocation() == null) {
+                        if (playerInf.getTeamLeader() == null) {
+                            // Player cannot be in a team - try to clean up
+                            playerInf.setLeaveTeam();
+                        } else {
+                            final Players leaderInf = new Players(plugin, playerInf.getTeamLeader());
+                            playerInf.setTeamIslandLocation(leaderInf.getIslandLocation());
+                        }
+                        playerInf.save(false);
+                    }
+                    // Add this player to the online cache
+                    //plugin.getLogger().info("DEBUG: added player " + p.getUniqueId());
+                    playerCache.put(p.getUniqueId(), playerInf);
+                } catch (Exception e) {
+                    plugin.getLogger().severe("Player add tried for a player with null UUID");
                 }
-                // Add this player to the online cache
-                playerCache.put(p.getUniqueId(), playerInf);
             }
         }
     }
 
-    public static List<Player> getOnlinePlayers() {
-        List<Player> list = Lists.newArrayList();
-        for (World world : Bukkit.getWorlds()) {
-            list.addAll(world.getPlayers());
-        }
+    /**
+     * @return list of all online cached players
+     */
+    public List<UUID> getOnlineCachedPlayers() {
+        List<UUID> list = plugin.getServer().getOnlinePlayers().stream()
+                .filter(p -> playerCache.containsKey(p.getUniqueId())).map(Entity::getUniqueId)
+                .collect(Collectors.toList());
         return Collections.unmodifiableList(list);
     }
 
@@ -76,23 +90,28 @@ public class PlayerCache {
      */
 
     public void addPlayer(final UUID playerUUID) {
-        // plugin.getLogger().info("DEBUG: added player");
+        //plugin.getLogger().info("DEBUG: added player " + playerUUID);
         if (!playerCache.containsKey(playerUUID)) {
-            final Players player = new Players(plugin, playerUUID);
-            playerCache.put(playerUUID, player);
+            try {
+                final Players player = new Players(plugin, playerUUID);
+                playerCache.put(playerUUID, player);
+            } catch (Exception e) {
+                // Be silent.
+                //plugin.getLogger().severe("Player add request for a null UUID");
+            }
         }
     }
 
     /**
      * Stores the player's info to a file and removes the player from the list
      * of currently online players
-     * 
+     *
      * @param player
      *            - name of player
      */
     public void removeOnlinePlayer(final UUID player) {
         if (playerCache.containsKey(player)) {
-            playerCache.get(player).save();
+            playerCache.get(player).save(false);
             playerCache.remove(player);
             // plugin.getLogger().info("Removing player from cache: " + player);
         }
@@ -102,9 +121,8 @@ public class PlayerCache {
      * Removes all players on the server now from cache and saves their info
      */
     public void removeAllPlayers() {
-        for (UUID pl : playerCache.keySet()) {
-            playerCache.get(pl).save();
-        }
+        Map<UUID, Players> map = ImmutableMap.copyOf(playerCache);
+        map.keySet().forEach(player -> map.get(player).save(false));
         playerCache.clear();
     }
 
@@ -113,8 +131,8 @@ public class PlayerCache {
      */
     /**
      * Returns location of player's island from cache if available
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      * @return Location of player's island
      */
     /*
@@ -128,8 +146,8 @@ public class PlayerCache {
      */
     /**
      * Checks if the player is known or not by looking through the filesystem
-     * 
-     * @param uniqueID
+     *
+     * @param uniqueID - unique ID
      * @return true if player is know, otherwise false
      */
     public boolean isAKnownPlayer(final UUID uniqueID) {
@@ -140,30 +158,19 @@ public class PlayerCache {
             return true;
         } else {
             // Get the file system
-            final File folder = plugin.getPlayersFolder();
-            final File[] files = folder.listFiles();
-            // Go through the native YAML files
-            for (final File f : files) {
-                // Need to remove the .yml suffix
-                if (f.getName().endsWith(".yml")) {
-                    try {
-                        if (UUID.fromString(f.getName().substring(0, f.getName().length() - 4)).equals(uniqueID)) {
-                            return true;
-                        }
-                    } catch (Exception e) {
-                        plugin.getLogger().warning("Problem with " + f.getName());
-                    }
-                }
+            try {
+                final File file = new File(plugin.getPlayersFolder(), uniqueID.toString() + ".yml");
+                return file.exists();
+            } catch (Exception e) {
+                return false;
             }
         }
-        // Not found, sorry.
-        return false;
     }
 
     /**
      * Returns the player object for the named player
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      *            - String name of player
      * @return - player object
      */
@@ -174,8 +181,8 @@ public class PlayerCache {
 
     /**
      * Checks if player has island from cache if available
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      *            - string name of player
      * @return true if player has island
      */
@@ -188,8 +195,8 @@ public class PlayerCache {
 
     /**
      * Checks if player is in a Team from cache if available
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      * @return true if player in a team
      */
     public boolean inTeam(final UUID playerUUID) {
@@ -200,8 +207,8 @@ public class PlayerCache {
     /**
      * Removes any island associated with this player and generally cleans up
      * the player
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      */
     public void zeroPlayerData(UUID playerUUID) {
         addPlayer(playerUUID);
@@ -220,7 +227,7 @@ public class PlayerCache {
                     // Just remove them from the team
                     addPlayer(leader);
                     playerCache.get(leader).removeMember(playerUUID);
-                    playerCache.get(leader).save();
+                    playerCache.get(leader).save(false);
                 }
             }
         }
@@ -229,13 +236,13 @@ public class PlayerCache {
         playerCache.get(playerUUID).clearHomeLocations();
         playerCache.get(playerUUID).setIslandLocation(null);
         playerCache.get(playerUUID).setIslandLevel(0);
-        playerCache.get(playerUUID).save(); // Needed?
-        TopTen.topTenRemoveEntry(playerUUID);
+        playerCache.get(playerUUID).save(false); // Needed?
+        plugin.getTopTen().topTenRemoveEntry(playerUUID);
     }
 
     /**
      * Sets the home location for the player
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @param location
      * @param number - 1 is default. Can be any number.
      */
@@ -246,7 +253,7 @@ public class PlayerCache {
 
     /**
      * Set the default home location for player
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @param location
      */
     public void setHomeLocation(UUID playerUUID, Location location) {
@@ -256,7 +263,7 @@ public class PlayerCache {
 
     /**
      * Clears any home locations for player
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      */
     public void clearHomeLocations(UUID playerUUID) {
         addPlayer(playerUUID);
@@ -265,9 +272,9 @@ public class PlayerCache {
 
     /**
      * Returns the home location, or null if none
-     * 
-     * @param playerUUID
-     * @param number 
+     *
+     * @param playerUUID - the player's UUID
+     * @param number
      * @return Home location or null if none
      */
     public Location getHomeLocation(UUID playerUUID, int number) {
@@ -277,7 +284,7 @@ public class PlayerCache {
 
     /**
      * Gets the default home location for player
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @return Home location or null if none
      */
     public Location getHomeLocation(UUID playerUUID) {
@@ -287,7 +294,7 @@ public class PlayerCache {
 
     /**
      * Provides all home locations for player
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @return List of home locations
      */
     public HashMap<Integer, Location> getHomeLocations(UUID playerUUID) {
@@ -298,8 +305,8 @@ public class PlayerCache {
     /**
      * Returns the player's island location.
      * Returns an island location OR a team island location
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      * @return Location of player's island
      */
     public Location getIslandLocation(UUID playerUUID) {
@@ -319,15 +326,15 @@ public class PlayerCache {
         playerCache.get(playerUUID).setIslandLocation(islandLocation);
     }
 
-    public Integer getIslandLevel(UUID playerUUID) {
+    public long getIslandLevel(UUID playerUUID) {
         addPlayer(playerUUID);
         return playerCache.get(playerUUID).getIslandLevel();
     }
 
-    public void setIslandLevel(UUID playerUUID, int islandLevel) {
+    public void setIslandLevel(UUID playerUUID, long l) {
         addPlayer(playerUUID);
-        playerCache.get(playerUUID).setIslandLevel(islandLevel);
-        plugin.getChatListener().setPlayerLevel(playerUUID, islandLevel);
+        playerCache.get(playerUUID).setIslandLevel(l);
+        plugin.getChatListener().setPlayerLevel(playerUUID, l);
     }
 
     public void setTeamIslandLocation(UUID playerUUID, Location islandLocation) {
@@ -337,8 +344,8 @@ public class PlayerCache {
 
     /**
      * Checks if a challenge has been completed or not
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      * @param challenge
      * @return true if complete
      */
@@ -349,8 +356,8 @@ public class PlayerCache {
 
     /**
      * Checks how often a challenge has been completed
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      * @param challenge
      * @return number of times
      */
@@ -361,22 +368,22 @@ public class PlayerCache {
 
     /**
      * Provides the status of all challenges for this player
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      * @return Hashmap of challenges as key, boolean as state
      */
-    public HashMap<String, Boolean> getChallengeStatus(UUID playerUUID) {
+    public Map<String, Boolean> getChallengeStatus(UUID playerUUID) {
         addPlayer(playerUUID);
         return playerCache.get(playerUUID).getChallengeStatus();
     }
 
     /**
      * How many times a challenge has been completed
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID
      * @return map of completion times
      */
-    public HashMap<String, Integer> getChallengeTimes(UUID playerUUID) {
+    public Map<String, Integer> getChallengeTimes(UUID playerUUID) {
         addPlayer(playerUUID);
         return playerCache.get(playerUUID).getChallengeCompleteTimes();
     }
@@ -389,7 +396,7 @@ public class PlayerCache {
     /**
      * Resets all the player's challenges. If the boolean is true, then everything will be reset, if false
      * challenges that have the "resetallowed: false" flag in challenges.yml will not be reset
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @param resetAll
      */
     public void resetAllChallenges(UUID playerUUID, boolean resetAll) {
@@ -407,26 +414,35 @@ public class PlayerCache {
         }
     }
 
-    public void setJoinTeam(UUID playerUUID, UUID teamLeader, Location islandLocation) {
+    /**
+     * Puts a player in a team
+     * @param playerUUID - the player's UUID
+     * @param teamLeader
+     * @param islandLocation
+     * @return true if successful, false if not
+     */
+    public boolean setJoinTeam(UUID playerUUID, UUID teamLeader, Location islandLocation) {
         addPlayer(playerUUID);
         addPlayer(teamLeader);
-        playerCache.get(playerUUID).setJoinTeam(teamLeader, islandLocation);
+        return playerCache.get(playerUUID).setJoinTeam(teamLeader, islandLocation);
     }
 
     /**
      * Called when a player leaves a team Resets inTeam, teamLeader,
      * islandLevel, teamIslandLocation, islandLocation and members array
+     * @param playerUUID - the player's UUID - player's UUID
+     * @return - true if successful
      */
-    public void setLeaveTeam(UUID playerUUID) {
+    public boolean setLeaveTeam(UUID playerUUID) {
         addPlayer(playerUUID);
-        playerCache.get(playerUUID).setLeaveTeam();
+        return playerCache.get(playerUUID).setLeaveTeam();
     }
 
     /**
      * Returns a list of team member UUID's. If the player is not the leader,
      * then the leader's list is used
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      * @return List of team UUIDs
      */
     public List<UUID> getMembers(UUID playerUUID) {
@@ -447,16 +463,20 @@ public class PlayerCache {
     }
 
     public void removeMember(UUID teamLeader, UUID playerUUID) {
-        addPlayer(teamLeader);
+        if (teamLeader != null) {
+            addPlayer(teamLeader);
+        }
         addPlayer(playerUUID);
-        playerCache.get(teamLeader).removeMember(playerUUID);
+        if (teamLeader != null) {
+            playerCache.get(teamLeader).removeMember(playerUUID);
+        }
         // Remove from team chat too
         plugin.getChatListener().unSetPlayer(playerUUID);
     }
 
     /**
      * Provides UUID of this player's team leader or null if it does not exist
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return UUID of leader
      */
     public UUID getTeamLeader(UUID playerUUID) {
@@ -466,11 +486,11 @@ public class PlayerCache {
 
     /**
      * Saves the player's info to the file system
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      */
     public void save(UUID playerUUID) {
-        playerCache.get(playerUUID).save();
+        playerCache.get(playerUUID).save(false);
         // Save the name + UUID in the database if it ready
 
         if (plugin.getTinyDB() != null && plugin.getTinyDB().isDbReady()) {
@@ -490,7 +510,7 @@ public class PlayerCache {
 
     /**
      * Attempts to return a UUID for a given player's name. Only uses online or cached information.
-     * @param string
+     * @param string - player's name
      * @return UUID of player or null if unknown
      */
     public UUID getUUID(String string) {
@@ -499,24 +519,29 @@ public class PlayerCache {
 
     /**
      * Attempts to return a UUID for a given player's name
-     * @param string
+     * @param string - player's name
      * @param adminCheck - if made via an admin call, this will go out to the 'net and grab - may cause lag
      * @return UUID of player or null if unknown
      */
     @SuppressWarnings("deprecation")
     public UUID getUUID(String string, boolean adminCheck) {
-        for (UUID id : playerCache.keySet()) {
-            String name = playerCache.get(id).getPlayerName();
-            // plugin.getLogger().info("DEBUG: Testing name " + name);
-            if (name != null && name.equalsIgnoreCase(string)) {
-                return id;
-            }
-        }
         // Look in the database if it ready
         if (plugin.getTinyDB() != null && plugin.getTinyDB().isDbReady()) {
             UUID result = plugin.getTinyDB().getPlayerUUID(string);
             if (result != null) {
                 return result;
+            }
+        }
+        // This goes after the database because it is possible for islands that have a duplicate name to be in
+        // the cache. For example, Bill had an island but left. Bill changes his name to Bob. Then Alice changes
+        // her name to Bill and logs into the game. There are now two islands with owner names called "Bill"
+        // The name database will ensure the names are updated.
+        for (UUID id : playerCache.keySet()) {
+            String name = playerCache.get(id).getPlayerName();
+            //plugin.getLogger().info("DEBUG: Testing name " + name);
+            if (name != null && name.equalsIgnoreCase(string)) {
+                //plugin.getLogger().info("DEBUG: found it! " + id);
+                return id;
             }
         }
         // Try the server
@@ -527,15 +552,15 @@ public class PlayerCache {
     }
 
     /**
-     * Sets the player's name and updates the name>UUID database is up to date
-     * @param uniqueId
-     * @param name
+     * Sets the player's name and updates the name to UUID database so it is up to date
+     * @param uniqueId - unique ID  - player's UUID
+     * @param name - name to set
      */
     public void setPlayerName(UUID uniqueId, String name) {
         addPlayer(uniqueId);
         playerCache.get(uniqueId).setPlayerN(name);
         // Save the name in the name database. Note that the old name will still work until someone takes it
-        // This feature enables admins to locate 'fugitive' players even if they change their name        
+        // This feature enables admins to locate 'fugitive' players even if they change their name
         if (plugin.getTinyDB() != null && plugin.getTinyDB().isDbReady()) {
             plugin.getTinyDB().savePlayerName(name, uniqueId);
         }
@@ -544,8 +569,8 @@ public class PlayerCache {
     /**
      * Obtains the name of the player from their UUID
      * Player must have logged into the game before
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID - player's UUID
      * @return String - playerName
      */
     public String getName(UUID playerUUID) {
@@ -563,8 +588,8 @@ public class PlayerCache {
 
     /**
      * Reverse lookup - returns the owner of an island from the location
-     * 
-     * @param loc
+     *
+     * @param loc - location to query
      * @return UUID of owner of island
      */
     public UUID getPlayerFromIslandLocation(Location loc) {
@@ -580,8 +605,8 @@ public class PlayerCache {
 
     /**
      * Gets how many island resets the player has left
-     * 
-     * @param playerUUID
+     *
+     * @param playerUUID - the player's UUID  - player's UUID
      * @return number of resets
      */
     public int getResetsLeft(UUID playerUUID) {
@@ -591,9 +616,9 @@ public class PlayerCache {
 
     /**
      * Sets how many resets the player has left
-     * 
-     * @param playerUUID
-     * @param resets
+     *
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param resets - value to set
      */
     public void setResetsLeft(UUID playerUUID, int resets) {
         addPlayer(playerUUID);
@@ -603,9 +628,9 @@ public class PlayerCache {
     /**
      * Returns how long the player must wait before they can be invited to an
      * island with the location
-     * 
-     * @param playerUUID
-     * @param location
+     *
+     * @param playerUUID - the player's UUID  - player's UUID
+     * @param location - location to query
      * @return time to wait in minutes/hours
      */
     public long getInviteCoolDownTime(UUID playerUUID, Location location) {
@@ -617,9 +642,9 @@ public class PlayerCache {
      * Starts the timer for the player for this location before which they can
      * be invited
      * Called when they are kicked from an island or leave.
-     * 
-     * @param playerUUID
-     * @param location
+     *
+     * @param playerUUID - the player's UUID  - player's UUID
+     * @param location - location to set
      */
     public void startInviteCoolDownTimer(UUID playerUUID, Location location) {
         addPlayer(playerUUID);
@@ -628,7 +653,7 @@ public class PlayerCache {
 
     /**
      * Returns the locale for this player. If missing, will return nothing
-     * @param playerUUID
+     * @param playerUUID - the player's UUID  - player's UUID
      * @return name of the locale this player uses
      */
     public String getLocale(UUID playerUUID) {
@@ -641,8 +666,8 @@ public class PlayerCache {
 
     /**
      * Sets the locale this player wants to use
-     * @param playerUUID
-     * @param localeName
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param localeName - locale to set
      */
     public void setLocale(UUID playerUUID, String localeName) {
         addPlayer(playerUUID);
@@ -651,8 +676,8 @@ public class PlayerCache {
 
     /**
      * The rating of the initial starter island out of 100. Default is 50
-     * @param playerUUID
-     * @return rating
+     * @param playerUUID - the player's UUID  - player's UUID
+     * @return rating the rating
      */
     public int getStartIslandRating(UUID playerUUID) {
         addPlayer(playerUUID);
@@ -661,8 +686,8 @@ public class PlayerCache {
 
     /**
      * Record the island rating that the player started with
-     * @param playerUUID
-     * @param rating
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param rating the rating
      */
     public void setStartIslandRating(UUID playerUUID, int rating) {
         addPlayer(playerUUID);
@@ -671,7 +696,7 @@ public class PlayerCache {
 
     /**
      * Clear the starter island rating from the player's record
-     * @param playerUUID
+     * @param playerUUID - the player's UUID  - player's UUID
      */
     public void clearStartIslandRating(UUID playerUUID) {
         setStartIslandRating(playerUUID, 0);
@@ -679,8 +704,8 @@ public class PlayerCache {
 
     /**
      * Ban target from a player's island
-     * @param playerUUID
-     * @param targetUUID
+     * @param playerUUID - the player's UUID  - player's UUID
+     * @param targetUUID - target's UUID
      */
     public void ban(UUID playerUUID, UUID targetUUID) {
         addPlayer(playerUUID);
@@ -689,20 +714,20 @@ public class PlayerCache {
             // Player has island
             playerCache.get(playerUUID).addToBanList(targetUUID);
         } else if (playerCache.get(playerUUID).inTeam()) {
-            // Try to get the leader's 
+            // Try to get the leader's
             UUID leader = playerCache.get(playerUUID).getTeamLeader();
             if (leader != null) {
                 addPlayer(leader);
                 playerCache.get(leader).addToBanList(targetUUID);
-                playerCache.get(leader).save();
+                playerCache.get(leader).save(false);
             }
         }
     }
 
     /**
      * Unban target from player's island
-     * @param playerUUID
-     * @param targetUUID
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param targetUUID - target's UUID
      */
     public void unBan(UUID playerUUID, UUID targetUUID) {
         addPlayer(playerUUID);
@@ -711,19 +736,19 @@ public class PlayerCache {
             // Player has island
             playerCache.get(playerUUID).unBan(targetUUID);
         } else if (playerCache.get(playerUUID).inTeam()) {
-            // Try to get the leader's 
+            // Try to get the leader's
             UUID leader = playerCache.get(playerUUID).getTeamLeader();
             if (leader != null) {
                 addPlayer(leader);
                 playerCache.get(leader).unBan(targetUUID);
-                playerCache.get(leader).save();
+                playerCache.get(leader).save(false);
             }
         }
     }
 
     /**
-     * @param playerUUID
-     * @param targetUUID
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param targetUUID - target's UUID
      * @return true if target is banned from player's island
      */
     public boolean isBanned(UUID playerUUID, UUID targetUUID) {
@@ -742,7 +767,7 @@ public class PlayerCache {
             // Player has island
             return playerCache.get(playerUUID).isBanned(targetUUID);
         } else if (playerCache.get(playerUUID).inTeam()) {
-            // Try to get the leader's 
+            // Try to get the leader's
             UUID leader = playerCache.get(playerUUID).getTeamLeader();
             if (leader != null) {
                 addPlayer(leader);
@@ -753,7 +778,7 @@ public class PlayerCache {
     }
 
     /**
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return ban list for player
      */
     public List<UUID> getBanList(UUID playerUUID) {
@@ -763,17 +788,18 @@ public class PlayerCache {
 
     /**
      * Clears resets for online players or players in the cache
-     * @param resetLimit
+     * @param resetLimit - reset limit
      */
     public void clearResets(int resetLimit) {
         for (Players player : playerCache.values()) {
             player.setResetsLeft(resetLimit);
-        }	
+        }
     }
 
     /**
      * Sets whether the player uses the control panel or not when doing /island
-     * @param b
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param b - true = user control panel
      */
     public void setControlPanel(UUID playerUUID, boolean b) {
         addPlayer(playerUUID);
@@ -783,7 +809,7 @@ public class PlayerCache {
 
     /**
      * Gets whether the player uses the control panel or not when doing /island
-     * @param playerUUID
+     * @param playerUUID - the player's UUID  - player's UUID
      * @return true if they use the control panel
      */
     public boolean getControlPanel(UUID playerUUID) {
@@ -794,20 +820,20 @@ public class PlayerCache {
 
     /**
      * Map storing whether a player is officially teleporting or not
-     * @param uniqueId
-     * @param b
+     * @param uniqueId - unique ID  - player's UUID
+     * @param b - true if player is teleporting
      */
     public void setInTeleport(UUID uniqueId, boolean b) {
         if (b) {
             inTeleport.add(uniqueId);
         } else {
             inTeleport.remove(uniqueId);
-        }	
+        }
     }
 
     /**
      * Checks whether player with UUID uniqueId is in an official teleport or not
-     * @param uniqueId
+     * @param uniqueId - unique ID  - player's UUID
      * @return true if in teleport
      */
     public boolean isInTeleport(UUID uniqueId) {
@@ -816,7 +842,7 @@ public class PlayerCache {
 
     /**
      * Add death to player
-     * @param playerUUID
+     * @param playerUUID - the player's UUID  - player's UUID
      */
     public void addDeath(UUID playerUUID) {
         addPlayer(playerUUID);
@@ -825,8 +851,8 @@ public class PlayerCache {
 
     /**
      * Set death number for player
-     * @param playerUUID
-     * @param deaths
+     * @param playerUUID - the player's UUID - player's UUID
+     * @param deaths - number of deaths
      */
     public void setDeaths(UUID playerUUID, int deaths) {
         addPlayer(playerUUID);
@@ -835,8 +861,8 @@ public class PlayerCache {
 
     /**
      * Get number of times player has died in ASkyBlock worlds since counting began
-     * @param playerUUID
-     * @return
+     * @param playerUUID - the player's UUID - player's UUID
+     * @return number of deaths
      */
     public int getDeaths(UUID playerUUID) {
         addPlayer(playerUUID);
@@ -844,7 +870,7 @@ public class PlayerCache {
     }
 
     /**
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return List of challenges or levels done
      */
     public List<String> getChallengesDone(UUID playerUUID) {
@@ -853,7 +879,7 @@ public class PlayerCache {
     }
 
     /**
-     * @param playerUUID
+     * @param playerUUID - the player's UUID - player's UUID
      * @return List of challenges or levels not done
      */
     public List<String> getChallengesNotDone(UUID playerUUID) {
