@@ -16,16 +16,20 @@
  *******************************************************************************/
 package com.wasteofplastic.askyblock;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Map.Entry;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
+import org.bukkit.OfflinePlayer;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
 
@@ -39,15 +43,16 @@ import com.wasteofplastic.askyblock.util.Util;
  * @author tastybento
  */
 public class Players {
-    private ASkyBlock plugin;
+    private final ASkyBlock plugin;
     private YamlConfiguration playerInfo;
-    private HashMap<String, Boolean> challengeList;
-    private HashMap<String, Integer> challengeListTimes;
+    private Map<String, Boolean> challengeList;
+    private Map<String, Integer> challengeListTimes;
+    private Map<String, Long> challengeListTimestamp;
     private boolean hasIsland;
     private boolean inTeam;
     //private String homeLocation;
-    private HashMap<Integer, Location> homeLocations;
-    private int islandLevel;
+    private Map<Integer, Location> homeLocations;
+    private long islandLevel;
     private String islandLocation;
     private List<UUID> members;
     private String teamIslandLocation;
@@ -55,7 +60,7 @@ public class Players {
     private UUID uuid;
     private String playerName;
     private int resetsLeft;
-    private HashMap<Location, Date> kickedList;
+    private Map<Location, Date> kickedList;
     private List<UUID> banList;
     private String locale;
     private int startIslandRating;
@@ -63,30 +68,34 @@ public class Players {
     private int deaths;
 
     /**
-     * @param uuid
-     *            Constructor - initializes the state variables
-     * 
+     * @param aSkyBlock - plugin
+     * @param uuid - player's uuid
+     * @throws IOException - if uuid is null
      */
-    public Players(final ASkyBlock aSkyBlock, final UUID uuid) {
+    public Players(final ASkyBlock aSkyBlock, final UUID uuid) throws IOException {
         this.plugin = aSkyBlock;
         this.uuid = uuid;
-        this.members = new ArrayList<UUID>();
+        if (uuid == null) {
+            throw new IOException("UUID is null");
+        }
+        this.members = new ArrayList<>();
         this.hasIsland = false;
         this.islandLocation = null;
         //this.homeLocation = null;
-        this.homeLocations = new HashMap<Integer,Location>();
+        this.homeLocations = new HashMap<>();
         this.inTeam = false;
         this.teamLeader = null;
         this.teamIslandLocation = null;
-        this.challengeList = new HashMap<String, Boolean>();
-        this.challengeListTimes = new HashMap<String, Integer>();
+        this.challengeList = new HashMap<>();
+        this.challengeListTimes = new HashMap<>();
+        this.challengeListTimestamp = new HashMap<>();
         this.islandLevel = 0;
         this.playerName = "";
         this.resetsLeft = Settings.resetLimit;
-        this.kickedList = new HashMap<Location, Date>();
+        this.kickedList = new HashMap<>();
         this.locale = "";
         this.startIslandRating = 50;
-        this.banList = new ArrayList<UUID>();
+        this.banList = new ArrayList<>();
         this.useControlPanel = Settings.useControlPanel;
         load(uuid);
     }
@@ -102,16 +111,24 @@ public class Players {
         // Load in from YAML file
         this.playerName = playerInfo.getString("playerName", "");
         if (playerName.isEmpty()) {
-            try {
-                playerName = plugin.getServer().getOfflinePlayer(uuid).getName();
-            } catch (Exception e) {
-                plugin.getLogger().severe("Could not obtain a name for the player with UUID " + uuid.toString());
-                playerName = "";
+            Player player = plugin.getServer().getPlayer(uuid);
+            if (player != null && player.hasMetadata("NPC")) {
+                //plugin.getLogger().info("DEBUG: Entity is NPC");
+                playerName = player.getUniqueId().toString();
+            } else {
+                playerName = uuid.toString();
+                //plugin.getLogger().info("DEBUG: Entity is player");
+                try {
+                    OfflinePlayer offlinePlayer = plugin.getServer().getOfflinePlayer(uuid);
+                    if (offlinePlayer != null && offlinePlayer.getName() != null) {
+                        playerName = offlinePlayer.getName();
+                    }
+                } catch (Exception e) {
+                    plugin.getLogger().severe("Could not obtain a name for the player with UUID " + uuid.toString());
+                    playerName = uuid.toString();
+                }
             }
-            if (playerName == null) {
-                plugin.getLogger().severe("Could not obtain a name for the player with UUID " + uuid.toString());
-                playerName = "";
-            }
+
         }
         // Start island rating - how difficult the start island was. Default if 50/100
         this.startIslandRating = playerInfo.getInt("startIslandRating", 50);
@@ -122,7 +139,7 @@ public class Players {
         for (String uuidString : banListString) {
             try {
                 banList.add(UUID.fromString(uuidString));
-            } catch (Exception e) {}
+            } catch (Exception ignored) {}
         }
         // plugin.getLogger().info("Loading player..." + playerName);
         this.hasIsland = playerInfo.getBoolean("hasIsland", false);
@@ -168,11 +185,13 @@ public class Players {
             // If they are in the list, then use the value, otherwise use false
             challengeList.put(challenge.toLowerCase(), playerInfo.getBoolean("challenges.status." + challenge.toLowerCase().replace(".", "[dot]"), false));
             challengeListTimes.put(challenge.toLowerCase(), playerInfo.getInt("challenges.times." + challenge.toLowerCase().replace(".", "[dot]"), 0));
+            challengeListTimestamp.put(challenge.toLowerCase(), playerInfo.getLong("challenges.timestamp." + challenge.toLowerCase().replace(".", "[dot]"), 0));
         }
         for (String challenge : Settings.challengeLevels) {
             // If they are in the list, then use the value, otherwise use false
             challengeList.put(challenge.toLowerCase(), playerInfo.getBoolean("challenges.status." + challenge.toLowerCase().replace(".", "[dot]"), false));
             challengeListTimes.put(challenge.toLowerCase(), playerInfo.getInt("challenges.times." + challenge.toLowerCase().replace(".", "[dot]"), 0));
+            challengeListTimestamp.put(challenge.toLowerCase(), playerInfo.getLong("challenges.timestamp." + challenge.toLowerCase().replace(".", "[dot]"), 0));
         }
         // Load reset limit
         this.resetsLeft = playerInfo.getInt("resetsLeft", Settings.resetLimit);
@@ -215,13 +234,13 @@ public class Players {
     /**
      * Saves the player info to the file system
      */
-    public void save() {
+    public void save(boolean async) {
         //plugin.getLogger().info("Saving player..." + playerName);
         // Save the variables
         playerInfo.set("playerName", playerName);
         playerInfo.set("hasIsland", hasIsland);
         if (hasIsland && !banList.isEmpty()) {
-            List<String> banListString = new ArrayList<String>();
+            List<String> banListString = new ArrayList<>();
             for (UUID bannedUUID : banList) {
                 banListString.add(bannedUUID.toString());
             }
@@ -255,11 +274,16 @@ public class Players {
         // Save the challenges
         playerInfo.set("challenges",null);
         for (String challenge : challengeList.keySet()) {
-            //plugin.getLogger().info("DEBUG: " + challenge + "  --> " + challengeList.get(challenge));
-            playerInfo.set("challenges.status." + challenge.replace(".","[dot]"), challengeList.get(challenge));
+            if (!challenge.isEmpty())
+                playerInfo.set("challenges.status." + challenge.replace(".","[dot]"), challengeList.get(challenge));
         }
         for (String challenge : challengeListTimes.keySet()) {
-            playerInfo.set("challenges.times." + challenge.replace(".","[dot]"), challengeListTimes.get(challenge));
+            if (!challenge.isEmpty())
+                playerInfo.set("challenges.times." + challenge.replace(".","[dot]"), challengeListTimes.get(challenge));
+        }
+        for (String challenge : challengeListTimestamp.keySet()) {
+            if (!challenge.isEmpty())
+                playerInfo.set("challenges.timestamp." + challenge.replace(".","[dot]"), challengeListTimestamp.get(challenge));
         }
         // Check what the global limit is
         if (Settings.resetLimit < this.resetsLeft) {
@@ -287,10 +311,12 @@ public class Players {
         // Island info - to be used if the island.yml file is removed
         playerInfo.set("islandInfo",null);
         if (hasIsland) {
-            Island island = plugin.getGrid().getIsland(uuid);
-            if (island != null) {
-                playerInfo.set("islandInfo", island.save());
-            } 
+            if (plugin.getGrid() != null) {
+                Island island = plugin.getGrid().getIsland(uuid);
+                if (island != null) {
+                    playerInfo.set("islandInfo", island.save());
+                } 
+            }
         }
         // Control panel
         playerInfo.set("useControlPanel", useControlPanel);
@@ -298,7 +324,7 @@ public class Players {
         //playerInfo.set("coops", value);
 
         // Actually save the file
-        Util.saveYamlFile(playerInfo, "players/" + uuid.toString() + ".yml");
+        Util.saveYamlFile(playerInfo, "players/" + uuid.toString() + ".yml", async);
     }
 
     /**
@@ -327,11 +353,19 @@ public class Players {
      * @param challenge
      * @return true if the challenge is listed as complete, false if not
      */
-    public boolean checkChallenge(final String challenge) {
-        if (challengeList.containsKey(challenge.toLowerCase())) {
+    public boolean checkChallenge(String challenge) {
+        challenge = challenge.toLowerCase();
+        if (challengeList.containsKey(challenge)) {
+            // Check if the challenge has been globally reset
+            long timestamp = challengeListTimestamp.get(challenge);
+            if (plugin.getChallenges().isChallengeReset(challenge, timestamp)) {
+                this.resetChallenge(challenge);
+                return false;
+            }
+            // It has not been globally reset
             // plugin.getLogger().info("DEBUG: " + challenge + ":" +
             // challengeList.get(challenge.toLowerCase()).booleanValue() );
-            return challengeList.get(challenge.toLowerCase()).booleanValue();
+            return challengeList.get(challenge.toLowerCase());
         }
         return false;
     }
@@ -346,12 +380,12 @@ public class Players {
         if (challengeListTimes.containsKey(challenge.toLowerCase())) {
             // plugin.getLogger().info("DEBUG: check " + challenge + ":" +
             // challengeListTimes.get(challenge.toLowerCase()).intValue() );
-            return challengeListTimes.get(challenge.toLowerCase()).intValue();
+            return challengeListTimes.get(challenge.toLowerCase());
         }
         return 0;
     }
 
-    public HashMap<String, Boolean> getChallengeStatus() {
+    public Map<String, Boolean> getChallengeStatus() {
         return challengeList;
     }
 
@@ -363,7 +397,7 @@ public class Players {
      * @param challenge
      */
     public void completeChallenge(final String challenge) {
-        // plugin.getLogger().info("DEBUG: Complete challenge");
+        //plugin.getLogger().info("DEBUG: Complete challenge");
         challengeList.put(challenge.toLowerCase(), true);
         // Count how many times the challenge has been done
         int times = 0;
@@ -374,6 +408,8 @@ public class Players {
         challengeListTimes.put(challenge.toLowerCase(), times);
         // plugin.getLogger().info("DEBUG: complete " + challenge + ":" +
         // challengeListTimes.get(challenge.toLowerCase()).intValue() );
+        // Add timestamp
+        challengeListTimestamp.put(challenge, System.currentTimeMillis());
     }
 
     public boolean hasIsland() {
@@ -415,7 +451,7 @@ public class Players {
             }
         }
         if (members == null) {
-            members = new ArrayList<UUID>();
+            members = new ArrayList<>();
         }
         return inTeam;
     }
@@ -434,11 +470,7 @@ public class Players {
      * @return Location of this home or null if not available
      */
     public Location getHomeLocation(int number) {
-        if (homeLocations.containsKey(number)) {
-            return homeLocations.get(number);
-        } else {
-            return null;
-        }
+        return homeLocations.getOrDefault(number, null);
     }
 
     /**
@@ -457,7 +489,7 @@ public class Players {
      * @return The island level int. Note this function does not calculate the
      *         island level
      */
-    public int getIslandLevel() {
+    public long getIslandLevel() {
         return islandLevel;
     }
 
@@ -555,6 +587,7 @@ public class Players {
      * @param challenge
      */
     public void resetChallenge(final String challenge) {
+        //plugin.getLogger().info("DEBUG: reset challenge");
         challengeList.put(challenge, false);
         challengeListTimes.put(challenge, 0);
     }
@@ -590,12 +623,12 @@ public class Players {
     /**
      * Records the island's level. Does not calculate it
      * 
-     * @param i
+     * @param l
      */
-    public void setIslandLevel(final int i) {
-        islandLevel = i;
+    public void setIslandLevel(final long l) {
+        islandLevel = l;
         if (Settings.setTeamName) {
-            Scoreboards.getInstance().setLevel(uuid, i);
+            Scoreboards.getInstance().setLevel(uuid, l);
         }
     }
 
@@ -617,27 +650,42 @@ public class Players {
      * @param l
      *            - the Bukkit location of the team's island (converted to a
      *            String in this function)
+     * @return - true if successful
      */
-    public void setJoinTeam(final UUID leader, final Location l) {
+    public boolean setJoinTeam(final UUID leader, final Location l) {
         if(inTeam) {
-            Bukkit.getPluginManager().callEvent(new TeamLeaveEvent(uuid, teamLeader));
+            TeamLeaveEvent teamLeaveEvent = new TeamLeaveEvent(uuid, teamLeader);
+            Bukkit.getPluginManager().callEvent(teamLeaveEvent);
+            if (teamLeaveEvent.isCancelled()) {
+                return false;
+            }
         }
-
+        // Fire the event and give a chance for it to be cancelled.
+        TeamJoinEvent teamJoinEvent = new TeamJoinEvent(uuid, leader);
+        Bukkit.getPluginManager().callEvent(teamJoinEvent);
+        if (teamJoinEvent.isCancelled()) {
+            return false;
+        }
+        // Success
         inTeam = true;
         teamLeader = leader;
         teamIslandLocation = Util.getStringLocation(l);
 
-        Bukkit.getPluginManager().callEvent(new TeamJoinEvent(uuid, leader));
+        return true;
     }
 
     /**
      * Called when a player leaves a team Resets inTeam, teamLeader,
      * islandLevel, teamIslandLocation and members array
+     * @return true if successful, false if not
      */
-
-    public void setLeaveTeam() {
+    public boolean setLeaveTeam() {
         if(inTeam) {
-            Bukkit.getPluginManager().callEvent(new TeamLeaveEvent(uuid, teamLeader));
+            TeamLeaveEvent teamLeaveEvent = new TeamLeaveEvent(uuid, teamLeader);
+            Bukkit.getPluginManager().callEvent(teamLeaveEvent);
+            if (teamLeaveEvent.isCancelled()) {
+                return false;
+            }
         }
 
         inTeam = false;
@@ -645,6 +693,7 @@ public class Players {
         islandLevel = 0;
         teamIslandLocation = null;
         members = new ArrayList<>();
+        return true;
     }
 
     /**
@@ -658,17 +707,26 @@ public class Players {
     /**
      * @param leader
      *            a String name of the team leader
+     * @return true if successful
      */
-    public void setTeamLeader(final UUID leader) {
+    public boolean setTeamLeader(final UUID leader) {
         if(inTeam) {
             // Changing team leader changes the team identifier
-            Bukkit.getPluginManager().callEvent(new TeamLeaveEvent(uuid, teamLeader));
+            TeamLeaveEvent teamLeaveEvent = new TeamLeaveEvent(uuid, teamLeader);
+            Bukkit.getPluginManager().callEvent(teamLeaveEvent);
+            if (teamLeaveEvent.isCancelled()) {
+                return false;
+            }
         }
-
-        teamLeader = leader;
         if(leader != null) {
-            Bukkit.getPluginManager().callEvent(new TeamJoinEvent(uuid, leader));
+            TeamJoinEvent teamJoinEvent = new TeamJoinEvent(uuid, leader);
+            Bukkit.getPluginManager().callEvent(teamJoinEvent);
+            if (teamJoinEvent.isCancelled()) {
+                return false;
+            }
         }
+        teamLeader = leader;
+        return true;
     }
 
     /**
@@ -730,7 +788,7 @@ public class Players {
     /**
      * @return the challengeListTimes
      */
-    public HashMap<String, Integer> getChallengeCompleteTimes() {
+    public Map<String, Integer> getChallengeCompleteTimes() {
         return challengeListTimes;
     }
 
@@ -847,26 +905,16 @@ public class Players {
      * Used by the reset admin command
      */
     public List<String> getChallengesDone() {
-        List<String> result = new ArrayList<String>();
-        for (Entry<String, Boolean> en : challengeList.entrySet()) {
-           if (en.getValue()) {
-               result.add(en.getKey());
-           }
-        }
-        return result;
+        return challengeList.entrySet().stream().filter(Entry::getValue)
+            .map(Entry::getKey).collect(Collectors.toList());
     }
-    
+
     /**
      * @return a list of challenges this player has not completed
      * Used by the complete admin command
      */
     public List<String> getChallengesNotDone() {
-        List<String> result = new ArrayList<String>();
-        for (Entry<String, Boolean> en : challengeList.entrySet()) {
-           if (!en.getValue()) {
-               result.add(en.getKey());
-           }
-        }
-        return result;
+        return challengeList.entrySet().stream().filter(en -> !en.getValue())
+            .map(Entry::getKey).collect(Collectors.toList());
     }
 }
