@@ -20,6 +20,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
 
@@ -31,11 +32,11 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.material.MaterialData;
 import org.bukkit.potion.PotionEffectType;
 
-import com.wasteofplastic.askyblock.Island.Flags;
+import com.wasteofplastic.askyblock.Island.SettingsFlag;
 
 /**
  * Where all the settings are
- * 
+ *
  * @author tastybento
  */
 public class Settings {
@@ -56,7 +57,7 @@ public class Settings {
     public static final String CHALLENGECOMMAND = "aic";
     // Admin command
     public static final String ADMINCOMMAND = "acid";
-    */
+     */
     public final static GameType GAMETYPE = GameType.ASKYBLOCK;
     // Permission prefix
     public final static String PERMPREFIX = "askyblock.";
@@ -82,55 +83,36 @@ public class Settings {
     public static int monsterSpawnLimit;
     public static int animalSpawnLimit;
     public static int waterAnimalSpawnLimit;
-    // IslandGuard settings
-    public static HashMap<Flags, Boolean> defaultIslandSettings = new HashMap<Flags, Boolean>();
-    public static boolean allowAnvilUse;
-    public static boolean allowArmorStandUse;
-    public static boolean allowBeaconAccess;
-    public static boolean allowBedUse;
-    public static boolean allowBreakBlocks;
-    public static boolean allowBreeding;
-    public static boolean allowBrewing;
-    public static boolean allowBucketUse;
-    public static boolean allowChestAccess;
-    public static boolean allowCrafting;
-    public static boolean allowCropTrample;
-    public static boolean allowDoorUse;
-    public static boolean allowEnchanting;
-    public static boolean allowEnderPearls;
-    public static boolean allowFurnaceUse;
-    public static boolean allowGateUse;
-    public static boolean allowHorseInvAccess;
-    public static boolean allowHorseRiding;
-    public static boolean allowHurtMobs;
-    public static boolean allowLeashUse;
-    public static boolean allowLeverButtonUse;
-    public static boolean allowMusic;
-    public static boolean allowPlaceBlocks;
-    public static boolean allowPortalUse;
-    public static boolean allowPressurePlate;
-    public static boolean allowPvP;
-    public static boolean allowNetherPvP;
-    public static boolean allowRedStone;
-    public static boolean allowShearing;
-    public static boolean allowVillagerTrading;
-    public static boolean allowChorusFruit;
-    public static boolean enableJoinAndLeaveIslandMessages;
-    public static boolean allowMobSpawning;
+    /**
+     * Default world protection settings
+     */
+    public static final Map<SettingsFlag, Boolean> defaultWorldSettings = new HashMap<>();
+
+    /**
+     * Default island protection settings
+     */
+    public static final Map<SettingsFlag, Boolean> defaultIslandSettings = new HashMap<>();
+    /**
+     * Default spawn protection settings
+     */
+    public static final Map<SettingsFlag, Boolean> defaultSpawnSettings = new HashMap<>();
+    /**
+     * Visitors settings to show in the GUI
+     */
+    public static final Map<SettingsFlag, Boolean> visitorSettings = new HashMap<>();
+    // Flymode
+    public static int flyTimeOutside;
+
+    // Temporary Permissions
+    public static List<String> temporaryPermissions;
 
     // System settings
     public static boolean allowChestDamage;
     public static boolean allowCreeperDamage;
     public static boolean allowCreeperGriefing;
     public static boolean allowEndermanGriefing;
-    public static boolean allowFire;
-    public static boolean allowFireSpread;
-    public static boolean allowHurtMonsters;
-    public static boolean allowMonsterEggs;
     public static boolean allowPistonPush;
     public static boolean allowTNTDamage;
-    public static boolean allowVisitorItemDrop;
-    public static boolean allowVisitorItemPickup;
     public static boolean allowVisitorKeepInvOnDeath;
     public static boolean restrictWither;
 
@@ -138,16 +120,18 @@ public class Settings {
     public static int islandDistance;
     public static int islandXOffset;
     public static int islandZOffset;
-    public static int sea_level;
-    public static int island_protectionRange;
+    public static int seaHeight;
+    public static int islandProtectionRange;
     public static int abandonedIslandLevel;
     public static Double startingMoney;
     public static double netherSpawnRadius;
-    public static List<PotionEffectType> acidDamageType = new ArrayList<PotionEffectType>();
+    public static List<PotionEffectType> acidDamageType = new ArrayList<>();
     public static boolean resetMoney;
     public static boolean damageOps;
     public static boolean endermanDeathDrop;
-    
+    public static boolean onlyLeaderCanCoop;
+    public static boolean coopIsRequest;
+
     // Invincible visitor
     public static boolean invincibleVisitors;
     public static HashSet<DamageCause> visitorDamagePrevention;
@@ -155,52 +139,12 @@ public class Settings {
     // public static boolean ultraSafeBoats;
     public static boolean logInRemoveMobs;
     public static boolean islandRemoveMobs;
-    public static int island_level;
+    public static int islandHeight;
     public static boolean resetChallenges;
-    // Spawn fields
-    public static boolean allowSpawnDoorUse;
-    public static boolean allowSpawnLeverButtonUse;
-    public static boolean allowSpawnChestAccess;
-    public static boolean allowSpawnFurnaceUse;
-    public static boolean allowSpawnRedStone;
-    public static boolean allowSpawnMusic;
-    public static boolean allowSpawnCrafting;
-    public static boolean allowSpawnBrewing;
-    public static boolean allowSpawnGateUse;
-    public static boolean allowSpawnMobSpawn;
-    public static boolean allowSpawnNoAcidWater;
-    public static boolean allowSpawnEnchanting;
-    public static boolean allowSpawnAnvilUse;
-    public static boolean allowSpawnBeaconAccess;
-    public static boolean allowSpawnAnimalSpawn;
-    public static boolean allowSpawnAnimalKilling;
-    public static boolean allowSpawnMobKilling;
-    public static boolean allowSpawnMilking;
-    public static boolean allowSpawnBreakBlocks;
-    public static boolean allowSpawnPlaceBlocks;
-    public static boolean allowSpawnEggs;
-    public static boolean allowSpawnPVP;
-    public static boolean allowSpawnLavaCollection;
-    public static boolean allowSpawnWaterCollection;
-    public static boolean allowSpawnMonsterEggs;
-    public static boolean allowSpawnHorseInvAccess;
-    public static boolean allowSpawnHorseRiding;
-    public static boolean allowSpawnPressurePlate;
-    public static boolean allowSpawnVisitorItemDrop;
-    public static boolean allowSpawnVisitorItemPickup;
-    public static boolean allowSpawnArmorStandUse;
-    public static boolean allowSpawnBedUse;
-    public static boolean allowSpawnBreeding;
-    public static boolean allowSpawnCropTrample;
-    public static boolean allowSpawnEnderPearls;
-    public static boolean allowSpawnLeashUse;
-    public static boolean allowSpawnShearing;
-    public static boolean allowSpawnChorusFruit;
 
     // Levels
     public static HashMap<MaterialData, Integer> blockLimits;
     public static HashMap<MaterialData, Integer> blockValues;
-    public static boolean fastLevelCalc;
 
     // Challenge completion broadcast
     public static boolean broadcastMessages;
@@ -216,10 +160,10 @@ public class Settings {
     public static Biome defaultBiome;
 
     // Island reset commands
-    public static List<String> resetCommands = new ArrayList<String>();
+    public static List<String> resetCommands = new ArrayList<>();
     // Mob limits
     public static int breedingLimit;
-    
+
     // Console shows teamchat messages
     public static boolean logTeamChat;
 
@@ -236,7 +180,7 @@ public class Settings {
 
     // Use Economy
     public static boolean useEconomy;
-    
+
     // Use Minishop
     public static boolean useMinishop;
 
@@ -245,20 +189,24 @@ public class Settings {
 
     // Use physics when pasting schematic blocks
     public static boolean usePhysics;
-    
+
     // Use old display (chat instead of GUI) for Island top ten
     public static boolean displayIslandTopTenInChat;
 
     // Need a certain amount of island levels to create a warp sign
     public static int warpLevelsRestriction;
-    
+
     // Magic Cobble Generator
     public static boolean useMagicCobbleGen;
-    public static TreeMap<Integer,TreeMap<Double,Material>> magicCobbleGenChances;
-    
+    public static boolean magicCobbleGenOnlyAtSpawn;
+    public static TreeMap<Long,TreeMap<Double,Material>> magicCobbleGenChances;
+
     // Disable offline redstone
     public static boolean disableOfflineRedstone;
-    
+
+    // Fancy island level display
+    public static boolean fancyIslandLevelDisplay;
+
     // Falling blocked commands
     public static List<String> fallingCommandBlockList;
     public static List<String> leaveCommands;
@@ -275,7 +223,6 @@ public class Settings {
     public static boolean immediateTeleport;
     public static boolean makeIslandIfNone;
     public static boolean setTeamName;
-    public static int debug;
     public static boolean useSchematicPanel;
     public static boolean chooseIslandRandomly;
     public static double underWaterMultiplier;
@@ -290,7 +237,6 @@ public class Settings {
     public static boolean useWarpPanel;
     public static List<EntityType> mobWhiteList = new ArrayList<EntityType>();
     public static int villagerLimit;
-    public static int mobLimit;
     public static int hopperLimit;
     public static List<String> visitorCommandBlockList;
     public static boolean muteDeathMessages;
@@ -298,7 +244,6 @@ public class Settings {
     public static HashMap<String,Integer> limitedBlocks;
     public static long pvpRestartCooldown;
     public static long backupDuration;
-    public static boolean cleanUpBlocks;
     public static boolean acidBottle;
     public static boolean useOwnGenerator;
     public static List<String> freeLevels = new ArrayList<String>();
@@ -307,7 +252,7 @@ public class Settings {
     public static boolean recoverSuperFlat;
     protected static boolean levelLogging;
     public static boolean persistantCoops;
-    public static boolean allowSpawnCreeperPain;
+    //public static boolean allowSpawnCreeperPain;
     public static List<String> teamStartCommands;
     public static int minNameLength;
     public static int maxNameLength;
@@ -316,16 +261,31 @@ public class Settings {
     public static int maxDeaths;
     public static boolean islandResetDeathReset;
     public static boolean teamJoinDeathReset;
-    public static boolean allowAutoActivator;
+    public static List<String> allowedFakePlayers;
     public static boolean netherRoof;
-    public static boolean allowSpawnVillagerTrading;
+    //public static boolean allowSpawnVillagerTrading;
     public static String chatLevelPrefix;
     public static String chatChallengeLevelPrefix;
     public static String chatIslandPlayer;
     public static boolean allowObsidianScooping;
     public static boolean allowFireExtinguish;
-    public static boolean allowSpawnFireExtinguish;
+    //public static boolean allowSpawnFireExtinguish;
     public static boolean allowMobDamageToItemFrames;
     public static boolean kickedKeepInv;
     public static boolean hackSkeletonSpawners;
+    public static HashMap<EntityType, Integer> entityLimits;
+    public static long acidItemDestroyTime;
+    public static boolean helmetProtection;
+    public static boolean fullArmorProtection;
+    public static String defaultLanguage;
+    public static boolean showInActionBar;
+    public static boolean leaversLoseReset;
+    public static int maxPurge;
+    public static boolean allowTNTPushing;
+    public static boolean silenceCommandFeedback;
+    public static long inviteTimeout;
+    public static boolean warpHeads;
+    public static boolean saveEntities;
+    public static boolean coopsCanCreateWarps;
+    public static boolean deleteProtectedOnly;
 }
