@@ -24,6 +24,7 @@ import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.nio.file.Files;
+import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Scanner;
 import java.util.UUID;
@@ -38,17 +39,17 @@ import org.bukkit.scheduler.BukkitRunnable;
  *
  */
 public class TinyDB {
-    private ASkyBlock plugin;
-    private ConcurrentHashMap<String,UUID> treeMap;
+    private final ASkyBlock plugin;
+    private final Map<String,UUID> treeMap;
     private boolean dbReady;
     private boolean savingFlag;
     /**
      * Opens the database
-     * @param plugin
+     * @param plugin - ASkyBlock plugin object
      */
     public TinyDB(ASkyBlock plugin) {       
         this.plugin = plugin;
-        this.treeMap = new ConcurrentHashMap<String,UUID>();
+        this.treeMap = new ConcurrentHashMap<>();
         File database = new File(plugin.getDataFolder(), "name-uuid.txt");
         if (!database.exists()) {
             // Import from player files. Done async so may take a while
@@ -102,14 +103,12 @@ public class TinyDB {
                             // Read next lines
                             line = br.readLine();
                             uuid = br.readLine();
-                        } 
-                        br.close();
+                        }
                     }
-                    
-                }                
-                out.close();      
+
+                }
             }
-            
+
             // Move files around
             try {
                 Files.move(newDB.toPath(), oldDB.toPath(), java.nio.file.StandardCopyOption.REPLACE_EXISTING);
@@ -131,14 +130,12 @@ public class TinyDB {
             public void run() {
                 try {
                     // Load all the files from the player folder
-                    FilenameFilter ymlFilter = new FilenameFilter() {
-                        public boolean accept(File dir, String name) {
-                            String lowercaseName = name.toLowerCase();
-                            if (lowercaseName.endsWith(".yml")) {
-                                return true;
-                            } else {
-                                return false;
-                            }
+                    FilenameFilter ymlFilter = (dir, name) -> {
+                        String lowercaseName = name.toLowerCase();
+                        if (lowercaseName.endsWith(".yml")) {
+                            return true;
+                        } else {
+                            return false;
                         }
                     };
                     int count = 0;
@@ -153,18 +150,18 @@ public class TinyDB {
                             //System.out.println("DEBUG: UUID is " + uuid);
                             final UUID playerUUID = UUID.fromString(uuid);
                             // Get the player's name
-                            Scanner scanner = new Scanner(file);
-                            while (scanner.hasNextLine()) {
-                                final String lineFromFile = scanner.nextLine();
-                                if (lineFromFile.contains("playerName:")) { 
-                                    // Check against potentialUnowned list
-                                    String playerName = lineFromFile.substring(lineFromFile.indexOf(' ')).trim();
-                                    //System.out.println("DEBUG: Player name is " + playerName);
-                                    treeMap.put(playerName.toLowerCase(), playerUUID);
-                                    break;
+                            try (Scanner scanner = new Scanner(file)) {
+                                while (scanner.hasNextLine()) {
+                                    final String lineFromFile = scanner.nextLine();
+                                    if (lineFromFile.contains("playerName:")) { 
+                                        // Check against potentialUnowned list
+                                        String playerName = lineFromFile.substring(lineFromFile.indexOf(' ')).trim();
+                                        //System.out.println("DEBUG: Player name is " + playerName);
+                                        treeMap.put(playerName.toLowerCase(), playerUUID);
+                                        break;
+                                    }
                                 }
                             }
-                            scanner.close();
                         } catch (Exception ex) {
                             System.err.println("[ASkyBlock/AcidIsland]: Problem reading " + file.getName() + " skipping...");
                             //ex.printStackTrace();
@@ -174,7 +171,7 @@ public class TinyDB {
                     saveDB();
                     // Save memory
                     treeMap.clear();
-                    System.out.println("[ASkyBlock]: Complete. Processed " + count + " names to database");
+                    System.out.println("Complete. Processed " + count + " names to database");
                     // Set flag
                     dbReady = true;
                 } catch (Exception e) {
@@ -193,9 +190,12 @@ public class TinyDB {
     /**
      * Saves the player name to the database. Case insensitive!
      * @param playerName
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      */
     public void savePlayerName(String playerName, UUID playerUUID) {
+        if (playerName == null) {
+            return;
+        }
         treeMap.put(playerName.toLowerCase(), playerUUID);
         // This will be saved when everything shuts down
     }
@@ -206,6 +206,9 @@ public class TinyDB {
      * @return UUID of player, or null if unknown
      */
     public UUID getPlayerUUID(String playerName) {
+        if (playerName == null) {
+            return null;
+        }
         // Try cache
         if (treeMap.containsKey(playerName.toLowerCase())) {
             //plugin.getLogger().info("DEBUG: found in UUID cache");
@@ -233,13 +236,16 @@ public class TinyDB {
         } 
         return null;
     }
-    
+
     /**
      * Gets players name from tiny database
-     * @param playerUuid
+     * @param playerUuid - the player's UUID
      * @return Name or empty string if unknown
      */
     public String getPlayerName(UUID playerUuid) {
+        if (playerUuid == null) {
+            return "";
+        }
         // Names and UUID's are stored in line pairs.
         try(BufferedReader br = new BufferedReader(new FileReader(new File(plugin.getDataFolder(), "name-uuid.txt")))) {
             String line = br.readLine();
