@@ -17,9 +17,13 @@
 
 package com.wasteofplastic.askyblock.listeners;
 
+import java.math.BigInteger;
+import java.text.DecimalFormat;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
+import java.util.TreeMap;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -33,6 +37,7 @@ import org.bukkit.event.player.AsyncPlayerChatEvent;
 
 import com.wasteofplastic.askyblock.ASkyBlock;
 import com.wasteofplastic.askyblock.Settings;
+import com.wasteofplastic.askyblock.util.Util;
 
 /**
  * This class is to catch chats and do two things: (1) substitute in the island level to the chat string
@@ -49,21 +54,21 @@ import com.wasteofplastic.askyblock.Settings;
  */
 public class ChatListener implements Listener {
 
-    private ASkyBlock plugin;
-    private ConcurrentHashMap<UUID,Boolean> teamChatUsers;
-    private ConcurrentHashMap<UUID,String> playerLevels;
-    private ConcurrentHashMap<UUID,String> playerChallengeLevels;
+    private final ASkyBlock plugin;
+    private final Map<UUID,Boolean> teamChatUsers;
+    private final Map<UUID,String> playerLevels;
+    private final Map<UUID, String> playerChallengeLevels;
     // List of which admins are spying or not on team chat
-    private Set<UUID> spies;
+    private final Set<UUID> spies;
     private static final boolean DEBUG = false;
 
     /**
-     * @param plugin
+     * @param plugin - ASkyBlock plugin object
      */
     public ChatListener(ASkyBlock plugin) {
-        this.teamChatUsers = new ConcurrentHashMap<UUID,Boolean>();
-        this.playerLevels = new ConcurrentHashMap<UUID,String>();
-        this.playerChallengeLevels = new ConcurrentHashMap<UUID,String>();
+        this.teamChatUsers = new ConcurrentHashMap<>();
+        this.playerLevels = new ConcurrentHashMap<>();
+        this.playerChallengeLevels = new ConcurrentHashMap<>();
         this.plugin = plugin;
         // Add all online player Levels
         for (Player player : plugin.getServer().getOnlinePlayers()) {
@@ -71,11 +76,25 @@ public class ChatListener implements Listener {
             playerChallengeLevels.put(player.getUniqueId(), plugin.getChallenges().getChallengeLevel(player));
         }
         // Initialize spies
-        spies = new HashSet<UUID>();
+        spies = new HashSet<>();
     }
 
+    private static final BigInteger THOUSAND = BigInteger.valueOf(1000);
+    /**
+     * Provides an easy way to "fancy" the island level in chat
+     * @since 3.0.8.3
+     */
+    private static final TreeMap<BigInteger, String> LEVELS;
+    static {
+        LEVELS = new TreeMap<>();
 
-    @EventHandler(priority = EventPriority.LOW, ignoreCancelled = true)
+        LEVELS.put(THOUSAND, "k");
+        LEVELS.put(THOUSAND.pow(2), "M");
+        LEVELS.put(THOUSAND.pow(3), "G");
+        LEVELS.put(THOUSAND.pow(4), "T");
+    }
+
+    @EventHandler(priority = EventPriority.LOWEST, ignoreCancelled = true)
     public void onChat(final AsyncPlayerChatEvent event) {
         if (DEBUG)
             plugin.getLogger().info("DEBUG: " + event.getEventName());
@@ -83,16 +102,47 @@ public class ChatListener implements Listener {
         String level = "";
         if (playerLevels.containsKey(event.getPlayer().getUniqueId())) {
             level = playerLevels.get(event.getPlayer().getUniqueId());
+            if(Settings.fancyIslandLevelDisplay) {
+                BigInteger levelValue = BigInteger.valueOf(Long.valueOf(level));
+
+                Map.Entry<BigInteger, String> stage = LEVELS.floorEntry(levelValue);
+
+                if (stage != null) { // level > 1000
+                    // 1 052 -> 1.0k
+                    // 1 527 314 -> 1.5M
+                    // 3 874 130 021 -> 3.8G
+                    // 4 002 317 889 -> 4.0T
+                    level = new DecimalFormat("#.#").format(levelValue.divide(stage.getKey().divide(THOUSAND)).doubleValue()/1000.0) + stage.getValue();
+                }
+            }
         }
-        String format = event.getFormat().replace(Settings.chatLevelPrefix, level);
-        level = "";
-        if (playerChallengeLevels.containsKey(event.getPlayer().getUniqueId())) {
-            level = playerChallengeLevels.get(event.getPlayer().getUniqueId());
+        if (DEBUG) {
+            plugin.getLogger().info("DEBUG: player level = " + level);
+            plugin.getLogger().info("DEBUG: getFormat = " + event.getFormat());
+            plugin.getLogger().info("DEBUG: getMessage = " + event.getMessage());
         }
-        format = format.replace(Settings.chatChallengeLevelPrefix, level);
+        String format = event.getFormat();
+        if (!Settings.chatLevelPrefix.isEmpty()) {
+            format = format.replace(Settings.chatLevelPrefix, level);
+            if (DEBUG)
+                plugin.getLogger().info("DEBUG: format (island level substitute) = " + format);
+        }
+        if (!Settings.chatChallengeLevelPrefix.isEmpty()) {
+            level = "";
+            if (playerChallengeLevels.containsKey(event.getPlayer().getUniqueId())) {
+                level = playerChallengeLevels.get(event.getPlayer().getUniqueId());
+            }
+            format = format.replace(Settings.chatChallengeLevelPrefix, level);
+            if (DEBUG)
+                plugin.getLogger().info("DEBUG: format (challenge level sub) = " + format);           
+        }
         event.setFormat(format);
+        if (DEBUG)
+            plugin.getLogger().info("DEBUG: format set");
         // Team chat
         if (Settings.teamChat && teamChatUsers.containsKey(event.getPlayer().getUniqueId())) {
+            if (DEBUG)
+                plugin.getLogger().info("DEBUG: team chat");
             // Cancel the event
             event.setCancelled(true);
             // Queue the sync task because you cannot use HashMaps asynchronously. Delaying to the next tick
@@ -116,11 +166,15 @@ public class ChatListener implements Listener {
             List<UUID> teamMembers = plugin.getPlayers().getMembers(player.getUniqueId());
             // Tell only the team members if they are online
             boolean onLine = false;
-            message = plugin.myLocale(playerUUID).teamChatPrefix.replace(Settings.chatIslandPlayer,player.getDisplayName()) + message;
+            if (Settings.chatIslandPlayer.isEmpty()) {
+                message = plugin.myLocale(playerUUID).teamChatPrefix + message;
+            } else {
+                message = plugin.myLocale(playerUUID).teamChatPrefix.replace(Settings.chatIslandPlayer,player.getDisplayName()) + message;
+            }
             for (UUID teamMember : teamMembers) {
                 Player teamPlayer = plugin.getServer().getPlayer(teamMember);
                 if (teamPlayer != null) {
-                    teamPlayer.sendMessage(message);
+                    Util.sendMessage(teamPlayer, message);
                     if (!teamMember.equals(playerUUID)) {
                         onLine = true;
                     }
@@ -130,20 +184,20 @@ public class ChatListener implements Listener {
             if (onLine) {
                 for (Player onlinePlayer: plugin.getServer().getOnlinePlayers()) {
                     if (spies.contains(onlinePlayer.getUniqueId()) && onlinePlayer.hasPermission(Settings.PERMPREFIX + "mod.spy")) {
-                        onlinePlayer.sendMessage(ChatColor.RED + "[TCSpy] " + ChatColor.WHITE + message);
+                        Util.sendMessage(onlinePlayer, ChatColor.RED + "[TCSpy] " + ChatColor.WHITE + message);
                     }
                 }
                 //Log teamchat
                 if(Settings.logTeamChat) plugin.getLogger().info(ChatColor.stripColor(message));
             }
             if (!onLine) {
-                player.sendMessage(ChatColor.RED + plugin.myLocale(playerUUID).teamChatNoTeamAround);
-                player.sendMessage(ChatColor.RED + plugin.myLocale(playerUUID).teamChatStatusOff);
+                Util.sendMessage(player, ChatColor.RED + plugin.myLocale(playerUUID).teamChatNoTeamAround);
+                Util.sendMessage(player, ChatColor.RED + plugin.myLocale(playerUUID).teamChatStatusOff);
                 teamChatUsers.remove(playerUUID);
             }
         } else {
-            player.sendMessage(ChatColor.RED + plugin.myLocale(playerUUID).teamChatNoTeamAround);
-            player.sendMessage(ChatColor.RED + plugin.myLocale(playerUUID).teamChatStatusOff);
+            Util.sendMessage(player, ChatColor.RED + plugin.myLocale(playerUUID).teamChatNoTeamAround);
+            Util.sendMessage(player, ChatColor.RED + plugin.myLocale(playerUUID).teamChatStatusOff);
             // Not in a team any more so delete
             teamChatUsers.remove(playerUUID);
         }
@@ -151,7 +205,7 @@ public class ChatListener implements Listener {
 
     /**
      * Adds player to team chat
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      */
     public void setPlayer(UUID playerUUID) {
         this.teamChatUsers.put(playerUUID,true);
@@ -159,7 +213,7 @@ public class ChatListener implements Listener {
 
     /**
      * Removes player from team chat
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      */
     public void unSetPlayer(UUID playerUUID) {
         this.teamChatUsers.remove(playerUUID);
@@ -167,7 +221,7 @@ public class ChatListener implements Listener {
 
     /**
      * Whether the player has team chat on or not
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @return true if team chat is on
      */
     public boolean isTeamChat(UUID playerUUID) {
@@ -176,12 +230,12 @@ public class ChatListener implements Listener {
 
     /**
      * Store the player's level for use in their chat tag
-     * @param playerUUID
-     * @param level
+     * @param playerUUID - the player's UUID
+     * @param l
      */
-    public void setPlayerLevel(UUID playerUUID, int level) {
+    public void setPlayerLevel(UUID playerUUID, long l) {
         //plugin.getLogger().info("DEBUG: putting " + playerUUID.toString() + " Level " + level);
-        playerLevels.put(playerUUID, String.valueOf(level));
+        playerLevels.put(playerUUID, String.valueOf(l));
     }
 
     /**
@@ -192,10 +246,10 @@ public class ChatListener implements Listener {
         //plugin.getLogger().info("DEBUG: setting player's challenge level to " + plugin.getChallenges().getChallengeLevel(player));
         playerChallengeLevels.put(player.getUniqueId(), plugin.getChallenges().getChallengeLevel(player));
     }
-    
+
     /**
      * Return the player's level for use in chat - async safe
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @return Player's level as string
      */
     public String getPlayerLevel(UUID playerUUID) {
@@ -204,16 +258,18 @@ public class ChatListener implements Listener {
 
     /**
      * Return the player's challenge level for use in chat - async safe
-     * @param playerUUID
-     * @return challenge level as string
+     * @param playerUUID - the player's UUID
+     * @return challenge level as string or empty string none
      */
     public String getPlayerChallengeLevel(UUID playerUUID) {
-        return playerChallengeLevels.get(playerUUID);
+        if (playerChallengeLevels.containsKey(playerUUID))
+            return playerChallengeLevels.get(playerUUID);
+        return "";
     }
 
     /**
      * Toggles team chat spy. Spy must also have the spy permission to see chats
-     * @param playerUUID
+     * @param playerUUID - the player's UUID
      * @return true if toggled on, false if toggled off
      */
     public boolean toggleSpy(UUID playerUUID) {
