@@ -17,39 +17,47 @@
 package com.wasteofplastic.askyblock.listeners;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
 import java.util.UUID;
 
 import org.bukkit.ChatColor;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
 import org.bukkit.Material;
+import org.bukkit.World.Environment;
 import org.bukkit.block.BlockFace;
 import org.bukkit.entity.Player;
+import org.bukkit.entity.Projectile;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
+import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.EntityDamageEvent.DamageCause;
 import org.bukkit.event.entity.FoodLevelChangeEvent;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.PlayerCommandPreprocessEvent;
 import org.bukkit.event.player.PlayerDropItemEvent;
+import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerMoveEvent;
-import org.bukkit.event.player.PlayerPickupItemEvent;
+import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.player.PlayerRespawnEvent;
 import org.bukkit.event.player.PlayerTeleportEvent;
 import org.bukkit.event.player.PlayerTeleportEvent.TeleportCause;
-import org.bukkit.util.Vector;
 
 import com.wasteofplastic.askyblock.ASkyBlock;
 import com.wasteofplastic.askyblock.InventorySave;
 import com.wasteofplastic.askyblock.Island;
-import com.wasteofplastic.askyblock.Island.Flags;
+import com.wasteofplastic.askyblock.Island.SettingsFlag;
 import com.wasteofplastic.askyblock.Settings;
 import com.wasteofplastic.askyblock.events.IslandEnterEvent;
 import com.wasteofplastic.askyblock.events.IslandExitEvent;
+import com.wasteofplastic.askyblock.util.Util;
 import com.wasteofplastic.askyblock.util.VaultHelper;
 
 /**
@@ -60,17 +68,77 @@ public class PlayerEvents implements Listener {
     private final ASkyBlock plugin;
     private static final boolean DEBUG = false;
     // A set of falling players
-    private static HashSet<UUID> fallingPlayers = new HashSet<UUID>();
+    private static Set<UUID> fallingPlayers = new HashSet<>();
+    private static Map<UUID, List<String>> temporaryPerms = new HashMap<>();
     private List<UUID> respawn;
 
     public PlayerEvents(final ASkyBlock plugin) {
         this.plugin = plugin;
-        respawn = new ArrayList<UUID>();
+        respawn = new ArrayList<>();
+    }
+
+    /**
+     * Removes all temp perms. Used when the server is shutting down and players are still online.
+     */
+    public void removeAllTempPerms() {
+        for (Entry<UUID, List<String>> en : temporaryPerms.entrySet()) {
+            Player player = plugin.getServer().getPlayer(en.getKey());
+            if (player != null) {
+                for (String perm: en.getValue()) {
+                    VaultHelper.removePerm(player, perm, ASkyBlock.getIslandWorld());
+                    if (Settings.createNether && Settings.newNether && ASkyBlock.getNetherWorld() != null) {
+                        VaultHelper.removePerm(player, perm, ASkyBlock.getNetherWorld());
+                    }
+                }
+            } else {
+                String playerName = plugin.getTinyDB().getPlayerName(en.getKey());
+                plugin.getLogger().warning(playerName + ", (" + en.getKey().toString() + ")");
+                plugin.getLogger().warning("Had these temporary perms that could not be removed because they are offline:");
+                for (String perm: en.getValue()) {
+                    plugin.getLogger().warning(perm);
+                }
+                plugin.getLogger().warning("These should be removed manually.");
+            }
+        }
+    }
+
+    /**
+     * Gives temporary perms to players who are online when the server is reloaded or the plugin reloaded.
+     */
+    public void giveAllTempPerms() {
+        if (plugin.getGrid() == null) {
+            return;
+        }
+        if (DEBUG)
+            plugin.getLogger().info("DEBUG: Giving all temp perms");
+        for (Player player : plugin.getServer().getOnlinePlayers()) {
+            if(player != null && !player.hasMetadata("NPC") && plugin.getGrid().playerIsOnIsland(player)){
+                if(VaultHelper.checkPerm(player, Settings.PERMPREFIX + "islandfly")){
+                    if (DEBUG)
+                        plugin.getLogger().info("DEBUG: Fly enable");
+                    player.setAllowFlight(true);
+                    player.setFlying(true);
+                }
+
+                for(String perm : Settings.temporaryPermissions){
+                    if(!VaultHelper.checkPerm(player, perm)){
+                        VaultHelper.addPerm(player, perm, ASkyBlock.getIslandWorld());
+                        if (Settings.createNether && Settings.newNether && ASkyBlock.getNetherWorld() != null) {
+                            VaultHelper.addPerm(player, perm, ASkyBlock.getNetherWorld());
+                        }
+                        List<String> perms = new ArrayList<String>();
+                        if(temporaryPerms.containsKey(player.getUniqueId())) perms = temporaryPerms.get(player.getUniqueId());
+                        perms.add(perm);
+                        temporaryPerms.put(player.getUniqueId(), perms);
+                    }
+                }
+            }
+        }
     }
 
     /**
      * Prevents changing of hunger while having a special permission and being on your island
-     * @param e
+     * @param e - event
      */
     @EventHandler(priority = EventPriority.LOW, ignoreCancelled = true)
     public void onHungerChange(final FoodLevelChangeEvent e) {
@@ -90,10 +158,217 @@ public class PlayerEvents implements Listener {
     }
 
     /**
-     * Places player back on their island if the setting is true
-     * @param e
+     * Gives temporary perms
+     * Gives flymode if player has a specific permission and is on his island
+     * @param e - event
      */
-    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = false)
+    @EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
+    public void onPlayerEnterOnIsland(final IslandEnterEvent e){
+        Player player = plugin.getServer().getPlayer(e.getPlayer());
+        if (player != null && !player.hasMetadata("NPC")) {
+            if (DEBUG) {
+                plugin.getLogger().info("DEBUG: player entered island");
+                plugin.getLogger().info("DEBUG: island center is " + e.getIslandLocation());
+                if (e.getIslandOwner() != null && plugin.getPlayers().isAKnownPlayer(e.getIslandOwner())) {
+                    plugin.getLogger().info("DEBUG: island owner is " + plugin.getPlayers().getName(e.getIslandOwner()));
+                } else {
+                    plugin.getLogger().info("DEBUG: island is unowned or owner unknown");
+                }
+            }
+            processPerms(player, e.getIsland());
+        }
+    }
+
+    /**
+     * Adds perms or fly for player on island
+     * @param player
+     * @param island
+     */
+    private void processPerms(final Player player, final Island island) {
+        if (island != null && island.getMembers().contains(player.getUniqueId())) {
+            if (DEBUG)
+                plugin.getLogger().info("DEBUG: player on island " + player.getName());
+            if(VaultHelper.checkPerm(player, Settings.PERMPREFIX + "islandfly", island.getCenter().getWorld())) {
+                if (DEBUG)
+                    plugin.getLogger().info("DEBUG: player has fly");
+                plugin.getServer().getScheduler().runTask(plugin, () -> {
+                    player.setAllowFlight(true);
+                    player.setFlying(true);
+
+                });
+
+            }
+            if (DEBUG)
+                plugin.getLogger().info("DEBUG: adding temp perms");
+            for(String perm : Settings.temporaryPermissions) {
+                if(!VaultHelper.checkPerm(player, perm, island.getCenter().getWorld())){
+                    VaultHelper.addPerm(player, perm, ASkyBlock.getIslandWorld());
+                    if (Settings.createNether && Settings.newNether && ASkyBlock.getNetherWorld() != null) {
+                        VaultHelper.addPerm(player, perm, ASkyBlock.getNetherWorld());
+                    }
+                    List<String> perms = new ArrayList<>();
+                    if(temporaryPerms.containsKey(player.getUniqueId())) perms = temporaryPerms.get(player.getUniqueId());
+                    perms.add(perm);
+                    temporaryPerms.put(player.getUniqueId(), perms);
+                    if (DEBUG)
+                        plugin.getLogger().info("DEBUG: adding perm " + perm);
+                } else {
+                    if (DEBUG)
+                        plugin.getLogger().info("DEBUG: player already has perm " + perm);
+                }
+            }
+        }
+    }
+
+    /**
+     * Handle player joining
+     * @param event
+     */
+    @EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
+    public void onPlayerJoin(final PlayerJoinEvent event) {
+        final Island island = plugin.getGrid().getProtectedIslandAt(event.getPlayer().getLocation());
+        if (island != null) {
+            processPerms(event.getPlayer(), island);
+            // Fire entry event
+            final IslandEnterEvent e = new IslandEnterEvent(event.getPlayer().getUniqueId(), island, event.getPlayer().getLocation());
+            plugin.getServer().getPluginManager().callEvent(e);
+        }
+    }
+
+    /**
+     * Revoke temporary perms
+     * Removes flymode with a delay if player leave his island.
+     * @param e - event
+     */
+    @EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
+    public void onPlayerLeaveIsland(final IslandExitEvent e) {
+        if (DEBUG)
+            plugin.getLogger().info("DEBUG: island exit event");
+        final Player player = plugin.getServer().getPlayer(e.getPlayer());
+        if (player != null && !player.hasMetadata("NPC")) {
+            if (DEBUG) {
+                plugin.getLogger().info("DEBUG: player left island. e.getLocation = " + e.getLocation());
+                plugin.getLogger().info("DEBUG: player location = " + player.getLocation());
+            }
+            removeTempPerms(player, e.getIsland(), plugin.getGrid().getIslandAt(e.getLocation()));
+        }
+    }
+
+    /**
+     * Removes perms for a player who was on one island and is now elsewhere
+     * If fromIsland and toIsland are the same, then the player is just out of their protection zone
+     * and if timing is allowed, they will keep their fly capability
+     * @param player
+     * @param fromIsland
+     * @param toIsland
+     */
+    public void removeTempPerms(final Player player, Island fromIsland, Island toIsland) {
+        if (DEBUG)
+            plugin.getLogger().info("DEBUG: Removing temp perms");
+        if (player == null || player.hasMetadata("NPC")) {
+            return;
+        }
+        // Check if the player has left the island completely
+        if(VaultHelper.checkPerm(player, Settings.PERMPREFIX + "islandfly")) {
+            // If the player has teleported to another world or island
+            if (fromIsland.equals(toIsland)) {
+                if (player.isFlying() && player.getGameMode().equals(GameMode.SURVIVAL)) {
+                    if (DEBUG)
+                        plugin.getLogger().info("DEBUG: player is flying timer is " + Settings.flyTimeOutside + "s");
+                    if (Settings.flyTimeOutside == 0) {
+                        player.setAllowFlight(false);
+                        player.setFlying(false);
+                        if (DEBUG)
+                            plugin.getLogger().info("DEBUG: removed fly");
+                    } else {
+                        plugin.getServer().getScheduler().scheduleSyncDelayedTask(plugin, new Runnable() {
+
+                            @Override
+                            public void run() {
+                                if(!plugin.getGrid().playerIsOnIsland(player) && player.isFlying()){
+                                    // Check they didn't enable creative
+                                    if (player.getGameMode().equals(GameMode.SURVIVAL)) {
+                                        player.setAllowFlight(false);
+                                        player.setFlying(false);
+                                        if (DEBUG)
+                                            plugin.getLogger().info("DEBUG: removed fly");
+                                    }
+                                }
+
+                            }
+                        }, 20L*Settings.flyTimeOutside);
+                    }
+                }
+            } else {
+                if (DEBUG)
+                    plugin.getLogger().info("DEBUG: Removing flight immediately");
+                if (player.getGameMode().equals(GameMode.SURVIVAL)) {
+                    // Remove fly immediately
+                    player.setAllowFlight(false);
+                    player.setFlying(false);
+                    if (DEBUG)
+                        plugin.getLogger().info("DEBUG: removed fly");
+                }
+            }
+        }
+
+
+        for(String perm : Settings.temporaryPermissions){
+            if(temporaryPerms.containsKey(player.getUniqueId()) && VaultHelper.checkPerm(player, perm)){
+                VaultHelper.removePerm(player, perm, ASkyBlock.getIslandWorld());
+                if (Settings.createNether && Settings.newNether && ASkyBlock.getNetherWorld() != null) {
+                    VaultHelper.removePerm(player, perm, ASkyBlock.getNetherWorld());
+                }
+                if (DEBUG)
+                    plugin.getLogger().info("DEBUG: removed temp perm " + perm);
+                List<String> perms = temporaryPerms.get(player.getUniqueId());
+                perms.remove(perm);
+                if(perms.isEmpty()) temporaryPerms.remove(player.getUniqueId());
+                else temporaryPerms.put(player.getUniqueId(), perms);
+            }
+        }
+    }
+
+
+
+    /**
+     * Removes temporary perms when the player log out
+     * @param event
+     */
+    @EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
+    public void onPlayerLeave(PlayerQuitEvent event){
+        if (DEBUG)
+            plugin.getLogger().info("DEBUG: Removing fly and all temp perms");
+        Player player = event.getPlayer();
+
+        if(temporaryPerms.containsKey(player.getUniqueId())){
+            for(String perm : temporaryPerms.get(player.getUniqueId())){
+                VaultHelper.removePerm(player, perm, ASkyBlock.getIslandWorld());
+                if (Settings.createNether && Settings.newNether && ASkyBlock.getNetherWorld() != null) {
+                    VaultHelper.removePerm(player, perm, ASkyBlock.getNetherWorld());
+                }
+            }
+            temporaryPerms.remove(player.getUniqueId());
+        }
+        if(VaultHelper.checkPerm(player, Settings.PERMPREFIX + "islandfly")) {
+            if (player.getGameMode().equals(GameMode.SURVIVAL)) {
+                player.setAllowFlight(false);
+                player.setFlying(false);
+            }
+        }
+        final Island island = plugin.getGrid().getProtectedIslandAt(event.getPlayer().getLocation());
+        if (island != null) {
+            // Fire exit event
+            final IslandExitEvent e = new IslandExitEvent(event.getPlayer().getUniqueId(), island, event.getPlayer().getLocation());
+            plugin.getServer().getPluginManager().callEvent(e);
+        }
+    }
+
+    /**
+     * Places player back on their island if the setting is true
+     * @param e - event
+     */
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
     public void onPlayerRespawn(final PlayerRespawnEvent e) {
         if (DEBUG) {
             plugin.getLogger().info(e.getEventName());
@@ -107,6 +382,12 @@ public class PlayerEvents implements Listener {
             if (respawnLocation != null) {
                 //plugin.getLogger().info("DEBUG: Setting respawn location to " + respawnLocation);
                 e.setRespawnLocation(respawnLocation);
+                // Get island
+                Island island = plugin.getGrid().getIslandAt(respawnLocation);
+                if (island != null) {
+                    // Run perms etc.
+                    processPerms(e.getPlayer(), island);
+                }
             }
         }
     }
@@ -114,9 +395,9 @@ public class PlayerEvents implements Listener {
     /**
      * Registers death of player.
      * Places the player on the island respawn list if set
-     * @param e
+     * @param e - event
      */
-    @EventHandler(priority = EventPriority.LOW, ignoreCancelled = false)
+    @EventHandler(priority = EventPriority.LOW, ignoreCancelled = true)
     public void onPlayerDeath(final PlayerDeathEvent e) {
         if (DEBUG) {
             plugin.getLogger().info(e.getEventName());
@@ -141,14 +422,14 @@ public class PlayerEvents implements Listener {
                 }
             }
         }
-    } 
+    }
 
     /*
      * Prevent dropping items if player dies on another island
      * This option helps reduce the down side of dying due to traps, etc.
      * Also handles muting of death messages
      */
-    @EventHandler(priority = EventPriority.LOW, ignoreCancelled = false)
+    @EventHandler(priority = EventPriority.LOW, ignoreCancelled = true)
     public void onVistorDeath(final PlayerDeathEvent e) {
         if (DEBUG) {
             plugin.getLogger().info(e.getEventName());
@@ -160,18 +441,19 @@ public class PlayerEvents implements Listener {
         if (Settings.muteDeathMessages) {
             e.setDeathMessage(null);
         }
-        // If the player is on their island then they die and lose everything -
-        // sorry :-(
-        if (plugin.getGrid().playerIsOnIsland(e.getEntity())) {
-            return;
-        }
         // If visitors will keep items and their level on death
         // This will override any global settings
         if (Settings.allowVisitorKeepInvOnDeath) {
-            InventorySave.getInstance().savePlayerInventory(e.getEntity());
-            e.getDrops().clear();
-            e.setKeepLevel(true);
-            e.setDroppedExp(0);
+            // If the player is not a visitor then they die and lose everything -
+            // sorry :-(
+            Island island = plugin.getGrid().getProtectedIslandAt(e.getEntity().getLocation());
+            if (island != null && !island.getMembers().contains(e.getEntity().getUniqueId())) {
+                // They are a visitor
+                InventorySave.getInstance().savePlayerInventory(e.getEntity());
+                e.getDrops().clear();
+                e.setKeepLevel(true);
+                e.setDroppedExp(0);
+            }
         }
     }
 
@@ -186,29 +468,6 @@ public class PlayerEvents implements Listener {
             InventorySave.getInstance().clearSavedInventory(e.getPlayer());
         }
     }
-    /*
-     * Prevent item pickup by visitors
-     */
-    @EventHandler(priority = EventPriority.LOW, ignoreCancelled = true)
-    public void onVisitorPickup(final PlayerPickupItemEvent e) {
-        if (DEBUG) {
-            plugin.getLogger().info(e.getEventName());
-        }
-        if (!IslandGuard.inWorld(e.getPlayer())) {
-            return;
-        }
-        if (plugin.getGrid().isAtSpawn(e.getItem().getLocation())) {
-            if (Settings.allowSpawnVisitorItemPickup || e.getPlayer().isOp() || VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypassprotect")
-                    || plugin.getGrid().locationIsOnIsland(e.getPlayer(), e.getItem().getLocation())) {
-                return;
-            }
-        }
-        if (Settings.allowVisitorItemPickup || e.getPlayer().isOp() || VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypassprotect")
-                || plugin.getGrid().locationIsOnIsland(e.getPlayer(), e.getItem().getLocation())) {
-            return;
-        }
-        e.setCancelled(true);
-    }
 
     /*
      * Prevent item drop by visitors
@@ -221,17 +480,13 @@ public class PlayerEvents implements Listener {
         if (!IslandGuard.inWorld(e.getPlayer())) {
             return;
         }
-        if (plugin.getGrid().isAtSpawn(e.getItemDrop().getLocation())) {
-            if (Settings.allowSpawnVisitorItemDrop || e.getPlayer().isOp() || VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypassprotect")
-                    || plugin.getGrid().locationIsOnIsland(e.getPlayer(), e.getItemDrop().getLocation())) {
-                return;
-            }
-        }
-        if (Settings.allowVisitorItemDrop || e.getPlayer().isOp() || VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypassprotect")
+        Island island = plugin.getGrid().getIslandAt(e.getItemDrop().getLocation());
+        if ((island != null && island.getIgsFlag(SettingsFlag.VISITOR_ITEM_DROP))
+                || e.getPlayer().isOp() || VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypassprotect")
                 || plugin.getGrid().locationIsOnIsland(e.getPlayer(), e.getItemDrop().getLocation())) {
             return;
         }
-        e.getPlayer().sendMessage(ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).islandProtected);
+        Util.sendMessage(e.getPlayer(), ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).islandProtected);
         e.setCancelled(true);
     }
 
@@ -251,10 +506,10 @@ public class PlayerEvents implements Listener {
          * plugin.getLogger().info(e.getEventName());
          * }
          */
-        if (!IslandGuard.inWorld(e.getPlayer()) 
-                || Settings.allowTeleportWhenFalling 
-                || !e.getPlayer().getGameMode().equals(GameMode.SURVIVAL) 
-                || e.getPlayer().isOp() 
+        if (!IslandGuard.inWorld(e.getPlayer())
+                || Settings.allowTeleportWhenFalling
+                || !e.getPlayer().getGameMode().equals(GameMode.SURVIVAL)
+                || e.getPlayer().isOp()
                 || e.getPlayer().isFlying()) {
             fallingPlayers.remove(e.getPlayer().getUniqueId());
             return;
@@ -274,10 +529,10 @@ public class PlayerEvents implements Listener {
 
     /**
      * Prevents teleporting when falling based on setting by stopping commands
-     * 
-     * @param e
+     *
+     * @param e - event
      */
-    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = false)
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
     public void onPlayerTeleport(final PlayerCommandPreprocessEvent e) {
         if (DEBUG) {
             plugin.getLogger().info(e.getEventName());
@@ -292,25 +547,25 @@ public class PlayerEvents implements Listener {
         // e.getMessage().substring(1).toLowerCase() + "'");
         if (isFalling(e.getPlayer().getUniqueId()) && (Settings.fallingCommandBlockList.contains("*") || Settings.fallingCommandBlockList.contains(e.getMessage().substring(1).toLowerCase()))) {
             // Sorry you are going to die
-            e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).errorNoPermission); 
-            e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).islandcannotTeleport);
+            Util.sendMessage(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).errorNoPermission);
+            Util.sendMessage(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).islandcannotTeleport);
             e.setCancelled(true);
         }
     }
 
     /**
      * Prevents teleporting when falling based on setting and teleporting to locked islands
-     * 
-     * @param e
+     *
+     * @param e - event
      */
-    @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = false)
+    @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
     public void onPlayerTeleport(final PlayerTeleportEvent e) {
         if (DEBUG) {
             plugin.getLogger().info(e.getEventName());
             plugin.getLogger().info("DEBUG: to = " + e.getTo());
             plugin.getLogger().info("DEBUG: from = " + e.getFrom());
         }
-        // Options - 
+        // Options -
         // Player is in an island world and trying to teleport out - handle
         // Player is in an island world and trying to teleport within - handle
         // Player is not in an island world and trying to teleport in - handle
@@ -342,7 +597,7 @@ public class PlayerEvents implements Listener {
                 if (DEBUG)
                     plugin.getLogger().info("DEBUG: player is falling");
                 // Sorry you are going to die
-                e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).islandcannotTeleport);
+                Util.sendMessage(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).islandcannotTeleport);
                 e.setCancelled(true);
                 // Check if the player is in the void and kill them just in case
                 if (e.getPlayer().getLocation().getBlockY() < 0) {
@@ -360,11 +615,7 @@ public class PlayerEvents implements Listener {
         if (DEBUG)
             plugin.getLogger().info("DEBUG: getting islands for to from");
         Island islandFrom = plugin.getGrid().getProtectedIslandAt(e.getFrom());
-        if (DEBUG && islandFrom != null)
-            plugin.getLogger().info("DEBUG: islandFrom is not null");
         Island islandTo = plugin.getGrid().getProtectedIslandAt(e.getTo());
-        if (DEBUG && islandTo != null)
-            plugin.getLogger().info("DEBUG: islandTo is not null");
 
         // Ender pearl and chorus fruit teleport checks
         if (e.getCause() != null) {
@@ -373,36 +624,23 @@ public class PlayerEvents implements Listener {
                     plugin.getLogger().info("DEBUG: Enderpearl");
 
                 if (islandTo == null) {
-                    if (Settings.allowEnderPearls) {
+                    if (Settings.defaultWorldSettings.get(SettingsFlag.ENDER_PEARL)) {
                         return;
                     }
                 } else {
                     if (DEBUG )
                         plugin.getLogger().info("DEBUG: islandTo is not null enderpearl");
-
-                    if (islandTo.isSpawn()) {
-                        if (DEBUG)
-                            plugin.getLogger().info("DEBUG: islandTo is spawn");
-
-                        if (Settings.allowSpawnEnderPearls) {
-                            if (DEBUG )
-                                plugin.getLogger().info("DEBUG: spawn enderpearl allowed");
-                            return;
-                        }
-                    } else {
+                    if (DEBUG )
+                        plugin.getLogger().info("DEBUG: islandTo is regular island");
+                    if (islandTo.getIgsFlag(SettingsFlag.ENDER_PEARL) || islandTo.getMembers().contains(e.getPlayer().getUniqueId())) {
                         if (DEBUG )
-                            plugin.getLogger().info("DEBUG: islandTo is regular island");
-                        // Regular island
-                        if (islandTo.getIgsFlag(Flags.allowEnderPearls) || islandTo.getMembers().contains(e.getPlayer().getUniqueId())) {
-                            if (DEBUG )
-                                plugin.getLogger().info("DEBUG: enderpearl allowed");
-                            return;
-                        }
+                            plugin.getLogger().info("DEBUG: enderpearl allowed");
+                        return;
                     }
                 }
                 if (DEBUG )
                     plugin.getLogger().info("DEBUG: enderpearl not allowed");
-                e.getPlayer().sendMessage(ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).islandProtected);
+                Util.sendMessage(e.getPlayer(), ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).islandProtected);
                 e.setCancelled(true);
                 return;
             } else if (!plugin.getServer().getVersion().contains("(MC: 1.8")
@@ -416,39 +654,25 @@ public class PlayerEvents implements Listener {
                     boolean cancel = false;
                     // Check both from and to islands
                     if (islandTo == null) {
-                        if (!Settings.allowChorusFruit) {
+                        if (!Settings.defaultWorldSettings.get(SettingsFlag.CHORUS_FRUIT)) {
                             cancel = true;
                         }
                     } else {
-                        if (islandTo.isSpawn()) {
-                            if (!Settings.allowSpawnChorusFruit) {
-                                cancel = true;
-                            }
-                        } else {
-                            // Regular island
-                            if (!islandTo.getIgsFlag(Flags.allowChorusFruit) && !islandTo.getMembers().contains(e.getPlayer().getUniqueId())) {
-                                cancel = true;
-                            }
+                        if (!islandTo.getIgsFlag(SettingsFlag.CHORUS_FRUIT) && !islandTo.getMembers().contains(e.getPlayer().getUniqueId())) {
+                            cancel = true;
                         }
                     }
                     if (islandFrom == null) {
-                        if (!Settings.allowChorusFruit) {
+                        if (!Settings.defaultWorldSettings.get(SettingsFlag.CHORUS_FRUIT)) {
                             cancel = true;
                         }
                     } else {
-                        if (islandFrom.isSpawn()) {
-                            if (!Settings.allowSpawnChorusFruit) {
-                                cancel = true;
-                            }
-                        } else {
-                            // Regular island
-                            if (!islandFrom.getIgsFlag(Flags.allowChorusFruit) && !islandFrom.getMembers().contains(e.getPlayer().getUniqueId())) {
-                                cancel = true;
-                            }
+                        if (!islandFrom.getIgsFlag(SettingsFlag.CHORUS_FRUIT) && !islandFrom.getMembers().contains(e.getPlayer().getUniqueId())) {
+                            cancel = true;
                         }
                     }
                     if (cancel) {
-                        e.getPlayer().sendMessage(ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).islandProtected);
+                        Util.sendMessage(e.getPlayer(), ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).islandProtected);
                         e.setCancelled(true);
                     }
                     return;
@@ -466,14 +690,16 @@ public class PlayerEvents implements Listener {
          * islandTo != null && islandFrom != null - same PlayerIsland or teleport?
          * islandTo == islandFrom
          */
+
         if (DEBUG )
             plugin.getLogger().info("DEBUG: announcements");
-        if (islandTo != null && islandFrom == null && (islandTo.getOwner() != null || islandTo.isSpawn())) {
+        if (islandTo != null && islandFrom == null) {
             if (DEBUG )
                 plugin.getLogger().info("DEBUG: entering");
             // Entering
-            if (islandTo.isLocked() || plugin.getPlayers().isBanned(islandTo.getOwner(),e.getPlayer().getUniqueId())) {
-                e.getPlayer().sendMessage(ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).lockIslandLocked);
+            if (islandTo.getOwner() != null && (islandTo.isLocked() || plugin.getPlayers().isBanned(islandTo.getOwner(),e.getPlayer().getUniqueId()))) {
+                // Locked island
+                Util.sendMessage(e.getPlayer(), ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).lockIslandLocked);
                 if (!plugin.getGrid().locationIsOnIsland(e.getPlayer(), e.getTo()) && !e.getPlayer().isOp()
                         && !VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypassprotect")
                         && !VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypasslock")) {
@@ -486,14 +712,18 @@ public class PlayerEvents implements Listener {
             if (islandTo.isSpawn()) {
                 if (DEBUG )
                     plugin.getLogger().info("DEBUG: islandTo is locked spawn");
-                if (!plugin.myLocale(e.getPlayer().getUniqueId()).lockEnteringSpawn.isEmpty()) {
-                    e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).lockEnteringSpawn);
+                if(islandTo.getIgsFlag(SettingsFlag.ENTER_EXIT_MESSAGES)) {
+                    if (!plugin.myLocale(e.getPlayer().getUniqueId()).lockEnteringSpawn.isEmpty()) {
+                        Util.sendEnterExit(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).lockEnteringSpawn);
+                    }
                 }
             } else {
                 if (DEBUG )
                     plugin.getLogger().info("DEBUG: islandTo is locked regular");
-                if (!plugin.myLocale(e.getPlayer().getUniqueId()).lockNowEntering.isEmpty()) {
-                    e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).lockNowEntering.replace("[name]", plugin.getGrid().getIslandName(islandTo.getOwner())));
+                if(islandTo.getOwner() != null && islandTo.getIgsFlag(SettingsFlag.ENTER_EXIT_MESSAGES)) {
+                    if (!plugin.myLocale(e.getPlayer().getUniqueId()).lockNowEntering.isEmpty()) {
+                        Util.sendEnterExit(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).lockNowEntering.replace("[name]", plugin.getGrid().getIslandName(islandTo.getOwner())));
+                    }
                 }
             }
             // Fire entry event
@@ -507,16 +737,22 @@ public class PlayerEvents implements Listener {
                 if (DEBUG )
                     plugin.getLogger().info("DEBUG: leaving spawn");
                 // Leaving
-                if (!plugin.myLocale(e.getPlayer().getUniqueId()).lockLeavingSpawn.isEmpty()) {
-                    e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).lockLeavingSpawn);
+                if(islandFrom.getIgsFlag(SettingsFlag.ENTER_EXIT_MESSAGES)) {
+                    if (!plugin.myLocale(e.getPlayer().getUniqueId()).lockLeavingSpawn.isEmpty()) {
+                        Util.sendEnterExit(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).lockLeavingSpawn);
+                    }
                 }
             } else {
                 if (DEBUG )
                     plugin.getLogger().info("DEBUG: leaving locked");
-                if (!plugin.myLocale(e.getPlayer().getUniqueId()).lockNowLeaving.isEmpty()) {
-                    e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).lockNowLeaving.replace("[name]", plugin.getGrid().getIslandName(islandFrom.getOwner())));
+                if(islandFrom.getIgsFlag(SettingsFlag.ENTER_EXIT_MESSAGES)) {
+                    if (!plugin.myLocale(e.getPlayer().getUniqueId()).lockNowLeaving.isEmpty()) {
+                        Util.sendEnterExit(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).lockNowLeaving.replace("[name]", plugin.getGrid().getIslandName(islandFrom.getOwner())));
+                    }
                 }
             }
+            // Remove temp perms
+            removeTempPerms(e.getPlayer(), islandFrom, islandTo);
             // Fire exit event
             final IslandExitEvent event = new IslandExitEvent(e.getPlayer().getUniqueId(), islandFrom, e.getFrom());
             plugin.getServer().getPluginManager().callEvent(event);
@@ -526,7 +762,7 @@ public class PlayerEvents implements Listener {
             // Teleporting from one islands to another
             // Entering
             if (islandTo.isLocked() || plugin.getPlayers().isBanned(islandTo.getOwner(),e.getPlayer().getUniqueId())) {
-                e.getPlayer().sendMessage(ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).lockIslandLocked);
+                Util.sendMessage(e.getPlayer(), ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).lockIslandLocked);
                 if (!plugin.getGrid().locationIsOnIsland(e.getPlayer(), e.getTo()) && !e.getPlayer().isOp()
                         && !VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypassprotect")
                         && !VaultHelper.checkPerm(e.getPlayer(), Settings.PERMPREFIX + "mod.bypasslock")) {
@@ -535,28 +771,46 @@ public class PlayerEvents implements Listener {
                     e.setCancelled(true);
                     return;
                 }
-            }            
+            }
             if (islandFrom.isSpawn()) {
                 // Leaving
-                e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).lockLeavingSpawn);
-            } else if (islandFrom.getOwner() != null) {
-                e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).lockNowLeaving.replace("[name]", plugin.getGrid().getIslandName(islandFrom.getOwner())));
+                if(islandFrom.getIgsFlag(SettingsFlag.ENTER_EXIT_MESSAGES) && !plugin.myLocale(e.getPlayer().getUniqueId()).lockLeavingSpawn.isEmpty()) {
+                    Util.sendEnterExit(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).lockLeavingSpawn);
+                }
+            } else if (islandFrom.getOwner() != null && !plugin.myLocale(e.getPlayer().getUniqueId()).lockNowLeaving.isEmpty()) {
+                if(islandFrom.getIgsFlag(SettingsFlag.ENTER_EXIT_MESSAGES)) {
+                    Util.sendEnterExit(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).lockNowLeaving.replace("[name]", plugin.getGrid().getIslandName(islandFrom.getOwner())));
+                }
             }
             if (islandTo.isSpawn()) {
-                e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).lockEnteringSpawn);
-            } else if (islandTo.getOwner() != null) {
-                e.getPlayer().sendMessage(plugin.myLocale(e.getPlayer().getUniqueId()).lockNowEntering.replace("[name]", plugin.getGrid().getIslandName(islandTo.getOwner())));
+                if(islandTo.getIgsFlag(SettingsFlag.ENTER_EXIT_MESSAGES) && !plugin.myLocale(e.getPlayer().getUniqueId()).lockEnteringSpawn.isEmpty()) {
+                    Util.sendEnterExit(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).lockEnteringSpawn);
+                }
+            } else if (islandTo.getOwner() != null && !plugin.myLocale(e.getPlayer().getUniqueId()).lockNowEntering.isEmpty()) {
+                if(islandTo.getIgsFlag(SettingsFlag.ENTER_EXIT_MESSAGES)) {
+                    Util.sendEnterExit(e.getPlayer(), plugin.myLocale(e.getPlayer().getUniqueId()).lockNowEntering.replace("[name]", plugin.getGrid().getIslandName(islandTo.getOwner())));
+                }
+            }
+            // Remove temp perms
+            if (!islandTo.getMembers().contains(e.getPlayer().getUniqueId())) {
+                removeTempPerms(e.getPlayer(), islandFrom, islandTo);
             }
             // Fire exit event
-            final IslandExitEvent event = new IslandExitEvent(e.getPlayer().getUniqueId(), islandTo, e.getTo());
+            final IslandExitEvent event = new IslandExitEvent(e.getPlayer().getUniqueId(), islandFrom, e.getFrom());
             plugin.getServer().getPluginManager().callEvent(event);
             // Fire entry event
-            final IslandEnterEvent event2 = new IslandEnterEvent(e.getPlayer().getUniqueId(), islandFrom, e.getFrom());
+            final IslandEnterEvent event2 = new IslandEnterEvent(e.getPlayer().getUniqueId(), islandTo, e.getTo());
+            plugin.getServer().getPluginManager().callEvent(event2);
+        } else if (islandTo != null && islandFrom != null && (islandTo.equals(islandFrom) && !e.getFrom().getWorld().equals(e.getTo().getWorld()))) {
+            if (DEBUG )
+                plugin.getLogger().info("DEBUG: jumping from dimension to another - same island");
+            // Fire exit event
+            final IslandExitEvent event = new IslandExitEvent(e.getPlayer().getUniqueId(), islandFrom, e.getFrom());
+            plugin.getServer().getPluginManager().callEvent(event);
+            // Fire entry event
+            final IslandEnterEvent event2 = new IslandEnterEvent(e.getPlayer().getUniqueId(), islandTo, e.getTo());
             plugin.getServer().getPluginManager().callEvent(event2);
         }
-
-
-
     }
 
 
@@ -564,8 +818,8 @@ public class PlayerEvents implements Listener {
 
     /**
      * Used to prevent teleporting when falling
-     * 
-     * @param uniqueId
+     *
+     * @param uniqueId - unique ID
      * @return true or false
      */
     public static boolean isFalling(UUID uniqueId) {
@@ -574,8 +828,8 @@ public class PlayerEvents implements Listener {
 
     /**
      * Used to prevent teleporting when falling
-     * 
-     * @param uniqueId
+     *
+     * @param uniqueId - unique ID
      */
     public static void setFalling(UUID uniqueId) {
         fallingPlayers.add(uniqueId);
@@ -583,8 +837,8 @@ public class PlayerEvents implements Listener {
 
     /**
      * Unset the falling flag
-     * 
-     * @param uniqueId
+     *
+     * @param uniqueId - unique ID
      */
     public static void unsetFalling(UUID uniqueId) {
         // getLogger().info("DEBUG: unset falling");
@@ -593,9 +847,9 @@ public class PlayerEvents implements Listener {
 
     /**
      * Prevents visitors from using commands on islands, like /spawner
-     * @param e
+     * @param e - event
      */
-    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = false)
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
     public void onVisitorCommand(final PlayerCommandPreprocessEvent e) {
         if (DEBUG) {
             plugin.getLogger().info("Visitor command " + e.getEventName() + ": " + e.getMessage());
@@ -610,50 +864,119 @@ public class PlayerEvents implements Listener {
         //plugin.getLogger().info(Settings.visitorCommandBlockList.toString());
         String[] args = e.getMessage().substring(1).toLowerCase().split(" ");
         if (Settings.visitorCommandBlockList.contains(args[0])) {
-            e.getPlayer().sendMessage(ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).islandProtected);
+            Util.sendMessage(e.getPlayer(), ChatColor.RED + plugin.myLocale(e.getPlayer().getUniqueId()).islandProtected);
+            e.setCancelled(true);
+        }
+    }
+
+
+    /**
+     * Prevents visitors from getting damage if invinciblevisitors option is set to TRUE
+     * @param e - event
+     */
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
+    public void onVisitorGetDamage(EntityDamageByEntityEvent e){
+        if(!Settings.invincibleVisitors
+                || !Settings.visitorDamagePrevention.contains(e.getCause())
+                || !(e.getEntity() instanceof Player)
+                || !e.getCause().equals(DamageCause.ENTITY_ATTACK)) {
+            return;
+        }
+        Player p = (Player) e.getEntity();
+        if (!IslandGuard.inWorld(p) || plugin.getGrid().locationIsOnIsland(p, p.getLocation())) return;
+        // Find out who the attacker is
+        Player attacker = null;
+        if (e.getDamager() instanceof Player) {
+            attacker = (Player)e.getDamager();
+        }
+        if (e.getDamager() instanceof Projectile) {
+            Projectile projectile = (Projectile)e.getDamager();
+            if (projectile.getShooter() instanceof Player) {
+                attacker = (Player)projectile.getShooter();
+            }
+        }
+        // If we cannot work out who is the attacker then just prevent the damage and return
+        if (attacker == null) {
+            e.setCancelled(true);
+            return;
+        }
+        // Check if at spawn
+        if (plugin.getGrid().isAtSpawn(e.getEntity().getLocation())) {
+            // Check if PVP is allowed at spawn or not
+            if ((e.getEntity().getWorld().getEnvironment().equals(Environment.NORMAL)
+                    && plugin.getGrid().getIslandAt(e.getEntity().getLocation()).getIgsFlag(SettingsFlag.PVP))
+                    || (e.getEntity().getWorld().getEnvironment().equals(Environment.NETHER)
+                            && plugin.getGrid().getIslandAt(e.getEntity().getLocation()).getIgsFlag(SettingsFlag.NETHER_PVP))) {
+                // Damage allowed
+                return;
+            }
+            // PVP is not allowed
+            attacker.sendMessage(ChatColor.RED + plugin.myLocale(attacker.getUniqueId()).visitorsProtected);
+            e.setCancelled(true);
+        } else {
+            // Not at spawn, protect visitor
+            attacker.sendMessage(ChatColor.RED + plugin.myLocale(attacker.getUniqueId()).visitorsProtected);
             e.setCancelled(true);
         }
     }
 
     /**
      * Prevents visitors from getting damage if invinciblevisitors option is set to TRUE
-     * @param e
+     * @param e - event
      */
-    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = false)
+    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
     public void onVisitorGetDamage(EntityDamageEvent e){
-        if(!Settings.invincibleVisitors) return;
-        if(!(e.getEntity() instanceof Player)) return;
-
+        if(!Settings.invincibleVisitors
+                || !(e.getEntity() instanceof Player)
+                || !e.getCause().equals(DamageCause.ENTITY_ATTACK)) {
+            return;
+        }
         Player p = (Player) e.getEntity();
         if (!IslandGuard.inWorld(p) || plugin.getGrid().locationIsOnIsland(p, p.getLocation())) return;
 
         if (Settings.visitorDamagePrevention.contains(e.getCause())) e.setCancelled(true);
 
-        else if(e.getCause().equals(DamageCause.VOID)) {
-            if(plugin.getPlayers().hasIsland(p.getUniqueId())) {
+        if(e.getCause().equals(DamageCause.VOID)) {
+            if (plugin.getPlayers().hasIsland(p.getUniqueId()) || plugin.getPlayers().inTeam(p.getUniqueId())) {
                 Location safePlace = plugin.getGrid().getSafeHomeLocation(p.getUniqueId(), 1);
                 if (safePlace != null) {
+                    unsetFalling(p.getUniqueId());
                     p.teleport(safePlace);
                     // Set their fall distance to zero otherwise they crash onto their island and die
                     p.setFallDistance(0);
                     e.setCancelled(true);
                     return;
-                } 
+                }
             }
-            // No island, or no safe spot on island
             if (plugin.getGrid().getSpawn() != null) {
                 p.teleport(plugin.getGrid().getSpawnPoint());
                 // Set their fall distance to zero otherwise they crash onto their island and die
                 p.setFallDistance(0);
                 e.setCancelled(true);
                 return;
-            }
-            // No island spawn, try regular spawn
-            if (!p.performCommand("spawn")) {
+            } else if (!p.performCommand("spawn")) {
                 // If this command doesn't work, let them die otherwise they may get trapped in the void forever
                 return;
+            } else {
+                // Set their fall distance to zero otherwise they crash onto their island and die
+                p.setFallDistance(0);
+                e.setCancelled(true);
             }
-            // Set their fall distance to zero otherwise they crash onto their island and die
+        }
+    }
+
+    /**
+     * Protect players from damage when teleporting
+     * @param e - event
+     */
+    @EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
+    public void onPlayerTeleportDamage(EntityDamageEvent e){
+        if(!(e.getEntity() instanceof Player)) return;
+
+        Player p = (Player) e.getEntity();
+        if (plugin.getPlayers().isInTeleport(p.getUniqueId())) {
+            if (DEBUG)
+                plugin.getLogger().info("DEBUG: protecting player from teleport damage");
             p.setFallDistance(0);
             e.setCancelled(true);
         }
