@@ -49,12 +49,14 @@ public class MiniShopItem {
     private String extra;
     private String description;
     private ItemStack item;
+    private EntityType entityType;
 
     // private ASkyBlock plugin = ASkyBlock.getPlugin();
 
     /**
      * 
      */
+    @SuppressWarnings("deprecation")
     public MiniShopItem(Material material, String extra, int slot, String description, int quantity, Double price, Double sellPrice) {
         this.slot = slot;
         this.material = material;
@@ -74,7 +76,7 @@ public class MiniShopItem {
             item.setAmount(quantity);
            // Deal with extras
             if (!extra.isEmpty()) {
-                // plugin.getLogger().info("DEBUG: extra is not empty");
+                // plugin.getLogger().info("DEBUG: extra is not empty");                
                 // If it not a potion, then the extras should just be durability
                 if (!material.name().contains("POTION")) {
                     if (material.equals(Material.MONSTER_EGG)) {
@@ -99,7 +101,7 @@ public class MiniShopItem {
                                 }
                             }
                         }
-                    } else {
+                    } else if (!material.equals(Material.MOB_SPAWNER)) {
                         item.setDurability(Short.parseShort(extra));
                     }
                 } else {
@@ -115,6 +117,16 @@ public class MiniShopItem {
             List<String> desc = new ArrayList<String>(Arrays.asList(description.split("\\|")));
             meta.setDisplayName(desc.get(0));
             ArrayList<String> buyAndSell = new ArrayList<String>();
+            if (material.equals(Material.MOB_SPAWNER) && !extra.isEmpty()) {
+                //Bukkit.getLogger().info("DEBUG: mob spawner and extra is " + extra);
+                // Get the entity type
+                for (EntityType type : EntityType.values()) {
+                    if (extra.toUpperCase().equals(type.name())) {
+                        entityType = type;
+                        break;
+                    }
+                }
+            }
             if (desc.size() > 1) {
                 desc.remove(0);// Remove the name
                 buyAndSell.addAll(desc); // Add the rest to the description
@@ -164,7 +176,11 @@ public class MiniShopItem {
         ItemStack temp = this.item.clone();
         ItemMeta meta = temp.getItemMeta();
         meta.setDisplayName(null);
-        meta.setLore(null);
+        List<String> lore = new ArrayList<String>(1);
+        if (item.getType().equals(Material.MOB_SPAWNER)) {  
+            lore.add(Util.prettifyText(entityType.name()));
+        }
+        meta.setLore(lore);
         temp.setItemMeta(meta);
         return temp;
     }
@@ -490,6 +506,7 @@ public class MiniShopItem {
             case 3:
                 return "JUNGLE_LEAVES";
             } // Note Acacia and Dark Oak are LEAVES_2 for some reason...
+            return mat.toString();
         case COAL:
             switch (damage) {
             case 0:
@@ -943,8 +960,6 @@ public class MiniShopItem {
             break;
         case FIREWORK:
             return "Firework Rocket";
-        case FISHING_ROD:
-            break;
         case FLINT:
             break;
         case FLINT_AND_STEEL:
@@ -1504,7 +1519,6 @@ public class MiniShopItem {
         switch (mat) {
         case BOW:
         case SHEARS:
-        case FISHING_ROD:
         case FLINT_AND_STEEL:
 
         case CHAINMAIL_BOOTS:
@@ -1564,6 +1578,13 @@ public class MiniShopItem {
             return false;
         }
 
+    }
+
+    /**
+     * @return the entityType
+     */
+    public EntityType getEntityType() {
+        return entityType;
     }
 
 }
