@@ -35,6 +35,7 @@ import org.bukkit.inventory.Inventory;
 import com.wasteofplastic.askyblock.ASkyBlock;
 import com.wasteofplastic.askyblock.Settings;
 import com.wasteofplastic.askyblock.schematics.Schematic;
+import com.wasteofplastic.askyblock.util.Util;
 import com.wasteofplastic.askyblock.util.VaultHelper;
 
 public class SchematicsPanel implements Listener {
@@ -42,7 +43,7 @@ public class SchematicsPanel implements Listener {
     private HashMap<UUID, List<SPItem>> schematicItems = new HashMap<UUID, List<SPItem>>();
 
     /**
-     * @param plugin
+     * @param plugin - ASkyBlock plugin object
      */
     public SchematicsPanel(ASkyBlock plugin) {
         this.plugin = plugin;
@@ -82,7 +83,7 @@ public class SchematicsPanel implements Listener {
             }
             return newPanel;
         } else {
-            player.sendMessage(ChatColor.RED + plugin.myLocale().errorCommandNotReady);
+            Util.sendMessage(player, ChatColor.RED + plugin.myLocale().errorCommandNotReady);
         }
         return null;
     }
@@ -139,16 +140,16 @@ public class SchematicsPanel implements Listener {
             if (item.getCost() > 0) {
                 if (Settings.useEconomy && VaultHelper.setupEconomy() && !VaultHelper.econ.has(player, item.getCost())) {
                     // Too expensive
-                    player.sendMessage(ChatColor.RED + plugin.myLocale(player.getUniqueId()).minishopYouCannotAfford.replace("[description]", item.getName()));        
+                    Util.sendMessage(player, ChatColor.RED + plugin.myLocale(player.getUniqueId()).minishopYouCannotAfford.replace("[description]", item.getName()));        
                 } else {
                     // Do something
                     if (Settings.useEconomy && VaultHelper.setupEconomy()) {
                         VaultHelper.econ.withdrawPlayer(player, item.getCost());                   
                     } 
-                    player.performCommand(Settings.ISLANDCOMMAND + " make " + item.getHeading());
+                    Util.runCommand(player, Settings.ISLANDCOMMAND + " make " + item.getHeading());
                 }
             } else {
-                player.performCommand(Settings.ISLANDCOMMAND + " make " + item.getHeading());
+                Util.runCommand(player, Settings.ISLANDCOMMAND + " make " + item.getHeading());
             }
             schematicItems.remove(player.getUniqueId());
             thisPanel.clear();   
