@@ -32,7 +32,6 @@ import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Biome;
 import org.bukkit.block.Block;
-import org.bukkit.block.Chest;
 import org.bukkit.block.CreatureSpawner;
 import org.bukkit.block.DoubleChest;
 import org.bukkit.block.Sign;
@@ -63,9 +62,9 @@ public class IslandBlock {
     private PotBlock pot;
     private EntityType spawnerBlockType;
     // Chest contents
-    private HashMap<Byte,ItemStack> chestContents = new HashMap<Byte,ItemStack>();
-    public static final HashMap<String, Material> WEtoM = new HashMap<String, Material>();
-    public static final HashMap<String, EntityType> WEtoME = new HashMap<String, EntityType>();
+    private final Map<Byte,ItemStack> chestContents = new HashMap<>();
+    protected static final Map<String, Material> WEtoM = new HashMap<>();
+    protected static final Map<String, EntityType> WEtoME = new HashMap<>();
 
     static {
         // Establish the World Edit to Material look up
@@ -85,6 +84,7 @@ public class IslandBlock {
         WEtoM.put("CARROT_ON_A_STICK",Material.CARROT_STICK);
         WEtoM.put("CARROT",Material.CARROT_ITEM);
         WEtoM.put("CAULDRON", Material.CAULDRON_ITEM);
+        WEtoM.put("CHEST_MINECART", Material.STORAGE_MINECART);
         WEtoM.put("CLOCK", Material.WATCH);
         WEtoM.put("COBBLESTONE_WALL",Material.COBBLE_WALL);
         WEtoM.put("COMMAND_BLOCK",Material.COMMAND);
@@ -117,6 +117,7 @@ public class IslandBlock {
         WEtoM.put("GOLDEN_HOE", Material.GOLD_HOE);
         WEtoM.put("GOLDEN_AXE", Material.GOLD_AXE);
         WEtoM.put("GOLDEN_BOOTS", Material.GOLD_BOOTS);
+        WEtoM.put("GUNPOWDER", Material.SULPHUR);
         WEtoM.put("HARDENED_CLAY",Material.HARD_CLAY);
         WEtoM.put("HEAVY_WEIGHTED_PRESSURE_PLATE",Material.GOLD_PLATE);
         WEtoM.put("IRON_BARS",Material.IRON_FENCE);
@@ -132,6 +133,8 @@ public class IslandBlock {
         WEtoM.put("NETHER_WART",Material.NETHER_STALK);
         WEtoM.put("NETHERBRICK",Material.NETHER_BRICK_ITEM);
         WEtoM.put("OAK_STAIRS",Material.WOOD_STAIRS);
+        WEtoM.put("PISTON",Material.PISTON_BASE);
+        WEtoM.put("PLANKS",Material.WOOD);
         WEtoM.put("POTATO", Material.POTATO_ITEM);
         WEtoM.put("RAIL",Material.RAILS);
         WEtoM.put("RECORD_11",Material.RECORD_11);
@@ -151,6 +154,7 @@ public class IslandBlock {
         WEtoM.put("REPEATER",Material.DIODE);
         WEtoM.put("SKULL", Material.SKULL_ITEM);
         WEtoM.put("SPAWN_EGG",Material.MONSTER_EGG);
+        WEtoM.put("STICKY_PISTON",Material.PISTON_STICKY_BASE);
         WEtoM.put("STONE_BRICK_STAIRS",Material.BRICK_STAIRS);
         WEtoM.put("STONE_BRICK_STAIRS",Material.SMOOTH_STAIRS);
         WEtoM.put("STONE_SHOVEL",Material.STONE_SPADE);
@@ -221,7 +225,6 @@ public class IslandBlock {
         skull = null;
         pot = null;
         spawnerBlockType = null;
-        chestContents = new HashMap<Byte,ItemStack>();
     }
     /**
      * @return the type
@@ -538,19 +541,24 @@ public class IslandBlock {
      * @param nms
      * @param blockLoc
      */
+    //@SuppressWarnings("deprecation")
+    @SuppressWarnings("deprecation")
     public void paste(NMSAbstraction nms, Location blockLoc, boolean usePhysics, Biome biome) {
         // Only paste air if it is below the sea level and in the overworld
         Block block = new Location(blockLoc.getWorld(), x, y, z).add(blockLoc).getBlock();
         block.setBiome(biome);
         nms.setBlockSuperFast(block, typeId, data, usePhysics);
         if (signText != null) {
+            if (block.getTypeId() != typeId) {
+                block.setTypeId(typeId);
+            }
             // Sign
             Sign sign = (Sign) block.getState();
             int index = 0;
             for (String line : signText) {
                 sign.setLine(index++, line);
             }
-            sign.update();
+            sign.update(true, false);
         } else if (banner != null) {
             banner.set(block);
         } else if (skull != null){
@@ -558,21 +566,32 @@ public class IslandBlock {
         } else if (pot != null){
             pot.set(nms, block);
         } else if (spawnerBlockType != null) {
+            if (block.getTypeId() != typeId) {
+                block.setTypeId(typeId);
+            }
             CreatureSpawner cs = (CreatureSpawner)block.getState();
             cs.setSpawnedType(spawnerBlockType);
+            //Bukkit.getLogger().info("DEBUG: setting spawner");
+            cs.update(true, false);
         } else if (!chestContents.isEmpty()) {
+            if (block.getTypeId() != typeId) {
+                block.setTypeId(typeId);
+            }
+            //Bukkit.getLogger().info("DEBUG: inventory holder "+ block.getType());
             // Check if this is a double chest
-            Chest chestBlock = (Chest) block.getState();
-            InventoryHolder iH = chestBlock.getInventory().getHolder();
-            if (iH instanceof DoubleChest) {
+            
+            InventoryHolder chestBlock = (InventoryHolder) block.getState();
+            //InventoryHolder iH = chestBlock.getInventory().getHolder();
+            if (chestBlock instanceof DoubleChest) {
                 //Bukkit.getLogger().info("DEBUG: double chest");
-                DoubleChest doubleChest = (DoubleChest) iH;
+                DoubleChest doubleChest = (DoubleChest) chestBlock;
                 for (ItemStack chestItem: chestContents.values()) {
                     doubleChest.getInventory().addItem(chestItem);
                 }
             } else {
-                // Signle chest
+                // Single chest
                 for (Entry<Byte, ItemStack> en : chestContents.entrySet()) {
+                    //Bukkit.getLogger().info("DEBUG: " + en.getKey() + ","  + en.getValue());
                     chestBlock.getInventory().setItem(en.getKey(), en.getValue());
                 }
             }
