@@ -58,11 +58,10 @@ public class SkullBlock {
     private String skullTextureValue = null;
     private String skullTextureSignature = null;
 
-    private static HashMap<Integer, SkullType> skullTypeList;
-    private static HashMap<Integer, BlockFace> skullRotationList;
+    private static final Map<Integer, SkullType> skullTypeList = new HashMap<>();
+    private static final Map<Integer, BlockFace> skullRotationList = new HashMap<>();
 
     static {
-        skullTypeList = new HashMap<Integer, SkullType>();
         skullTypeList.put(0, SkullType.SKELETON);
         skullTypeList.put(1, SkullType.WITHER);
         skullTypeList.put(2, SkullType.ZOMBIE);
@@ -71,7 +70,6 @@ public class SkullBlock {
     }
 
     static {
-        skullRotationList = new HashMap<Integer, BlockFace>();
         skullRotationList.put(0, BlockFace.NORTH);
         skullRotationList.put(1, BlockFace.NORTH_NORTH_EAST);
         skullRotationList.put(2, BlockFace.NORTH_EAST);
@@ -103,7 +101,7 @@ public class SkullBlock {
         if(skullTextureValue != null){
             setSkullWithNonPlayerProfile(skullTextureValue, skullTextureSignature, skullOwnerUUID, skullOwnerName, skull);
         }
-        skull.update();
+        skull.update(true, false);
         return true;
     }
 
