@@ -16,13 +16,14 @@
  *******************************************************************************/
 package com.wasteofplastic.askyblock.util;
 
-import net.milkbowl.vault.economy.Economy;
-import net.milkbowl.vault.permission.Permission;
-
+import org.bukkit.World;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.RegisteredServiceProvider;
 
 import com.wasteofplastic.askyblock.ASkyBlock;
+
+import net.milkbowl.vault.economy.Economy;
+import net.milkbowl.vault.permission.Permission;
 
 /**
  * Helper class for Vault Economy and Permissions
@@ -60,7 +61,7 @@ public class VaultHelper {
     }
 
     /**
-     * Checks permission of player in world or in any world
+     * Checks permission of player in the world the player is in now
      * 
      * @param player
      * @param perm
@@ -71,6 +72,18 @@ public class VaultHelper {
     }
 
     /**
+     * Checks permission of player in world 
+     * 
+     * @param player
+     * @param perm
+     * @param world
+     * @return
+     */
+    public static boolean checkPerm(final Player player, final String perm, final World world) {
+        return permission.playerHas(world.getName(), player, perm);
+    }
+
+    /**
      * Adds permission to player
      * 
      * @param player
@@ -78,6 +91,16 @@ public class VaultHelper {
      */
     public static void addPerm(final Player player, final String perm) {
         permission.playerAdd(player, perm);
+    }
+    
+    /**
+     * Add permission to player in world
+     * @param player
+     * @param perm
+     * @param world
+     */
+    public static void addPerm(final Player player, final String perm, final World world) {
+        permission.playerAdd(world.getName(), player, perm);
     }
 
     /**
@@ -90,4 +113,14 @@ public class VaultHelper {
         permission.playerRemove(player, perm);
     }
 
+    /**
+     * Removes a player's permission in world
+     * 
+     * @param player
+     * @param perm
+     * @param world
+     */
+    public static void removePerm(final Player player, final String perm, World world) {
+        permission.playerRemove(world.getName(), player, perm);
+    }
 }
